% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{AFNI_HRF}
\alias{AFNI_HRF}
\title{AFNI HRF Constructor Function}
\usage{
AFNI_HRF(name, nbasis, params)
}
\arguments{
\item{name}{A string specifying the name of the AFNI HRF.}

\item{nbasis}{An integer representing the number of basis functions for the AFNI HRF.}

\item{params}{A list containing the parameter values for the AFNI HRF.}
}
\value{
An AFNI_HRF object with the specified properties.
}
\description{
The \code{AFNI_HRF} function creates an object representing an AFNI-specific hemodynamic response function (HRF). It is a class constructor for AFNI HRFs.
}
\seealso{
HRF
}
