% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-afni.R
\name{afni_trialwise}
\alias{afni_trialwise}
\title{construct a native AFNI hrf specification for '3dDeconvolve' and individually modulated events using the 'stim_times_IM' argument.}
\usage{
afni_trialwise(
  label,
  basis = c("spmg1", "block", "dmblock", "gamma", "wav"),
  onsets = NULL,
  durations = 0,
  subset = NULL,
  id = NULL,
  start = 0,
  stop = 22
)
}
\arguments{
\item{label}{name of regressor}

\item{basis}{the impulse response function or the name of a pre-supplied function,
one of: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian", "tent", "bs".
Can also be an \code{HRF} object.}

\item{onsets}{optional onsets override. If missing, onsets will be taken from the LHS of the main model formula.}

\item{durations}{optional durations override. If missing, durations argument from \code{event_model} is used.}

\item{subset}{an expression indicating the subset of 'onsets' to keep.}

\item{id}{a unique \code{character} identifier used to refer to term, otherwise will be determined from variable names.}

\item{start}{start of hrf (for multiple basis hrfs)}

\item{stop}{end of hrf (for multiple basis hrfs)}
}
\value{
an \code{afni_trialwise_hrfspec} instance
}
\description{
construct a native AFNI hrf specification for '3dDeconvolve' and individually modulated events using the 'stim_times_IM' argument.
}
\examples{


tw <- afni_trialwise("trialwise", basis="gamma", onsets=seq(1,100,by=5))

}
