% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/regressor.R
\name{amplitudes}
\alias{amplitudes}
\alias{amplitudes.regressor}
\title{get amplitude vector}
\usage{
amplitudes(x)

\method{amplitudes}{regressor}(x)
}
\arguments{
\item{x}{the object}
}
\description{
get the intensity or amplitudes associated with each event
}
