% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta.R
\name{apply_huber_weights}
\alias{apply_huber_weights}
\title{Apply Huber Robust Weights}
\usage{
apply_huber_weights(y, se, X, w)
}
\arguments{
\item{y}{Effect sizes}

\item{se}{Standard errors}

\item{X}{Design matrix}

\item{w}{Initial weights}
}
\value{
Updated weights
}
\description{
Apply Huber Robust Weights
}
\keyword{internal}
