% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar_whiten_inplace}
\alias{ar_whiten_inplace}
\title{AR(p) whitening of data and design matrices}
\usage{
ar_whiten_inplace(Y, X, phi_coeffs, exact_first_ar1 = FALSE)
}
\arguments{
\item{Y}{Numeric matrix of data (time x voxels)}

\item{X}{Numeric matrix of design (time x predictors)}

\item{phi_coeffs}{Numeric vector of AR coefficients (length p)}

\item{exact_first_ar1}{Logical, apply exact scaling of first sample for AR(1)}
}
\description{
Applies a causal AR filter defined by \code{phi_coeffs} to both \code{Y} and \code{X}
matrices in place.
}
\keyword{internal}
