% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroim2-methods.R
\name{as.array.NeuroVec}
\alias{as.array.NeuroVec}
\title{Coerce NeuroVec to base array}
\usage{
\method{as.array}{NeuroVec}(x, ...)
}
\arguments{
\item{x}{A NeuroVec object from neuroim2}

\item{...}{Additional arguments (currently unused)}
}
\value{
A base R array containing the data from the NeuroVec object
}
\description{
This method allows NeuroVec objects from neuroim2 to be converted to base R arrays
using the standard as.array() function. This is particularly useful in testing
and data manipulation contexts.
}
