% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{as_hrf}
\alias{as_hrf}
\title{Turn any function into an HRF object}
\usage{
as_hrf(
  f,
  name = deparse(substitute(f)),
  nbasis = 1L,
  span = 24,
  params = list()
)
}
\arguments{
\item{f}{The function to be turned into an HRF object. It must accept a single argument \code{t} (time).}

\item{name}{The name for the HRF object. Defaults to the deparsed name of \code{f}.}

\item{nbasis}{The number of basis functions represented by \code{f}. Defaults to 1L.}

\item{span}{The nominal time span (duration in seconds) of the HRF. Defaults to 24.}

\item{params}{A named list of parameters associated with the HRF function \code{f}. Defaults to an empty list.}

\item{...}{Additional arguments to pass to the function f.}
}
\value{
A new HRF object.
}
\description{
This is the core constructor for creating HRF objects in the refactored system.
It takes a function \code{f(t)} and attaches standard HRF attributes.
}
\keyword{internal}
