% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline}
\alias{baseline}
\title{Create a model specification for modeling low-frequency drift in fmri time series.}
\usage{
baseline(
  degree = 1,
  basis = c("constant", "poly", "bs", "ns"),
  name = NULL,
  intercept = c("runwise", "global", "none")
)
}
\arguments{
\item{degree}{number of basis terms for each image block (ignored for 'constant' basis)}

\item{basis}{the type of polynomial basis.}

\item{name}{the name of the term}

\item{intercept}{the type of intercept to include}

\item{constant}{whether to include an intercept term}
}
\description{
Create a model specification for modeling low-frequency drift in fmri time series.
}
