% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline_term}
\alias{baseline_term}
\title{baseline_term}
\usage{
baseline_term(varname, mat, colind, rowind)
}
\arguments{
\item{varname}{the name of the term}

\item{mat}{the matrix of covariates}

\item{colind}{the column indices}

\item{rowind}{the row indices}
}
\description{
baseline_term
}
