% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{baseline_terms}
\alias{baseline_terms}
\title{baseline_terms}
\usage{
baseline_terms(x)
}
\arguments{
\item{x}{the model object}
}
\value{
A list containing the 'baseline_term's.
}
\description{
This function extracts a list of 'baseline_term's from the provided model object \code{x}.
}
\details{
Extract a list of 'baseline_term's from the model object.
}
