% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_stats.R
\name{beta_stats}
\alias{beta_stats}
\title{Beta Statistics for Linear Model}
\usage{
beta_stats(lmfit, varnames, se = TRUE)
}
\arguments{
\item{lmfit}{Fitted linear model object.}

\item{varnames}{Vector of variable names.}

\item{se}{Logical flag indicating whether to calculate standard errors (default: TRUE).}
}
\value{
A list containing the following elements:
\itemize{
\item{\code{estimate}: Estimated beta coefficients.}
\item{\code{stat}: t-statistics of the beta coefficients (if \code{se = TRUE}).}
\item{\code{se}: Standard errors of the beta coefficients (if \code{se = TRUE}).}
\item{\code{prob}: Probabilities associated with the t-statistics (if \code{se = TRUE}).}
\item{\code{stat_type}: Type of the statistics calculated.}
}
}
\description{
This function calculates beta statistics for a linear model fit.
}
\examples{
data(mtcars)
lm_fit <- lm(mpg ~ wt + qsec + am, data = mtcars)
beta_stats(lm_fit, c("Intercept", "wt", "qsec", "am"))
}
