% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{block}
\alias{block}
\title{Create a Block Variable.}
\usage{
block(x)
}
\arguments{
\item{x}{The block variable.}
}
\value{
An object of class "blockspec".
}
\description{
Returns a block variable that is constant over the span of a scanning run.
}
