% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{block_term}
\alias{block_term}
\title{block_term}
\usage{
block_term(
  varname,
  blockids,
  expanded_blockids,
  mat,
  type = c("runwise", "global")
)
}
\arguments{
\item{varname}{the name of the term}

\item{blockids}{the ordered sequence of blockids}

\item{expanded_blockids}{the vector of blockids expanded by run}

\item{mat}{the \code{matrix} of covariates}
}
\description{
construct a constant term
}
