% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame.R
\name{blockids.sampling_frame}
\alias{blockids.sampling_frame}
\title{Extract Block IDs from a Sampling Frame}
\usage{
\method{blockids}{sampling_frame}(x)
}
\arguments{
\item{x}{A sampling_frame object.}
}
\value{
An integer vector.
}
\description{
Returns an integer vector indicating the block ID for each scan.
}
