% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{blocklens}
\alias{blocklens}
\title{get block lengths}
\usage{
blocklens(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
get the number of scans/images in each block
}
