% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{BSpline}
\alias{BSpline}
\title{B-spline basis}
\usage{
BSpline(x, degree)
}
\arguments{
\item{x}{a numeric vector at which to evaluate the spline. Missing values are not allowed in x}

\item{degree}{the degree of the piecewise polynomial}
}
\value{
an \code{BSpline} list instance
}
\description{
Generate the B-spline basis matrix for a polynomial spline.
}
\seealso{
\link[splines]{bs}
}
