% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R, R/all_generic.R
\name{build_afni_stims}
\alias{build_afni_stims}
\alias{build_afni_stims.event_term}
\title{Build AFNI stimulus definitions}
\usage{
build_afni_stims(x, ...)

\method{build_afni_stims}{event_term}(x, iresp = FALSE, tr_times = 1, ...)

build_afni_stims(x, ...)
}
\arguments{
\item{x}{the term}

\item{...}{extra args}

\item{iresp}{Logical; generate -iresp files when TRUE}

\item{tr_times}{Numeric; TR timing for stimuli}
}
\value{
List of stimulus descriptors or \code{NULL}
}
\description{
Generate stimulus files for AFNI's 3dDeconvolve.
}
\concept{AFNI}
\keyword{internal}
