% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fdr.R
\name{build_block_neighbors}
\alias{build_block_neighbors}
\title{Build Block Neighbor Adjacency}
\usage{
build_block_neighbors(n_blocks, connectivity = 26)
}
\arguments{
\item{n_blocks}{Grid dimensions}

\item{connectivity}{6 or 26}
}
\value{
List of neighbor indices
}
\description{
Build Block Neighbor Adjacency
}
\keyword{internal}
