% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{build_contrast_from_names}
\alias{build_contrast_from_names}
\title{Build Contrast from Named Vector}
\usage{
build_contrast_from_names(contrast, object)
}
\arguments{
\item{contrast}{Named numeric vector}

\item{object}{fmri_meta object}
}
\value{
Numeric vector of contrast weights
}
\description{
Build Contrast from Named Vector
}
\keyword{internal}
