% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R
\name{build_decon_command}
\alias{build_decon_command}
\title{Build 3dDeconvolve command for an fMRI model}
\usage{
build_decon_command(model, dataset, working_dir, opts)
}
\arguments{
\item{model}{The fMRI model, usually created using the fmri_model function}

\item{dataset}{The fMRI dataset, usually created using the fmri_dataset function}

\item{working_dir}{The working directory}

\item{opts}{A list of options for the 3dDeconvolve command}
}
\value{
A list containing:
- cmd: The 3dDeconvolve command string
- cmdlines: The command lines for the 3dDeconvolve command
- afni_stims: A list of AFNI stimulus objects
- afni_baseline_mats: A list of AFNI baseline matrices
- gltfiles: A list of GLT (general linear test) filenames
- gltnames: A list of GLT names
- glts: A list of GLT objects
- gltstr: A list of GLT strings
- censor: The censoring vector
}
\description{
This function constructs the command string and associated options required
to run 3dDeconvolve using the specified fMRI model, dataset, working directory,
and other options. This command string can then be used to perform the actual
fMRI analysis using the AFNI software.
}
\keyword{internal}
