% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{build_event_model}
\alias{build_event_model}
\title{Build event model from configuration components}
\usage{
build_event_model(
  events_data,
  events_spec,
  regressors_spec,
  hrfs,
  sampling_frame
)
}
\arguments{
\item{events_data}{Event data from bidser::read_events}

\item{events_spec}{Event specification from YAML}

\item{regressors_spec}{Regressor specification from YAML}

\item{hrfs}{List of HRF objects}

\item{sampling_frame}{Time series sampling frame}
}
\value{
An event_model object
}
\description{
Build event model from configuration components
}
\keyword{internal}
