% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{cells.baseline_model}
\alias{cells.baseline_model}
\title{Retrieve Cells of a Baseline Model.}
\usage{
\method{cells}{baseline_model}(x, ...)
}
\arguments{
\item{x}{A baseline_model object.}
}
\value{
A tibble with columns: term, level, basis, and index.
}
\description{
Combines the cells from all baseline terms into a tibble with an index.
}
