% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_vector.R
\name{cells}
\alias{cells}
\alias{cells.event_factor}
\alias{cells.event_term}
\title{cells}
\usage{
cells(x, ...)

\method{cells}{event_factor}(x, drop.empty = TRUE)

\method{cells}{event_term}(x, drop.empty = TRUE)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
return the experimental cells that are in a model term
}
