% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{check_dir_exists}
\alias{check_dir_exists}
\title{Check if a string field points to an existing directory.}
\usage{
check_dir_exists(data, field, path, errors)
}
\description{
Check if a string field points to an existing directory.
}
\keyword{internal}
