% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{check_enum}
\alias{check_enum}
\title{Check if a field's value(s) are within an allowed set.}
\usage{
check_enum(data, field, allowed_values, path, errors, allow_null = FALSE)
}
\description{
Check if a field's value(s) are within an allowed set.
}
\keyword{internal}
