% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{check_pattern}
\alias{check_pattern}
\title{Check if a string field's value(s) match a regex pattern.}
\usage{
check_pattern(data, field, pattern, path, errors, allow_null = FALSE)
}
\description{
Check if a string field's value(s) match a regex pattern.
}
\keyword{internal}
