% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_latent_lm.R
\name{chunkwise_lm.latent_dataset}
\alias{chunkwise_lm.latent_dataset}
\title{Internal function for performing chunkwise linear regression on latent datasets}
\usage{
\method{chunkwise_lm}{latent_dataset}(
  dset,
  model,
  conlist,
  nchunks,
  robust = FALSE,
  verbose = FALSE,
  autocor = c("auto", "ar1", "ar2", "arma", "none"),
  bootstrap = FALSE,
  nboot = 1000,
  boot_rows = FALSE
)
}
\arguments{
\item{dset}{A latent dataset object.}

\item{model}{The fmri model object.}

\item{conlist}{A list of contrasts.}

\item{nchunks}{The number of chunks to use for the regression.}

\item{robust}{Logical, if TRUE, robust regression will be performed.}

\item{verbose}{Logical, if TRUE, additional output will be printed.}

\item{autocor}{A character string specifying the autocorrelation method to use.}

\item{bootstrap}{Logical, if TRUE, bootstrapping will be performed.}

\item{nboot}{The number of bootstrap iterations.}

\item{boot_rows}{Logical, if TRUE, bootstrap row-wise.}
}
\value{
A list containing the results of the chunkwise linear regression.
}
\description{
Internal function for performing chunkwise linear regression on latent datasets
}
\details{
This function is an internal helper function for fmri_latent_lm, which performs
chunkwise linear regression on latent datasets. The function handles different
autocorrelation options, as well as robust regression and bootstrapping.
}
\seealso{
fmri_latent_lm
}
\keyword{internal}
