% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{chunkwise_lm}
\alias{chunkwise_lm}
\title{estimate a linear model sequentially for each "chunk" (a matrix of time-series) of data}
\usage{
chunkwise_lm(x, ...)
}
\arguments{
\item{x}{the dataset}

\item{...}{extra args}
}
\description{
estimate a linear model sequentially for each "chunk" (a matrix of time-series) of data
}
