% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{coef.fmri_meta}
\alias{coef.fmri_meta}
\title{Extract Coefficients from Meta-Analysis}
\usage{
\method{coef}{fmri_meta}(object, ...)
}
\arguments{
\item{object}{An fmri_meta object}

\item{...}{Additional arguments}
}
\value{
Matrix of coefficients
}
\description{
Extract Coefficients from Meta-Analysis
}
