% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_ttest_helpers.R
\name{coef_image.fmri_ttest_fit}
\alias{coef_image.fmri_ttest_fit}
\title{Extract Coefficient Image from fmri_ttest_fit}
\usage{
\method{coef_image}{fmri_ttest_fit}(object, coef = 1, statistic = c("estimate", "se", "z", "p"), ...)
}
\arguments{
\item{object}{An fmri_ttest_fit object}

\item{coef}{Character or integer; coefficient to extract}

\item{statistic}{Character string; type of statistic to extract:
\itemize{
\item "estimate": Coefficient estimates (beta values)
\item "se": Standard errors (if available)
\item "z": Z-scores
\item "p": P-values
}}

\item{...}{Additional arguments (e.g., mask to apply)}
}
\value{
NeuroVol object or numeric vector
}
\description{
Creates a NeuroVol image from coefficients in an fmri_ttest_fit object.
}
