% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{columns.ParametricBasis}
\alias{columns.ParametricBasis}
\alias{columns.Scale}
\alias{columns.ScaleWithin}
\alias{columns.RobustScale}
\title{Get Column Name Information for ParametricBasis Objects}
\usage{
\method{columns}{Scale}(x)

\method{columns}{ScaleWithin}(x)

\method{columns}{RobustScale}(x)
}
\description{
Get Column Name Information for ParametricBasis Objects
}
