% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hrf.R
\name{compute_best_hrfs}
\alias{compute_best_hrfs}
\title{Compute Best HRFs}
\usage{
compute_best_hrfs(
  X_proj,
  L_proj,
  L,
  cluster_series = TRUE,
  cluster_quantile = 0.5,
  dist_method = "euclidean"
)
}
\arguments{
\item{X_proj}{Typically the projection of the data}

\item{L_proj}{The projection of the library}

\item{L}{The full library matrix (rows= time, cols= candidates)}

\item{cluster_series}{If TRUE, we cluster with cutree}

\item{cluster_quantile}{The distance threshold quantile}

\item{dist_method}{The distance measure (by default "euclidean")}
}
\description{
This is an internal helper that tries to match each column of X_proj
(representing data-based PCs) to columns of L_proj (the library's PCs),
and optionally performs a hierarchical clustering to group them.
}
\keyword{internal}
