% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_specs.R
\name{compute_spec}
\alias{compute_spec}
\title{Mechanical execution specification}
\usage{
compute_spec(
  voxel_chunks = 10L,
  backend = c("matrix", "reference"),
  parallel = c("none", "voxels", "chunks"),
  progress = FALSE
)
}
\arguments{
\item{voxel_chunks}{Number of voxel partitions for joint fitting.}

\item{backend}{Matrix or formula/reference implementation.}

\item{parallel}{Parallelization dimension.}

\item{progress}{Display progress.}
}
\description{
Mechanical execution specification
}
