% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{conditions.event_term}
\alias{conditions.event_term}
\title{Extract conditions from an event_term.}
\usage{
\method{conditions}{event_term}(x, drop.empty = TRUE, ...)
}
\arguments{
\item{x}{An event_term object.}

\item{drop.empty}{Logical; if TRUE, conditions with no cells are dropped (default is TRUE).}

\item{...}{Additional arguments.}
}
\value{
A character vector of condition labels.
}
\description{
Constructs a vector of condition labels by combining the levels of
the categorical events.
}
