% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R,
%   R/event_vector.R
\name{conditions}
\alias{conditions}
\alias{conditions.event_model}
\alias{conditions.fmri_term}
\alias{conditions.convolved_term}
\alias{conditions.afni_hrf_convolved_term}
\alias{conditions.afni_trialwise_convolved_term}
\alias{conditions.event_term}
\title{Conditions}
\usage{
conditions(x, ...)

\method{conditions}{event_model}(x, ...)

\method{conditions}{fmri_term}(x, ...)

\method{conditions}{convolved_term}(x, ...)

\method{conditions}{afni_hrf_convolved_term}(x, ...)

\method{conditions}{afni_trialwise_convolved_term}(x, ...)

\method{conditions}{event_term}(x, drop.empty = TRUE, ...)
}
\arguments{
\item{x}{the model term}

\item{...}{extra args}
}
\description{
return the set of condition labels associated with a model term
}
