% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{conform_len}
\alias{conform_len}
\title{check vector is of length 1 or repeated for supplied length}
\usage{
conform_len(val, len)
}
\description{
check vector is of length 1 or repeated for supplied length
}
\keyword{internal}
