% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{construct}
\alias{construct}
\title{construct}
\usage{
construct(x, model_spec, ...)
}
\arguments{
\item{x}{the object}

\item{model_spec}{the model specification}
}
\description{
construct a term given a an hrf spec and model specification
}

