% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{construct_model}
\alias{construct_model}
\title{Construct an Event Model.}
\usage{
construct_model(x)
}
\arguments{
\item{x}{A model_spec list.}
}
\value{
An event_model object.
}
\description{
Given a model_spec list (constructed by event_model.list or event_model.formula),
this function constructs the final event model.
}
\keyword{internal}
