% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridesign-imports.R
\name{contrast_set}
\alias{contrast_set}
\title{Create a contrast set}
\usage{
contrast_set(...)
}
\arguments{
\item{...}{contrast specifications}
}
\value{
A list of class "contrast_set" containing the specified contrasts
}
\description{
Create a contrast set
}
