% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast_weights.poly_contrast_spec}
\alias{contrast_weights.poly_contrast_spec}
\title{Polynomial Contrast Weights}
\usage{
\method{contrast_weights}{poly_contrast_spec}(x, term, ...)
}
\arguments{
\item{x}{A poly_contrast_spec object.}

\item{term}{A term object.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list containing the polynomial contrast weights and related information.
}
\description{
Compute the contrast weights for a polynomial contrast specification.
}
