% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{contrasts.event_term}
\alias{contrasts.event_term}
\title{Retrieve contrast definitions for an event term}
\usage{
contrasts.event_term(x, ...)
}
\arguments{
\item{x}{An \code{event_term} object.}

\item{...}{Unused.}
}
\value{
A list of contrast specifications or \code{NULL} when none are defined.
}
\description{
This accessor returns the list of contrast specifications attached to the
term's originating hrfspec, if any.
}
