% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-formula.R
\name{contrasts.hrfspec}
\alias{contrasts.hrfspec}
\title{Retrieve contrast specifications from an hrfspec}
\usage{
contrasts.hrfspec(x, ...)
}
\arguments{
\item{x}{An \code{hrfspec} object.}

\item{...}{Unused.}
}
\value{
The list of contrast specifications attached to the hrfspec, or \code{NULL}.
}
\description{
Retrieve contrast specifications from an hrfspec
}
