% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{convolve.event_term}
\alias{convolve.event_term}
\title{Convolve an event-related design matrix with an HRF.}
\usage{
\method{convolve}{event_term}(
  x,
  hrf,
  sampling_frame,
  drop.empty = TRUE,
  summate = TRUE,
  precision = 0.3,
  ...
)
}
\arguments{
\item{x}{A data frame with the design matrix.}

\item{hrf}{The HRF function.}

\item{sampling_frame}{Data frame specifying the sampling frame.}

\item{drop.empty}{Logical; if TRUE, remove empty rows.}

\item{summate}{Logical; if TRUE, sum the convolved HRF.}

\item{precision}{Numeric; the convolution precision (default: 0.3).}

\item{...}{Additional arguments.}
}
\value{
A tibble of the convolved design matrix.
}
\description{
This function takes an event-related design matrix and convolves it with a given
HRF to produce a new design matrix suitable for fMRI analysis.
}
