% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{convolve_block}
\alias{convolve_block}
\title{a convolve hemodynamic response with a block duration}
\usage{
convolve_block(t, hrf = hrf_gaussian, width = 5, precision = 0.1,
  summate = TRUE, ...)
}
\arguments{
\item{t}{time in seconds}

\item{hrf}{the underlying hemodynamic response function}

\item{width}{the fixed width of the response in seconds.}

\item{precision}{the sampling precision in seconds}
}
\description{
apply a hemodynamic response function with times \code{t} and duration \code{width}
}
