% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{convolve_design}
\alias{convolve_design}
\title{Convolve HRF with Design Matrix.}
\usage{
convolve_design(hrf, dmat, globons, durations, summate = TRUE)
}
\arguments{
\item{hrf}{A function representing the HRF.}

\item{dmat}{Design matrix (with named columns).}

\item{globons}{Numeric vector of global onsets.}

\item{durations}{Numeric vector of event durations.}

\item{summate}{Logical; if TRUE, summate the convolved HRF (default: TRUE).}
}
\value{
A list of regressors (one for each column).
}
\description{
Convolves a HRF with a design matrix (one column per condition) to produce a
list of regressors.
}
