% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{convolve_event_term}
\alias{convolve_event_term}
\title{Convolve Event Term with HRF}
\usage{
convolve_event_term(event_term_obj, hrf_obj, sampling_frame)
}
\arguments{
\item{event_term_obj}{An event term object created by \code{event_term()}.}

\item{hrf_obj}{An HRF object created by \code{create_hrf_object()}.}

\item{sampling_frame}{A data frame defining the time points at which to sample the convolved signal. Should contain a 'time' column.}
}
\value{
A numeric vector containing the convolved signal sampled at the times specified in \code{sampling_frame}.
}
\description{
This function convolves an event term with the specified HRF object over the sampling frame.
}
\examples{
# Assuming 'event_term_obj' and 'hrf_obj' have been created as in previous examples
sampling_frame <- data.frame(time = seq(0, 30, by = 1))
convolved_signal <- convolve_event_term(event_term_obj, hrf_obj, sampling_frame)

plot(sampling_frame$time, convolved_signal, type = 'l', xlab = 'Time', ylab = 'Signal')

}
