% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{correlation_map.event_model}
\alias{correlation_map.event_model}
\title{correlation_map.event_model}
\usage{
\method{correlation_map}{event_model}(
  x,
  method = c("pearson", "spearman"),
  half_matrix = FALSE,
  absolute_limits = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{event_model}.}

\item{method}{Correlation method, "pearson" or "spearman".}

\item{half_matrix}{Logical; if TRUE, mask out the upper-triangle.}

\item{absolute_limits}{Logical; if TRUE, fix color scale to -1..+1.}

\item{...}{Further arguments passed to \code{\link{.correlation_map_common}}.}
}
\description{
Generates a correlation heatmap of the columns in an \code{event_model}'s design matrix.
}
