% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{covariate}
\alias{covariate}
\title{Construct a Covariate Term}
\usage{
covariate(..., data, id = NULL, prefix = NULL)
}
\arguments{
\item{...}{A variable argument set of covariate names.}

\item{data}{A data.frame containing the variables.}

\item{id}{An optional identifier for the covariate term.}

\item{prefix}{An optional prefix to add to the covariate names.}
}
\value{
A list containing information about the covariate term.
}
\description{
Create a covariate term that is not convolved with a hemodynamic response function (HRF).
}
\examples{
df1 <- data.frame(x=rnorm(100), y=rnorm(100))
cv <- covariate(x, y, data=df1)
}
