% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{create_hrfspec}
\alias{create_hrfspec}
\title{Create an HRF specification programmatically}
\usage{
create_hrfspec(
  variable,
  hrf,
  onsets,
  durations = NULL,
  blockids = NULL,
  data = NULL
)
}
\description{
Create an HRF specification programmatically
}
\keyword{internal}
