% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{create_parametric_basis}
\alias{create_parametric_basis}
\title{Create parametric basis from specification}
\usage{
create_parametric_basis(basis_spec, data)
}
\description{
Create parametric basis from specification
}
\keyword{internal}
