% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{create_parametric_basis_expression}
\alias{create_parametric_basis_expression}
\title{create_parametric_basis_expression}
\usage{
create_parametric_basis_expression(basis_spec, variable_name)
}
\description{
create_parametric_basis_expression
}
\keyword{internal}
