% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{design_env}
\alias{design_env}
\title{design_env}
\usage{
design_env(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
return regression design as a set of matrices stored in an environment
}
