% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_smoothing_kernel.R
\name{design_kernel}
\alias{design_kernel}
\title{onsets <- sort(runif(25, 0, 200))
fac <- factor(sample(letters[1:4], length(onsets), replace=TRUE))
sframe <- sampling_frame(200, TR=1)
des <- data.frame(onsets=onsets, fac=fac, constant = factor(rep(1, length(fac))), block=rep(1, length(fac)))
emod <- event_model(onsets ~ hrf(constant) + hrf(fac), data=des, sampling_frame=sframe, block = ~ block)
dmat <- design_matrix(emod)}
\usage{
design_kernel(dmat, deriv = FALSE)
}
\description{
onsets <- sort(runif(25, 0, 200))
fac <- factor(sample(letters[1:4], length(onsets), replace=TRUE))
sframe <- sampling_frame(200, TR=1)
des <- data.frame(onsets=onsets, fac=fac, constant = factor(rep(1, length(fac))), block=rep(1, length(fac)))
emod <- event_model(onsets ~ hrf(constant) + hrf(fac), data=des, sampling_frame=sframe, block = ~ block)
dmat <- design_matrix(emod)
}
