% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{design_map.event_model}
\alias{design_map.event_model}
\title{Visualize the entire design matrix of an event_model as a heatmap}
\usage{
\method{design_map}{event_model}(
  x,
  rescale_cols = TRUE,
  block_separators = TRUE,
  rotate_x_text = TRUE,
  fill_midpoint = NULL,
  fill_limits = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{event_model} object.}

\item{rescale_cols}{Logical; if TRUE, columns are rescaled to (-1,1) before plotting.}

\item{block_separators}{Logical; if \code{TRUE}, draw white horizontal lines between blocks.}

\item{rotate_x_text}{Logical; if \code{TRUE}, rotate x-axis labels by 45 degrees.}

\item{fill_midpoint}{Numeric or \code{NULL}; if not \code{NULL}, passed as the \code{midpoint}
argument to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient2()}} to center the color scale. If
your design matrix has both positive and negative values, you may want
\code{fill_midpoint = 0}.}

\item{fill_limits}{Numeric vector of length 2 or \code{NULL}; passed to \verb{limits=}
in the fill scale. Useful to clip color range.}

\item{...}{Other arguments passed on to \code{geom_tile()}.}
}
\value{
A ggplot2 plot object.
}
\description{
Produces a single heatmap of all columns in the design matrix for an
\code{event_model} object, with rows corresponding to scan number and columns
corresponding to regressors. A white horizontal line separates each block/run.
}
