% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{design_matrix.event_term}
\alias{design_matrix.event_term}
\title{Construct a design matrix for an event_term.}
\usage{
\method{design_matrix}{event_term}(x, drop.empty = TRUE, ...)
}
\arguments{
\item{x}{An event_term object.}

\item{drop.empty}{Logical; if TRUE, columns with no events are removed.}

\item{...}{Additional arguments.}
}
\value{
A tibble representing the design matrix.
}
\description{
This function creates a design matrix from an event_term object by first
populating a local environment with event values and then applying model.matrix
to the formula derived from the event_term.
}
