% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.R
\name{design_plot}
\alias{design_plot}
\title{design_plot}
\usage{
design_plot(fmrimod, longnames = FALSE)
}
\arguments{
\item{fmrimod}{the \code{fmri_model} object}

\item{longnames}{use longnames in legend}
}
\description{
design_plot
}
