% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{detect_group_data_format}
\alias{detect_group_data_format}
\title{Detect Input Format for Group Data}
\usage{
detect_group_data_format(data)
}
\arguments{
\item{data}{Input data to detect format from}
}
\value{
Character string indicating detected format
}
\description{
Detect Input Format for Group Data
}
\keyword{internal}
