% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{.all_symbols_in_call}
\alias{.all_symbols_in_call}
\title{Recursively find all symbol names used inside a call/expression.}
\usage{
.all_symbols_in_call(call)
}
\arguments{
\item{call}{An R call object or symbol}
}
\value{
A character vector of bare variable names
}
\description{
For example, given \code{Poly(modulator, 2)}, returns \code{c("modulator")}.
}
\keyword{internal}
