% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.compute_beta_volumes}
\alias{.compute_beta_volumes}
\title{Compute Beta Volumes from fmri_lm Object}
\usage{
.compute_beta_volumes(fmrilm_obj)
}
\description{
Compute Beta Volumes from fmri_lm Object
}
\keyword{internal}
