% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.compute_statistical_volumes}
\alias{.compute_statistical_volumes}
\title{Compute Statistical Map Volumes from Contrast Data}
\usage{
.compute_statistical_volumes(contrast_data, stat, brain_dims, mask, space)
}
\description{
Compute Statistical Map Volumes from Contrast Data
}
\keyword{internal}
