% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.extract_contrast_type}
\alias{.extract_contrast_type}
\title{Extract Contrast Type Information}
\usage{
.extract_contrast_type(contrast_row, fmrilm_obj)
}
\description{
Extract Contrast Type Information
}
\keyword{internal}
