% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.extract_degrees_of_freedom}
\alias{.extract_degrees_of_freedom}
\title{Extract Degrees of Freedom from fmri_lm Object}
\usage{
.extract_degrees_of_freedom(fmrilm_obj)
}
\description{
Extract Degrees of Freedom from fmri_lm Object
}
\keyword{internal}
