% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{.extract_expr_variables_into_events}
\alias{.extract_expr_variables_into_events}
\title{Detect variables inside exprs(...) calls and copy those columns into \code{events}.}
\usage{
.extract_expr_variables_into_events(event_terms, events)
}
\arguments{
\item{event_terms}{The list of event term specifications}

\item{events}{A data.frame of event variables}
}
\value{
The updated \code{events} data frame, ensuring all variables found in expressions exist as columns
}
\description{
Detect variables inside exprs(...) calls and copy those columns into \code{events}.
}
\keyword{internal}
