% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.extract_source_files}
\alias{.extract_source_files}
\title{Extract Source Files from fmri_lm Object}
\usage{
.extract_source_files(fmrilm_obj)
}
\description{
Extract Source Files from fmri_lm Object
}
\keyword{internal}
