% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.finalize_atomic_write}
\alias{.finalize_atomic_write}
\title{Finalize Atomic Write by Moving Files}
\usage{
.finalize_atomic_write(temp_dir, final_path, created_files)
}
\description{
Finalize Atomic Write by Moving Files
}
\keyword{internal}
