% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_plugin_api.R
\docType{data}
\name{.fmrireg_engine_registry}
\alias{.fmrireg_engine_registry}
\title{Plugin registration and helpers for fmrireg}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.fmrireg_engine_registry
}
\description{
Plugin registration and helpers for fmrireg
}
\keyword{datasets}
