% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.generate_bids_filename}
\alias{.generate_bids_filename}
\title{Generate BIDS-Compliant Filename}
\usage{
.generate_bids_filename(
  entities,
  desc = NULL,
  contrast = NULL,
  stat = NULL,
  suffix = NULL,
  extension = "h5"
)
}
\description{
Generate BIDS-Compliant Filename
}
\keyword{internal}
