% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.map_stat_name}
\alias{.map_stat_name}
\title{Map User Statistic Names to Internal Field Names}
\usage{
.map_stat_name(stat)
}
\description{
Map User Statistic Names to Internal Field Names
}
\keyword{internal}
