% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_engine_lowrank.R
\name{.run_lowrank_engine}
\alias{.run_lowrank_engine}
\title{Internal: run low-rank/sketched engine under fmri_lm}
\usage{
.run_lowrank_engine(fm, dataset, lowrank, ar_options = NULL)
}
\description{
Internal: run low-rank/sketched engine under fmri_lm
}
\keyword{internal}
