% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{.sanitizeName}
\alias{.sanitizeName}
\title{Sanitize a variable name.}
\usage{
.sanitizeName(name)
}
\arguments{
\item{name}{A character string.}
}
\value{
A sanitized version of the input name.
}
\description{
This function replaces problematic characters (such as colons, spaces,
parentheses, and commas) to produce a valid name. For potential performance
improvements, caching of sanitized names could be implemented.
}
\keyword{internal}
