% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.save_betas_json_metadata}
\alias{.save_betas_json_metadata}
\title{Save JSON Metadata for Betas}
\usage{
.save_betas_json_metadata(
  fmrilm_obj,
  path,
  entities,
  desc,
  regressor_names,
  h5_filepath
)
}
\description{
Save JSON Metadata for Betas
}
\keyword{internal}
