% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.save_contrasts_json_metadata}
\alias{.save_contrasts_json_metadata}
\title{Save JSON Metadata for Contrasts (by stat)}
\usage{
.save_contrasts_json_metadata(
  fmrilm_obj,
  path,
  entities,
  desc,
  stat,
  contrast_names,
  contrasts,
  h5_filepath
)
}
\description{
Save JSON Metadata for Contrasts (by stat)
}
\keyword{internal}
