% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.save_regressor_betas}
\alias{.save_regressor_betas}
\title{Save Raw Regressor Betas using LabeledVolumeSet}
\usage{
.save_regressor_betas(fmrilm_obj, path, entities, desc, overwrite)
}
\description{
Save Raw Regressor Betas using LabeledVolumeSet
}
\keyword{internal}
