% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.save_single_contrast_json_metadata}
\alias{.save_single_contrast_json_metadata}
\title{Save JSON Metadata for Single Contrast}
\usage{
.save_single_contrast_json_metadata(
  fmrilm_obj,
  path,
  entities,
  desc,
  contrast_name,
  stat_names,
  contrast_row,
  h5_filepath
)
}
\description{
Save JSON Metadata for Single Contrast
}
\keyword{internal}
