% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-formula.R
\name{.trial_factor}
\alias{.trial_factor}
\title{Internal helper for generating trial factors}
\usage{
.trial_factor(n)
}
\arguments{
\item{n}{Length of the factor to generate.}
}
\description{
Internal helper for generating trial factors
}
\keyword{internal}
