% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_export.R
\name{.validate_required_entities}
\alias{.validate_required_entities}
\title{Validate Required BIDS Entities}
\usage{
.validate_required_entities(entities)
}
\description{
Validate Required BIDS Entities
}
\keyword{internal}
