% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{elements.event_term}
\alias{elements.event_term}
\title{Extract Elements (Values or Labels) from an Event Term}
\usage{
\method{elements}{event_term}(x, what = c("values", "labels"), ...)
}
\arguments{
\item{what}{Character, either "values" to return numeric data or "labels" for descriptive names.}

\item{...}{Additional arguments passed down (e.g., \code{transformed} for \code{elements.event}).}
}
\description{
Extract Elements (Values or Labels) from an Event Term
}
