% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{estcon.contrast}
\alias{estcon.contrast}
\title{Estimated Contrast}
\usage{
\method{estcon}{contrast}(x, fit, indices, ...)
}
\arguments{
\item{x}{The contrast to estimate.}

\item{fit}{The fit object.}

\item{indices}{The indices to use.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The estimated contrast.
}
\description{
Compute the estimated contrast for a given fit and indices.
}
