% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_betas.R
\name{estimate_betas.fmri_dataset}
\alias{estimate_betas.fmri_dataset}
\title{estimate trialwise betas for an fMRI dataset.}
\usage{
\method{estimate_betas}{fmri_dataset}(
  dataset,
  fixed,
  ran,
  block,
  method = c("mixed", "pls", "pls_searchlight", "slm"),
  basedeg = 5,
  nuisance_list = NULL,
  radius = 8,
  niter = 8,
  ncomp = 4,
  lambda = 0.01
)
}
\arguments{
\item{dataset}{the fmri_dataset instance}

\item{fixed}{the fixed factors}

\item{ran}{the random (trialwise) factors}

\item{block}{the block factor}

\item{basedeg}{number of baseline dmodels degrees of freedom}

\item{nuisance_list}{a list of matrices containing nuisance regressors, one matrix per block.}

\item{niter}{number of searchlight iterations for method "pls_searchlight"}

\item{ncomp}{number of pls components for method "pls" and "pls_searchlight"}

\item{lambda}{lambda parameter (not currently used)}
}
\description{
estimate trialwise betas for an fMRI dataset.
}
\examples{

facedes <- read.table(system.file("extdata", "face_design.txt", package = "fmrireg"), header=TRUE)
facesdes$frun <- factor(facedes$run)
scans <- paste0("rscan0", 1:6, ".nii")

dset <- fmri_dataset(scans=scans,mask="mask.nii",TR=1.5,run_length=rep(436,6),event_table=facedes)
fixed = onset ~ hrf(run)
ran = onset ~ trialwise()
block = ~ run
}
