% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{estimate_contrast}
\alias{estimate_contrast}
\title{estimate contrast}
\usage{
estimate_contrast(x, fit, colind, ...)
}
\arguments{
\item{x}{the contrast}

\item{fit}{the model fit}

\item{colind}{the subset of column indices in the design matrix}

\item{...}{extra args}
}
\description{
estimate contrast
}
\concept{estimate_contrast}
