% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank1_estimation.R
\name{estimate_r1}
\alias{estimate_r1}
\title{keywords internal}
\usage{
estimate_r1(
  dset,
  xdat,
  hrf_basis,
  hrf_ref,
  maxit = 100,
  flip_sign = TRUE,
  use_box_constraints = FALSE
)
}
\description{
keywords internal
}
