% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta.R
\name{estimate_tau2}
\alias{estimate_tau2}
\title{Estimate Between-Study Variance (tau2)}
\usage{
estimate_tau2(y, se, X, method)
}
\arguments{
\item{y}{Effect sizes}

\item{se}{Standard errors}

\item{X}{Design matrix}

\item{method}{Estimation method}
}
\value{
tau2 estimate
}
\description{
Estimate Between-Study Variance (tau2)
}
\keyword{internal}
