% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{EV}
\alias{EV}
\title{EV}
\usage{
EV(vals, name, onsets, blockids, durations = 1)
}
\arguments{
\item{vals}{the event values}

\item{name}{the name of the event variable}

\item{onsets}{the event onsets.}

\item{blockids}{the block ids associated with each event (must be non-decreasing)}
}
\description{
factory function for creating 'event' types: event_factor, event_variable, event_basis, event_matrix.
}

