% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{evaluate.hrfspec}
\alias{evaluate.hrfspec}
\title{evaluate.hrfspec}
\usage{
\method{evaluate}{hrfspec}(x, grid, amplitude = 1, duration = 0, precision = 0.1, ...)
}
\arguments{
\item{x}{The hrfspec object containing the HRF function.}

\item{grid}{A vector of time points.}

\item{amplitude}{The scaling value for the event (default: 1).}

\item{duration}{The duration of the event (default: 0).}

\item{precision}{The temporal resolution used for computing summed responses when duration > 0 (default: 0.1).}

\item{...}{Additional arguments to be passed to the evaluate.HRF function.}
}
\value{
A vector of HRF values at the specified time points.
}
\description{
This function evaluates a hemodynamic response function (HRF) specified by an hrfspec object for a given set of time points (grid) and other parameters.
It is a wrapper function that calls the evaluate.HRF function with the HRF function contained in the hrfspec object.
}
\examples{
hrf_spec <- hrfspec(hrf = HRF_SPMG1)
hrf_values <- evaluate(hrfspec, grid=seq(0,20,by=1.5), duration=2, precision=.1)
}
