% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/regressor.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.regressor}
\title{evaluate a function over a sampling grid}
\usage{
evaluate(x, grid, ...)

\method{evaluate}{regressor}(x, grid, precision = 0.33, use_conv = FALSE, ...)
}
\arguments{
\item{x}{the object to evaluate}

\item{grid}{the sampling grid. A vector of real values in seconds.}

\item{...}{extra args}

\item{precision}{the sampling precision for the hrf. This parameter is passed to \code{evaluate.HRF}}

\item{use_conv}{use fast convolution approach}
}
\description{
given an object to be evaluated and an input sample "grid", evaluate the object.

evalute a regressor function over a sampling grid
}
\examples{
frame <- sampling_frame(blocklens=100, TR=2)
reg <- regressor(onsets=c(10,20,35, 47,52, 68, 79,86), amp=runif(8), duration=runif(8)*3, hrf=HRF_SPMG1)
e1 = evaluate(reg, samples(frame))
e2 = evaluate(reg, samples(frame), use_conv=TRUE)
}
