% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_factor}
\alias{event_factor}
\title{event_factor}
\usage{
event_factor(fac, name, onsets, blockids = 1, durations = NULL)
}
\arguments{
\item{fac}{}

\item{name}{}

\item{onsets}{}

\item{blockids}{}

\item{durations}{}
}
\description{
Create an categorical event sequence from a \code{factor}
}

