% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{event_model}
\alias{event_model}
\title{event_model}
\usage{
event_model(formula, data, block, sampling_frame, drop_empty = TRUE,
  durations = 0, contrasts = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{data}{the data containing experimental design}

\item{block}{formula for the block structure}

\item{sampling_frame}{the sampling frame defining the temporal and block structure}

\item{drop_empty}{whether to drop empty factor levels}

\item{durations}{the event durations}

\item{contrasts}{the set of contrasts}
}
\description{
event_model
}
