<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_set}
\alias{event_set}
\title{event_set}
\usage{
event_set(mat, name, onsets, durations = NULL, blockids = 1)
}
\arguments{
\item{mat}{a matrix of values, one row per event, indicating the amplitude/intensity of each event.}

\item{name}{the name of the variable}

\item{onsets}{the event onsets in seconds}

\item{durations}{the durations of each event in seconds}

\item{blockids}{the index of the block/scan in which the event occurs}
}
\description{
Create a continuous valued event set from a \code{matrix}
}
\examples{
mat <- matrix(rnorm(200), 100, 2)
onsets <- seq(1, 1000, length.out=100)
durations <- rep(1, 100)
blockids <- rep(1, 100)

eset <- event_set(mat,onsets,durations,blockids)
}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_set}
\alias{event_set}
\usage{
event_set(mat, name, onsets, durations = NULL, blockids = 1)
}
\description{
Create a continuous valued event set from a \code{matrix}
}
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc

