% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{event_table}
\alias{event_table}
\alias{event_table.convolved_term}
\title{event_table}
\usage{
event_table(x)

\method{event_table}{convolved_term}(x)
}
\arguments{
\item{x}{the term}
}
\description{
extract event_table as \code{data.frame}
}
