<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc
% Please edit documentation in R/event_vector.R
\name{event_term}
\alias{event_term}
\title{event_term}
\usage{
event_term(evlist, onsets, blockids, durations = 1, subset = NULL)
}
<<<<<<< HEAD
\arguments{
\item{onsets}{the onset times from the experimental events in seconds.}

\item{blockids}{the block number associated whcih each onset, should be in strict increasing order.}
}
\description{
construct an \code{event_term} instance from a named list of variables
=======
\description{
Create a event model term from a list of variables.
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc
}

