% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{event_terms}
\alias{event_terms}
\title{event_terms}
\usage{
event_terms(x)
}
\arguments{
\item{x}{the model object}
}
\value{
A list containing the 'event_term's.
}
\description{
This function extracts a list of 'event_term's from the provided model object \code{x}.
}
\details{
Extract a list of 'event_term's from the model object.
}
