% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nystrom_extend.R
\name{extend_betas_landmarks}
\alias{extend_betas_landmarks}
\title{Apply Nyström / barycentric extension: B (p x V) from B_L (p x L)}
\usage{
extend_betas_landmarks(BL, W)
}
\arguments{
\item{BL}{p x L dense matrix of betas on landmarks}

\item{W}{V x L sparse weight matrix (rows sum to 1)}
}
\value{
p x V dense matrix of extended betas
}
\description{
Apply Nyström / barycentric extension: B (p x V) from B_L (p x L)
}
\keyword{internal}
