% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_h5.R
\name{extract_subject_ids_from_paths}
\alias{extract_subject_ids_from_paths}
\title{Extract Subject IDs from File Paths}
\usage{
extract_subject_ids_from_paths(paths)
}
\arguments{
\item{paths}{Character vector of file paths}
}
\value{
Character vector of subject IDs
}
\description{
Extract Subject IDs from File Paths
}
\keyword{internal}
