% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{Fcontrasts.event_term}
\alias{Fcontrasts.event_term}
\title{Compute F-contrasts for Event Term}
\usage{
\method{Fcontrasts}{event_term}(x, ...)
}
\arguments{
\item{x}{An event term object}

\item{...}{Additional arguments passed to the function}
}
\value{
A list of contrast matrices for main effects and interactions
}
\description{
This function computes F-contrasts for an event term object, considering main effects and interactions.
}
