% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hrf.R
\name{find_best_hrf}
\alias{find_best_hrf}
\title{Generic S3 method for \code{find_best_hrf}}
\usage{
find_best_hrf(
  dataset,
  onset_var,
  hrflib = NULL,
  rsam = seq(0, 24, by = 1),
  ncomp = 5,
  nsplits = 3,
  block = NULL,
  basemod = NULL,
  cluster_series = TRUE,
  cluster_quantile = 0.5,
  dist_method = "euclidean",
  spatial_smoothing = FALSE,
  lambda = 0.1,
  n_iterations = 5,
  k_neighbors = 6,
  nboot = 1,
  ...
)
}
\arguments{
\item{dataset}{The dataset}

\item{onset_var}{The name of the onset variable}

\item{hrflib}{Possibly a function returning a matrix of candidate HRFs}

\item{rsam}{Time points at which HRFs are sampled}

\item{ncomp}{Number of basis components}

\item{nsplits}{Number of splits}

\item{block}{Possibly a formula specifying block structure}

\item{basemod}{Possibly a baseline model}

\item{cluster_series}{If TRUE, cluster the assigned HRFs}

\item{cluster_quantile}{The quantile used in the clustering}

\item{dist_method}{The distance measure}

\item{spatial_smoothing}{If TRUE, smooth HRF assignments in space}

\item{lambda}{Smoothing param}

\item{n_iterations}{Number of smoothing iterations}

\item{k_neighbors}{How many neighbors to consider}

\item{nboot}{Number of bootstrap iterations}

\item{...}{Other args}
}
\description{
This function is a generic for finding best HRFs across a dataset.
Usually, you implement \code{find_best_hrf.matrix_dataset} or
\code{find_best_hrf.fmri_mem_dataset} depending on your data type.
}
