% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_stats.R
\name{fit_contrasts.default}
\alias{fit_contrasts.default}
\title{Fit Contrasts for Linear Model (Default Method)}
\usage{
\method{fit_contrasts}{default}(object, conmat, colind, se = TRUE, ...)
}
\arguments{
\item{object}{The fitted linear model object.}

\item{conmat}{The contrast matrix or contrast vector.}

\item{colind}{The subset column indices in the design associated with the contrast.}

\item{se}{Whether to compute standard errors, t-statistics, and p-values (default: TRUE).}

\item{...}{Additional arguments (unused)}
}
\value{
A list containing the following elements:
\itemize{
\item{\code{conmat}: Contrast matrix.}
\item{\code{sigma}: Residual standard error.}
\item{\code{df.residual}: Degrees of freedom for residuals.}
\item{\code{estimate}: Estimated contrasts.}
\item{\code{se}: Standard errors of the contrasts (if \code{se = TRUE}).}
\item{\code{stat}: t-statistics for the contrasts (if \code{se = TRUE}).}
\item{\code{prob}: Probabilities associated with the t-statistics (if \code{se = TRUE}).}
\item{\code{stat_type}: Type of the statistics calculated.}
}
}
\description{
This function calculates contrasts for a fitted linear model.
}
