% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_stats.R
\name{fit_contrasts.fmri_lm}
\alias{fit_contrasts.fmri_lm}
\title{Fit Contrasts for fMRI Linear Model Objects}
\usage{
\method{fit_contrasts}{fmri_lm}(object, contrasts, ...)
}
\arguments{
\item{object}{An fmri_lm object}

\item{contrasts}{A list of contrast specifications}

\item{...}{Additional arguments (unused)}
}
\value{
A list of contrast results
}
\description{
S3 method for computing contrasts on fitted fmri_lm objects.
}
