% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_stats.R
\name{fit_Fcontrasts}
\alias{fit_Fcontrasts}
\title{Fit F-contrasts for Linear Model}
\usage{
fit_Fcontrasts(lmfit, conmat, colind)
}
\arguments{
\item{lmfit}{Fitted linear model object.}

\item{conmat}{Contrast matrix.}

\item{colind}{Column indices corresponding to the variables in the contrast matrix.}
}
\value{
A list containing the following elements:
\itemize{
\item{\code{estimate}: Estimated contrasts.}
\item{\code{se}: Residual variance.}
\item{\code{stat}: F-statistics for the contrasts.}
\item{\code{prob}: Probabilities associated with the F-statistics.}
\item{\code{stat_type}: Type of the statistics calculated.}
}
}
\description{
This function calculates F-contrasts for a fitted linear model.
}
\examples{
data(mtcars)
lm_fit <- lm(mpg ~ wt + qsec + am, data = mtcars)
contrast_matrix <- matrix(c(0, -1, 1), nrow = 1, byrow = TRUE)
fit_Fcontrasts(lm_fit, contrast_matrix, c(2, 3))
}
\keyword{internal}
