% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fit_lm_contrasts}
\alias{fit_lm_contrasts}
\title{Fit Linear Model Contrasts}
\usage{
fit_lm_contrasts(fit, conlist, fcon, vnames, se = TRUE)
}
\arguments{
\item{fit}{A fitted linear model object.}

\item{conlist}{A list of contrast matrices.}

\item{fcon}{A list of F-contrasts.}

\item{vnames}{Variable names corresponding to the model coefficients.}

\item{se}{Logical. Whether to compute standard errors. Default is \code{TRUE}.}
}
\value{
A list containing contrasts, beta statistics, and the fitted model.
}
\description{
This function computes contrasts and beta statistics for a fitted linear model.
}
\keyword{internal}
