% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta.R
\name{fit_meta_chunk}
\alias{fit_meta_chunk}
\title{Fit Meta-Analysis for a Chunk of Voxels}
\usage{
fit_meta_chunk(
  chunk_data,
  X,
  method,
  robust,
  weights,
  weights_custom,
  n_threads,
  combine = NULL,
  contrasts = NULL,
  return_cov = NULL
)
}
\arguments{
\item{chunk_data}{List with beta and se matrices}

\item{X}{Design matrix}

\item{method}{Meta-analysis method}

\item{robust}{Robust estimation method}

\item{weights}{Weighting scheme}

\item{weights_custom}{Custom weights}

\item{n_threads}{Number of threads}
}
\value{
List with meta-analysis results
}
\description{
Fit Meta-Analysis for a Chunk of Voxels
}
\keyword{internal}
