% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta.R
\name{fit_meta_single}
\alias{fit_meta_single}
\title{Fit Meta-Analysis for a Single Voxel/ROI}
\usage{
fit_meta_single(y, se, X, method, robust, weights, weights_custom)
}
\arguments{
\item{y}{Effect sizes}

\item{se}{Standard errors}

\item{X}{Design matrix}

\item{method}{Meta-analysis method}

\item{robust}{Robust estimation method}

\item{weights}{Weighting scheme}

\item{weights_custom}{Custom weights}
}
\value{
List with meta-analysis results
}
\description{
Fit Meta-Analysis for a Single Voxel/ROI
}
\keyword{internal}
