% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_methods.R, R/fmrilm.R
\name{fitted_hrf.fmri_lm}
\alias{fitted_hrf.fmri_lm}
\title{S3 Methods for fmri_lm Objects}
\usage{
\method{fitted_hrf}{fmri_lm}(x, sample_at = seq(0, 24, by = 1), ...)

\method{fitted_hrf}{fmri_lm}(x, sample_at = seq(0, 24, by = 1), ...)
}
\arguments{
\item{x}{An \code{fmri_lm} object for which the fitted HRFs should be computed.}

\item{sample_at}{A numeric vector of time points at which the HRFs should be sampled. Default is \code{seq(0, 24, by = 1)}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame with the fitted HRF

A list where each element corresponds to an event term in the \code{fmri_lm} object. Each element contains:
\describe{
\item{\code{pred}}{A matrix of predicted HRF values.}
\item{\code{design}}{A tibble containing the design matrix for the HRFs.}
}
}
\description{
Methods for extracting results and information from fitted fmri_lm objects

This method computes the fitted hemodynamic response functions (HRFs) for an \code{fmri_lm} object.
}
\keyword{internal}
