% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fmri_benchmark_datasets.R
\docType{data}
\name{fmri_benchmark_datasets}
\alias{fmri_benchmark_datasets}
\title{Benchmark fMRI datasets}
\format{A list with elements \code{BM_Canonical_HighSNR}, \code{BM_Canonical_LowSNR}, \code{BM_HRF_Variability_AcrossVoxels}, \code{BM_Trial_Amplitude_Variability}, \code{BM_Complex_Realistic}, and \code{metadata}. Each element contains simulated BOLD data and ground truth information.}
\source{Generated by \code{data-raw/generate_benchmark_datasets.R}}
\description{
A list of simulated datasets used for testing the package.
}
