% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fmri_glm}
\alias{fmri_glm}
\title{fmri_glm}
\usage{
fmri_glm(formula, block_formula, baseline_model = NULL, dataset, durations,
  drop_empty = TRUE, contrasts = NULL, strategy = c("runwise",
  "slicewise", "all"))
}
\description{
fmri_glm
}

