% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_engine_lowrank.R
\name{fmri_lm_lowrank_dispatch}
\alias{fmri_lm_lowrank_dispatch}
\title{Internal: dispatch fmri_lm low-rank/sketch engine}
\usage{
fmri_lm_lowrank_dispatch(
  formula_or_model,
  dataset,
  engine = NULL,
  lowrank = NULL,
  block = NULL,
  baseline_model = NULL,
  durations = 0,
  drop_empty = TRUE,
  ar_options = NULL
)
}
\description{
Internal: dispatch fmri_lm low-rank/sketch engine
}
\keyword{internal}
