% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_ttest_helpers.R
\name{fmri_meta_fit_extended}
\alias{fmri_meta_fit_extended}
\title{Extended Meta-Analysis Fit with Voxelwise Covariate}
\usage{
fmri_meta_fit_extended(
  Y,
  V,
  X,
  method = c("pm", "dl", "fe", "reml"),
  robust = c("none", "huber"),
  huber_c = 1.345,
  robust_iter = 2,
  voxelwise = NULL,
  center_voxelwise = TRUE,
  voxel_name = "voxel_cov",
  n_threads = getOption("fmrireg.num_threads", 0)
)
}
\arguments{
\item{Y}{Matrix of effect sizes (S x P)}

\item{V}{Matrix of variances (S x P)}

\item{X}{Design matrix (S x K)}

\item{method}{Meta-analysis method}

\item{robust}{Robust estimation method}

\item{huber_c}{Huber tuning constant}

\item{robust_iter}{Number of IRLS iterations}

\item{voxelwise}{Optional voxelwise covariate matrix (S x P)}

\item{center_voxelwise}{Logical; center voxelwise covariate per feature}

\item{voxel_name}{Name for voxelwise coefficient}

\item{n_threads}{Number of threads}
}
\value{
List with meta-analysis results
}
\description{
Wrapper for meta-analysis that supports an optional voxelwise covariate.
This extends the basic fmri_meta_fit to handle per-voxel covariates.
}
