% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrireg.R
\docType{package}
\name{fmrireg-package}
\alias{fmrireg}
\alias{fmrireg-package}
\title{fmrireg: regression tools for fMRI data}
\description{
fmrireg provides functions for generating experimental design matrices appropriate for analyzing fMRI data with regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/fmrireg/}
  \item \url{https://github.com/bbuchsbaum/fmrireg}
  \item Report bugs at \url{https://github.com/bbuchsbaum/fmrireg/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
