% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_empirical_hrf}
\alias{gen_empirical_hrf}
\title{Generate an Empirical Hemodynamic Response Function}
\usage{
gen_empirical_hrf(t, y, name = "empirical_hrf")
}
\arguments{
\item{t}{Time points.}

\item{y}{Values of HRF at time \code{t[i]}.}

\item{name}{Name of the generated HRF.}
}
\value{
An instance of type \code{HRF} inheriting from \code{function}.
}
\description{
\code{gen_empirical_hrf} generates an empirical hemodynamic response function (HRF) using provided time points and HRF values.
}
\examples{

y <- -poly(0:24, 2)[,2]
emphrf <- gen_empirical_hrf(0:24, y)
## plot(emphrf(seq(0,24,by=.5)), type='l')
}
