% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_hrf_lagged}
\alias{gen_hrf_lagged}
\title{gen_hrf_lagged}
\usage{
gen_hrf_lagged(hrf, lag = 2)
}
\arguments{
\item{hrf}{the underlying hrf function to shift}

\item{lag}{the lag/delay in seconds}
}
\description{
gen_hrf_lagged
}

