% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_hrf_library}
\alias{gen_hrf_library}
\title{Generate an HRF (hemodynamic response function) library}
\usage{
gen_hrf_library(fun, pgrid, ...)
}
\arguments{
\item{fun}{A function that generates an HRF, given a set of parameters.}

\item{pgrid}{A data frame containing the parameter grid, where each row represents a set of parameters to be passed to the HRF generating function.}

\item{...}{Additional arguments to be passed to the HRF generating function.}
}
\value{
A list of HRFs generated by applying the HRF generating function to each row of the parameter grid.
}
\description{
This internal function generates an HRF library by applying the provided HRF generating function (\code{fun}) to each row of the parameter grid (\code{pgrid}). Additional arguments can be passed to the generating function using \code{...}.
}
\seealso{
Other gen_hrf: 
\code{\link{gen_hrf_blocked}()},
\code{\link{gen_hrf_lagged}()},
\code{\link{gen_hrf_set}()}
}
\concept{gen_hrf}
\keyword{internal}
