% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_hrf_set}
\alias{gen_hrf_set}
\title{Generate an HRF Basis Set}
\usage{
gen_hrf_set(..., span = 32, name = "hrf_set")
}
\arguments{
\item{...}{A list of functions f(t) mapping from time to amplitude.}

\item{span}{The span in seconds of the HRF.}

\item{name}{The name of the HRF.}
}
\value{
An instance of type \code{HRF} inheriting from \code{function}.
}
\description{
\code{gen_hrf_set} constructs an HRF basis set from one or more component functions.
This function is used to create arbitrary sets of basis functions for fMRI modeling.
}
\examples{

hrf1 <- hrf_spmg1 |> gen_hrf(lag=0)
hrf2 <- hrf_spmg1 |> gen_hrf(lag=3)
hrf3 <- hrf_spmg1 |> gen_hrf(lag=6)

hset <- gen_hrf_set(hrf1, hrf2, hrf3)
}
\seealso{
Other gen_hrf: 
\code{\link{gen_hrf_blocked}()},
\code{\link{gen_hrf_lagged}()}
}
\concept{gen_hrf}
