% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_csv.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get Available Contrasts}
\usage{
get_contrasts(gd)
}
\arguments{
\item{gd}{A group_data_csv object}
}
\value{
Character vector of contrast names
}
\description{
Get Available Contrasts
}
\examples{
gd <- fmrireg:::.demo_group_data_csv()
get_contrasts(gd)
}
