% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{get_covariates}
\alias{get_covariates}
\title{Get Covariates}
\usage{
get_covariates(x)
}
\arguments{
\item{x}{A group_data object}
}
\value{
Data frame of covariates or NULL
}
\description{
Get Covariates
}
