% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_data_matrix}
\alias{get_data_matrix}
\title{get_data_matrix}
\usage{
get_data_matrix(x, ...)
}
\arguments{
\item{x}{the dataset}

\item{...}{extra args}
}
\description{
get_data_matrix
}
\keyword{internal}
