% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/fmri_model_utils.R, R/fmrilm.R
\name{get_formula}
\alias{get_formula}
\alias{get_formula.fmri_model}
\title{Retrieve the formula underlying a model object}
\usage{
get_formula(x, ...)

\method{get_formula}{fmri_model}(x, ...)

\method{get_formula}{fmri_model}(x, ...)
}
\arguments{
\item{x}{The object to extract a formula from.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A formula.
}
\description{
Generic used to expose the original modelling formula for fitted objects.

Functions for creating and manipulating fmri_model objects
}
\examples{
fm <- fmrireg:::.demo_fmri_model()
get_formula(fm)
}
\keyword{internal}
