% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_mask}
\alias{get_mask}
\title{get_mask}
\usage{
get_mask(x, ...)
}
\arguments{
\item{x}{the dataset}

\item{...}{extra args}
}
\description{
get the binary inclusion mask associated with a dataset
}
