% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_csv.R
\name{get_rois}
\alias{get_rois}
\title{Get Available ROIs}
\usage{
get_rois(gd)
}
\arguments{
\item{gd}{A group_data_csv object}
}
\value{
Character vector of ROI names
}
\description{
Get Available ROIs
}
