% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{get_scan_length}
\alias{get_scan_length}
\title{Get run length for a specific scan}
\usage{
get_scan_length(scan_name, task_name, scan_params)
}
\arguments{
\item{scan_name}{Full name of scan}

\item{task_name}{Task name}

\item{scan_params}{Scan parameters from config}
}
\value{
Run length for this scan
}
\description{
Get run length for a specific scan
}
\keyword{internal}
