% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{get_scan_tr}
\alias{get_scan_tr}
\title{Get TR for a specific scan}
\usage{
get_scan_tr(scan_name, scan_params)
}
\arguments{
\item{scan_name}{Full name of scan (e.g., "sub-001_task-memory_run-01")}

\item{scan_params}{Scan parameters from config}
}
\value{
TR value for this scan
}
\description{
Get TR for a specific scan
}
\keyword{internal}
