% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{get_subjects}
\alias{get_subjects}
\title{Get Subject IDs}
\usage{
get_subjects(x)
}
\arguments{
\item{x}{A group_data object}
}
\value{
Character vector of unique subject IDs
}
\description{
Get Subject IDs
}
\examples{
gd <- fmrireg:::.demo_group_data_csv()
get_subjects(gd)
}
