% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame.R
\name{global_onsets.sampling_frame}
\alias{global_onsets.sampling_frame}
\title{Compute global onsets from a sampling_frame}
\usage{
\method{global_onsets}{sampling_frame}(x, onsets, blockids, ...)
}
\arguments{
\item{x}{A sampling_frame object.}

\item{onsets}{A numeric vector of onsets within the specified blocks.}

\item{blockids}{A numeric vector of block IDs corresponding to the onsets.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of global onsets computed from the specified \code{sampling_frame}.
}
\description{
This function computes the global onsets (with respect to the first block) for a given \code{sampling_frame}.
}
\examples{
frame <- sampling_frame(blocklens = c(100, 100, 100), TR = 2, precision = 0.5)
onsets <- c(10, 20, 30)
blockids <- c(1, 2, 3)
global_onsets(frame, onsets, blockids)

}
\seealso{
Other global_onsets: 
\code{\link{global_onsets}()}
}
\concept{global_onsets}
