% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl.R
\name{\%||\%}
\alias{\%||\%}
\title{The \verb{\%||\%} operator}
\usage{
x \%||\% y
}
\description{
This small helper replicates the \verb{rlang::\%||\%} pattern, returning the left-hand
side if not \code{NULL}, otherwise the right-hand side.
}
\keyword{internal}
