% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{AFNI_HRF}
\alias{AFNI_HRF}
\alias{HRF_GAUSSIAN}
\alias{HRF_BSPLINE}
\alias{HRF_SPMG1}
\alias{HRF_SPMG2}
\alias{HRF_SPMG3}
\title{AFNI HRF Constructor Function}
\usage{
AFNI_HRF(name, nbasis, params)

HRF_GAUSSIAN(t, mean = 6, sd = 2)

HRF_BSPLINE(t, span = 20, N = 5, degree = 3)

HRF_SPMG1(t, P1 = 5, P2 = 15, A1 = 0.0833)

HRF_SPMG2(t)

HRF_SPMG3(t)
}
\arguments{
\item{name}{A string specifying the name of the AFNI HRF.}

\item{nbasis}{An integer representing the number of basis functions for the AFNI HRF.}

\item{params}{A list containing the parameter values for the AFNI HRF.}

\item{span}{A numeric value representing the span in seconds of the HRF. Default is 24.}
}
\value{
An AFNI_HRF object with the specified properties.
}
\description{
The \code{AFNI_HRF} function creates an object representing an AFNI-specific hemodynamic response function (HRF). It is a class constructor for AFNI HRFs.
}
\section{Functions}{
\itemize{
\item \code{AFNI_HRF()}: AFNI hrf

\item \code{HRF_GAUSSIAN()}: Gaussian hrf

\item \code{HRF_BSPLINE()}: B-spline hrf

\item \code{HRF_SPMG1()}: SPMG1 hrf

\item \code{HRF_SPMG2()}: SPMG2 hrf

\item \code{HRF_SPMG3()}: SPMG3 hrf

}}
\seealso{
HRF
}
\keyword{internal}
