% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_block}
\alias{hrf_block}
\title{hrf_block}
\usage{
hrf_block(t, hrf = hrf_gaussian, width = 5, precision = 0.1, ...)
}
\arguments{
\item{t}{time in seconds}

\item{hrf}{the underlying hemodynamic response function}

\item{width}{the fixed width of the response in seconds.}

\item{precision}{the sampling precision in seconds}
}
\description{
hrf_block
}
