% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{HRF_DAGUERRE_BASIS}
\alias{HRF_DAGUERRE_BASIS}
\title{Create Daguerre basis HRF set}
\usage{
HRF_DAGUERRE_BASIS(n_basis = 3, scale = 4)
}
\arguments{
\item{n_basis}{Number of basis functions (default: 3)}

\item{scale}{Scale parameter for the time axis (default: 4)}
}
\value{
An HRF object (function) that takes time \code{t} and returns the basis set matrix.
}
\description{
Creates an HRF object representing a set of hemodynamic response functions
using Daguerre spherical basis functions.
}
\examples{
hrf_dag_basis_obj <- HRF_DAGUERRE_BASIS(n_basis = 3)
t <- seq(0, 32, by=0.1)
basis_set <- hrf_dag_basis_obj(t)
matplot(t, basis_set, type='l', xlab='Time (s)', ylab='Response')
}
