% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_gamma}
\alias{hrf_gamma}
\title{gamma hemodynamic response function}
\usage{
hrf_gamma(t, shape = 6, rate = 1)
}
\arguments{
\item{t}{time}

\item{shape}{the shape parameter for gamma pdf}

\item{rate}{the rate parameter for gamma pdf}
}
\description{
A hemodynamic response function using the Gamma density function
}
