% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_time}
\alias{hrf_time}
\title{hrf_time}
\usage{
hrf_time(t, maxt = 22)
}
\arguments{
\item{t}{time in seconds}

\item{maxt}{the maximum time point in domain}
}
\description{
hemodyanmic response function that is a simple linear function of time t, when t > 0.
}
