% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{Ident}
\alias{Ident}
\title{Ident}
\usage{
Ident(...)
}
\arguments{
\item{...}{a list of variable names}
}
\value{
an instance of class \code{Ident} extending \code{ParametricBasis}
}
\description{
A basis that applies identity transform to a set of raw variables.
}
