% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{is.increasing}
\alias{is.increasing}
\title{Check if a vector is non-decreasing.}
\usage{
is.increasing(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
TRUE if the differences are all non-negative.
}
\description{
Check if a vector is non-decreasing.
}
\keyword{internal}
