% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{is.strictly.increasing}
\alias{is.strictly.increasing}
\title{Check if a vector is strictly increasing.}
\usage{
is.strictly.increasing(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
TRUE if all differences are strictly positive.
}
\description{
Check if a vector is strictly increasing.
}
\keyword{internal}
