% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-classes.R
\name{is_categorical.event}
\alias{is_categorical.event}
\title{Check if an event object represents categorical values}
\usage{
\method{is_categorical}{event}(x)
}
\arguments{
\item{x}{An object of class \code{event}.}
}
\value{
Logical \code{TRUE} if categorical (factor), \code{FALSE} if continuous.
}
\description{
Checks if an event is categorical (i.e., not continuous).
}
