% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{is_categorical}
\alias{is_categorical}
\title{is_categorical}
\usage{
is_categorical(x)
}
\arguments{
\item{x}{the object}
}
\description{
checks to see if a variable is categorical e.g. factor-based
}
