% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-classes.R
\name{is_continuous.event}
\alias{is_continuous.event}
\title{Check if an event object represents continuous values}
\usage{
\method{is_continuous}{event}(x)
}
\arguments{
\item{x}{An object of class \code{event}.}
}
\value{
Logical \code{TRUE} if continuous (numeric, matrix, basis), \code{FALSE} if categorical (factor).
}
\description{
Checks the \verb{$continuous} flag set by the \code{event()} constructor.
}
