% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-classes.R
\name{levels.event}
\alias{levels.event}
\alias{columns.event}
\title{Get levels or column names of an event object}
\usage{
\method{levels}{event}(x)

\method{columns}{event}(x)
}
\arguments{
\item{x}{An object of class \code{event}.}
}
\value{
A character vector of names.
}
\description{
Retrieves the appropriate names associated with the event values.
\itemize{
\item For factors (stored internally as integers), retrieves names from \verb{$meta$levels}.
\item For basis objects, retrieves names via \code{levels()} method of the stored basis object in \verb{$meta$basis}.
\item For matrices or numeric vectors, returns the column names of the internal \verb{$value} matrix.
}
}
\section{Functions}{
\itemize{
\item \code{columns(event)}: Alias for levels.event

}}
