% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{levels.ParametricBasis}
\alias{levels.ParametricBasis}
\alias{levels.Scale}
\alias{levels.ScaleWithin}
\alias{levels.RobustScale}
\title{Get Levels/Names for ParametricBasis Objects}
\usage{
\method{levels}{Scale}(x)

\method{levels}{ScaleWithin}(x)

\method{levels}{RobustScale}(x)
}
\arguments{
\item{x}{ParametricBasis object}
}
\description{
Get Levels/Names for ParametricBasis Objects
}
