% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_vector.R
\name{levels}
\alias{levels}
\alias{levels.event_factor}
\alias{levels.event_variable}
\alias{levels.event_matrix}
\alias{levels.event_set}
\alias{levels.event_basis}
\alias{levels.event_term}
\title{levels}
\usage{
levels(x)

\method{levels}{event_factor}(x)

\method{levels}{event_variable}(x)

\method{levels}{event_matrix}(x)

\method{levels}{event_set}(x)

\method{levels}{event_basis}(x)

\method{levels}{event_term}(x)
}
\arguments{
\item{x}{the term}
}
\description{
extract the levels of a term
}
