% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{matrix_term}
\alias{matrix_term}
\title{matrix_term}
\usage{
matrix_term(varname, mat)
}
\arguments{
\item{varname}{the name of the variable}

\item{mat}{the \code{matrix} of values}
}
\description{
A set of regression variables stored as a numeric matrix
}
\examples{
mat <- matrix(rnorm(100*10), 100 ,10)
mterm <- matrix_term("mterm", mat)
}
