% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafuns.R
\name{meta_contrasts}
\alias{meta_contrasts}
\title{Meta-analysis of contrasts}
\usage{
meta_contrasts(cres, weighting = c("inv_var", "equal"))
}
\arguments{
\item{cres}{A list of contrast results, where each element of the list contains the results for a particular study or test.}

\item{weighting}{A character string specifying the weighting method to use for the meta-analysis. Options are "inv_var" (inverse variance weighting) or "equal" (equal weighting). Default is "inv_var".}
}
\value{
A list containing meta-analysis results for each contrast, including:
\itemize{
\item{estimate}{A matrix of the combined contrast estimates.}
\item{se}{A matrix of the pooled standard errors.}
\item{stat}{A matrix of the combined z-scores.}
\item{prob}{A matrix of the combined p-values.}
\item{stat_type}{A character string indicating the type of statistic used, in this case "zstat".}
}
}
\description{
This function performs a meta-analysis on a list of contrast results obtained from different studies or tests.
}
\examples{
# Assuming `cres` is a list of contrast results obtained from different studies or tests
# meta_results <- meta_contrasts(cres)
# print(meta_results)
}
\keyword{internal}
