% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafuns.R
\name{meta_Fcontrasts}
\alias{meta_Fcontrasts}
\title{Meta-analysis of F-contrasts}
\usage{
meta_Fcontrasts(fres)
}
\arguments{
\item{fres}{A list of F-contrast results, where each element of the list contains the results for a particular study or test.}
}
\value{
A list containing meta-analysis results for each contrast. Each element of the list includes:
\itemize{
\item{estimate}{A numeric vector of the combined z-scores.}
\item{se}{A numeric vector of the pooled standard errors.}
\item{stat}{A numeric vector of the combined z-scores.}
\item{prob}{A numeric vector of the combined p-values.}
\item{stat_type}{A character string indicating the type of statistic used, in this case "zfstat".}
}
}
\description{
This function performs a meta-analysis on a list of F-contrasts results obtained from different studies or tests.
}
\examples{
# Assuming `fres` is a list of F-contrast results obtained from different studies or tests
# meta_results <- meta_Fcontrasts(fres)
# print(meta_results)
}
\keyword{internal}
