% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meta_fit_vcov_cpp}
\alias{meta_fit_vcov_cpp}
\title{Meta-regression with ONE voxelwise covariate}
\usage{
meta_fit_vcov_cpp(
  Y,
  V,
  X,
  C,
  method,
  robust,
  huber_c = 1.345,
  robust_iter = 2L,
  n_threads = 0L
)
}
\arguments{
\item{Y}{S x P matrix of effect sizes}

\item{V}{S x P matrix of variances}

\item{X}{S x K design matrix}

\item{C}{S x P matrix of voxelwise covariates}

\item{method}{Meta-analysis method}

\item{robust}{Robust estimation method}

\item{huber_c}{Huber tuning constant}

\item{robust_iter}{Number of IRLS iterations}

\item{n_threads}{Number of OpenMP threads}
}
\value{
List with results
}
\description{
Meta-regression with ONE voxelwise covariate
}
