% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafuns.R
\name{meta_stouffer}
\alias{meta_stouffer}
\title{Meta-analysis using Stouffer's method}
\usage{
meta_stouffer(pval, se)
}
\arguments{
\item{pval}{A numeric vector of p-values from multiple studies or tests.}

\item{se}{A numeric vector of standard errors corresponding to the p-values.}
}
\value{
A list containing the following elements:
\itemize{
\item{estimate}{A numeric vector of the weighted z-scores obtained using Stouffer's method.}
\item{se}{A numeric vector of the combined standard errors.}
\item{stat}{Same as the estimate, the weighted z-scores.}
\item{prob}{A numeric vector of the combined p-values.}
\item{stat_type}{A character string indicating the type of statistic used, in this case "zfstat".}
}
}
\description{
This function performs a meta-analysis on input p-values and standard errors using Stouffer's method.
Stouffer's method combines z-scores by weighting them by their inverse variance.
}
\examples{
pval <- c(0.05, 0.01, 0.03)
se <- c(0.2, 0.15, 0.25)
result <- meta_stouffer(pval, se)
print(result)
}
\keyword{internal}
