% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{multiresponse_arma}
\alias{multiresponse_arma}
\title{Fit a multiresponse ARMA model}
\usage{
multiresponse_arma(
  form,
  data_env,
  conlist,
  vnames,
  fcon,
  modmat,
  blockids,
  autocor = c("ar1", "ar2", "arma", "auto")
)
}
\arguments{
\item{form}{The model formula.}

\item{data_env}{The dataset environment containing the time-series data.}

\item{conlist}{The contrast list.}

\item{vnames}{The variable names.}

\item{fcon}{The F-contrast.}

\item{modmat}{The design matrix.}

\item{blockids}{The block ids.}

\item{autocor}{The autocorrelation structure, either "ar1", "ar2", "arma", or "auto".}
}
\value{
A list containing the contrasts, F-contrasts, and beta statistics for the fitted ARMA model.
}
\description{
This function fits a multiresponse ARMA model for fMRI data with specified autocorrelation
structure using the provided formula, dataset environment, contrast list, variable names,
F-contrast, design matrix, and block ids.
}
\seealso{
lm.fit, Arima, auto.arima
}
\keyword{internal}
