% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{multiresponse_bootstrap_lm}
\alias{multiresponse_bootstrap_lm}
\title{Multiresponse Bootstrap Linear Model}
\usage{
multiresponse_bootstrap_lm(
  form,
  data_env,
  conlist,
  vnames,
  fcon,
  modmat,
  block_size = 30,
  boot_rows = FALSE,
  nboot = 100,
  event_indices
)
}
\arguments{
\item{form}{Formula for the linear model.}

\item{data_env}{Environment containing the data for the linear model.}

\item{conlist}{List of contrasts.}

\item{vnames}{Vector of variable names.}

\item{fcon}{Contrasts for fixed effects.}

\item{modmat}{Model matrix.}

\item{block_size}{Size of the blocks for the bootstrap (default: 30).}

\item{boot_rows}{Logical flag indicating whether to bootstrap rows (default: FALSE).}

\item{nboot}{Number of bootstrap iterations (default: 100).}

\item{event_indices}{Indices of events.}
}
\value{
A list containing the original fitted model and bootstrap results.
}
\description{
This function performs multiresponse bootstrap for a linear model.
}
