% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{multiresponse_lm}
\alias{multiresponse_lm}
\title{Fit Multiresponse Linear Model}
\usage{
multiresponse_lm(form, data_env, conlist, vnames, fcon, modmat = NULL)
}
\arguments{
\item{form}{The formula used to define the linear model.}

\item{data_env}{The environment containing the data to be used in the linear model.}

\item{conlist}{The list of contrasts used in the analysis.}

\item{vnames}{The names of the variables used in the linear model.}

\item{fcon}{The F-contrasts used in the analysis.}

\item{modmat}{The model matrix (default is \code{NULL}, which will calculate the model matrix using the formula).}
}
\value{
A list containing the results from the multiresponse linear model analysis.
}
\description{
This function fits a linear model to multiple responses in an fMRI dataset.
}
\keyword{internal}
