% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{n_subjects}
\alias{n_subjects}
\title{Extract Number of Subjects}
\usage{
n_subjects(x)
}
\arguments{
\item{x}{A group_data object}
}
\value{
Integer number of unique subjects
}
\description{
Extract Number of Subjects
}
