% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R, R/regressor.R
\name{nbasis}
\alias{nbasis}
\alias{nbasis.convolved_term}
\alias{nbasis.hrfspec}
\title{return number of basis functions associated with hrf.}
\usage{
nbasis(x)

\method{nbasis}{convolved_term}(x)

\method{nbasis}{hrfspec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
return number of basis functions associated with hrf.
}
