% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hrf.R
\name{new_best_hrf_result}
\alias{new_best_hrf_result}
\title{Create a new "best_hrf_result" object}
\usage{
new_best_hrf_result(
  best_hrf_indices = NULL,
  clusters = NULL,
  cluster_representatives = NULL,
  reassigned_hrfs = NULL,
  L_best_hrfs = NULL,
  L_unique = NULL,
  L_proj = NULL,
  X_proj = NULL,
  reassigned_hrfs_smoothed = NULL
)
}
\arguments{
\item{best_hrf_indices}{numeric vector of unique best HRFs}

\item{clusters}{integer vector of clusters assigned to each unique HRF}

\item{cluster_representatives}{integer vector of cluster medoids}

\item{reassigned_hrfs}{integer vector of HRF IDs assigned to each voxel}

\item{L_best_hrfs}{matrix of all best HRFs found}

\item{L_unique}{matrix of unique HRFs (one per cluster/representative)}

\item{L_proj}{matrix of the library projected into data space}

\item{X_proj}{matrix of the data projection}

\item{reassigned_hrfs_smoothed}{optionally the smoothed HRF assignments}
}
\value{
An object of class \code{"best_hrf_result"}.
}
\description{
This constructor wraps up the result from \code{find_best_hrf}
into a proper S3 class with a pretty print method.
}
\keyword{internal}
