% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{nuisance}
\alias{nuisance}
\title{nuisance}
\usage{
nuisance(x)
}
\arguments{
\item{x}{a \code{matrix}}
}
\value{
a class of type \code{nuisancespec}
}
\description{
a 'nuisance' term that consists of an arbitrary numeric matrix with the same number of rows as image time points.
}
