% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ols_t_cpp}
\alias{ols_t_cpp}
\title{OLS t-test / ANCOVA across features}
\usage{
ols_t_cpp(Y, X)
}
\arguments{
\item{Y}{S x P matrix (subjects x features)}

\item{X}{S x K design matrix with intercept if desired}
}
\value{
List with beta (K x P), se (K x P), t (K x P), df (scalar), ok (P)
}
\description{
OLS t-test / ANCOVA across features
}
