% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ols_t_vcov_cpp}
\alias{ols_t_vcov_cpp}
\title{OLS with ONE voxelwise covariate}
\usage{
ols_t_vcov_cpp(Y, X, C)
}
\arguments{
\item{Y}{S x P matrix of outcomes}

\item{X}{S x K design matrix}

\item{C}{S x P matrix of voxelwise covariates}
}
\value{
List with beta ((K+1) x P), se, t, df
}
\description{
OLS with ONE voxelwise covariate
}
