% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{pair_contrast}
\alias{pair_contrast}
\title{pair_contrast}
\usage{
pair_contrast(A, B, name, where = TRUE, split_by = NULL)
}
\arguments{
\item{A}{the first \code{formula} expression in the contrast}

\item{B}{the second \code{formula} expression in the contrast}

\item{name}{the name of the contrast}

\item{where}{the subset over which the contrast is computed}

\item{split_by}{split the contrast by a second factor variable}
}
\description{
pair_contrast
}
