% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{pairwise_contrasts}
\usage{
pairwise_contrasts(levels, where = TRUE)
}
\arguments{
\item{where}{the subset over which the contrast is computed}
}
\description{
pairwise_contrasts
}
