% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_vector.R
\name{parent_terms}
\alias{parent_terms}
\alias{parent_terms.event_term}
\title{parent_terms}
\usage{
parent_terms(x)

\method{parent_terms}{event_term}(x)
}
\arguments{
\item{x}{the object}
}
\description{
parent_terms
}
\keyword{internal}
