% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{parse_contrast_formula}
\alias{parse_contrast_formula}
\title{Parse Contrast Formula}
\usage{
parse_contrast_formula(formula, object)
}
\arguments{
\item{formula}{Contrast formula}

\item{object}{fmri_meta object}
}
\value{
Numeric vector of contrast weights
}
\description{
Parse Contrast Formula
}
\keyword{internal}
