% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta.R
\name{parse_meta_formula}
\alias{parse_meta_formula}
\title{Parse Meta-Analysis Formula}
\usage{
parse_meta_formula(formula, data)
}
\arguments{
\item{formula}{Formula object}

\item{data}{Group data object}
}
\value{
List with design matrix X and other model info
}
\description{
Parse Meta-Analysis Formula
}
\keyword{internal}
