% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{plot.event_model}
\alias{plot.event_model}
\title{Plot an event_model object}
\usage{
\method{plot}{event_model}(x, y, term_name = NULL, longnames = TRUE, ...)
}
\arguments{
\item{x}{An event_model object.}

\item{y}{Unused.}

\item{term_name}{Name of the term to plot (optional).}

\item{longnames}{Logical flag; if TRUE, use long names for the plot (default: TRUE).}

\item{...}{Additional arguments to be passed to the ggplot2 function.}
}
\value{
A ggplot2 plot object.
}
\description{
This function creates a plot of an event_model object, visualizing the design matrix.
}
