% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{plot.regressor}
\alias{plot.regressor}
\title{plot a regressor object}
\usage{
\method{plot}{regressor}(x, samples, add = FALSE, ...)
}
\arguments{
\item{x}{the object}

\item{samples}{the times in seconds along which to plot regressor function}

\item{add}{whether to add to existing plot}

\item{...}{extra args to send to \code{plot}}
}
\description{
plot a regressor object
}
