% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{plot_contrasts}
\alias{plot_contrasts}
\title{plot_contrasts}
\usage{
plot_contrasts(x, ...)
}
\arguments{
\item{x}{Object containing contrast information}

\item{...}{Additional arguments passed to methods}
}
\description{
Generic function for plotting contrasts.
}
