% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{Poly}
\alias{Poly}
\title{Polynomila basis}
\usage{
Poly(x, degree)
}
\arguments{
\item{x}{a numeric vector at which to evaluate the polynomial. Missing values are not allowed in x.}

\item{degree}{the degree of the polynomial. Must be less than the number of unique points.}
}
\value{
an instance of class \code{Poly} extending \code{ParametricBasis}
}
\description{
Orthogonal polynomial expansion of a linear term based on \code{\link[stats]{poly}}
}
\seealso{
\link[stats]{poly}
}
