% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{poly_contrast}
\alias{poly_contrast}
\title{Polynomial Contrast}
\usage{
poly_contrast(A, name, where = NULL, degree = 1, value_map = NULL)
}
\arguments{
\item{A}{A formula describing the contrast.}

\item{name}{A character string representing the name of the contrast.}

\item{where}{An optional formula defining the subset over which the contrast is computed.}

\item{degree}{An integer representing the degree of the polynomial (default: 1).}

\item{value_map}{An optional list that maps between levels of a factor and a numeric value.}
}
\value{
A list containing the polynomial contrast specification.
}
\description{
Create a polynomial contrast evaluated over a set of (orderable) factor levels.
}
\examples{
pcon <- poly_contrast(~ time, name="poly_time_3", degree=3)

}
