% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{predict.ParametricBasis}
\alias{predict.ParametricBasis}
\alias{predict.Scale}
\alias{predict.ScaleWithin}
\alias{predict.RobustScale}
\title{Predict Method for ParametricBasis Objects}
\usage{
\method{predict}{Scale}(object, newdata, ...)

\method{predict}{ScaleWithin}(object, newdata, newgroup, ...)

\method{predict}{RobustScale}(object, newdata, ...)
}
\description{
Predict Method for ParametricBasis Objects
}
