% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_model.R
\name{prediction_matrix}
\alias{prediction_matrix}
\title{(Internal) Prediction Matrix}
\usage{
prediction_matrix(x)
}
\arguments{
\item{x}{An fmri_model object.}
}
\value{
(Not implemented)
}
\description{
This function is intended to compute a prediction matrix for the model.
(Currently a stub.)
}
\keyword{internal}
