% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.afni_hrf_convolved_term}
\alias{print.afni_hrf_convolved_term}
\title{Print afni_hrf_convolved_term objects.}
\usage{
\method{print}{afni_hrf_convolved_term}(x, ...)
}
\arguments{
\item{x}{An afni_hrf_convolved_term object.}

\item{...}{Additional arguments.}
}
\description{
Print afni_hrf_convolved_term objects.
}
