% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{print.baseline_model}
\alias{print.baseline_model}
\title{Print a Baseline Model.}
\usage{
\method{print}{baseline_model}(x, ...)
}
\arguments{
\item{x}{A baseline_model object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays key information about the baseline model components and a preview
of the design matrix.
}
