% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hrf.R
\name{print.best_hrf_result}
\alias{print.best_hrf_result}
\title{Print method for "best_hrf_result"}
\usage{
\method{print}{best_hrf_result}(x, ...)
}
\arguments{
\item{x}{A \code{best_hrf_result} object}

\item{...}{Not used}
}
\description{
Uses crayon for colorful output
}
