% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{print.BSplineHRF}
\alias{print.BSplineHRF}
\title{Print method for B-spline HRF objects}
\usage{
\method{print}{BSplineHRF}(x, ...)
}
\arguments{
\item{x}{A B-spline HRF object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for B-spline HRF objects
}
