% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{print.chunkiter}
\alias{print.chunkiter}
\title{Pretty Print a Chunk Iterator}
\usage{
\method{print}{chunkiter}(x, ...)
}
\arguments{
\item{x}{A chunkiter object.}

\item{...}{Additional arguments (ignored).}
}
\description{
This function prints a summary of a chunk iterator using colored output.
}
