% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{print.contrast}
\alias{print.contrast}
\title{Print Contrast}
\usage{
\method{print}{contrast}(x, ...)
}
\arguments{
\item{x}{The contrast to print.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a contrast.
}
