% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{print.contrast_diff_spec}
\alias{print.contrast_diff_spec}
\title{Print Contrast Difference Specification}
\usage{
\method{print}{contrast_diff_spec}(x, ...)
}
\arguments{
\item{x}{The contrast difference specification to print.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a contrast difference specification.
}
