% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{print.contrast_set}
\alias{print.contrast_set}
\title{Print Contrast Set}
\usage{
\method{print}{contrast_set}(x, ...)
}
\arguments{
\item{x}{The contrast set to print.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a contrast set.
}
