% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{print.contrast_spec}
\alias{print.contrast_spec}
\title{Print Contrast Specification}
\usage{
\method{print}{contrast_spec}(x, ...)
}
\arguments{
\item{x}{The contrast specification to print.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a contrast specification.
}
