% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.convolved_term}
\alias{print.convolved_term}
\title{Print convolved_term objects.}
\usage{
\method{print}{convolved_term}(x, ...)
}
\arguments{
\item{x}{A convolved_term object.}

\item{...}{Additional arguments.}
}
\description{
Print convolved_term objects.
}
