% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{print.data_chunk}
\alias{print.data_chunk}
\title{Pretty Print a Data Chunk Object}
\usage{
\method{print}{data_chunk}(x, ...)
}
\arguments{
\item{x}{A data_chunk object.}

\item{...}{Additional arguments (ignored).}
}
\description{
This function prints a summary of a data chunk using crayon for colored output.
}
