% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-classes.R
\name{print.event}
\alias{print.event}
\title{Print event objects}
\usage{
\method{print}{event}(x, ...)
}
\arguments{
\item{x}{An event object.}

\item{...}{Additional arguments (unused).}
}
\description{
Provides a concise summary of an event object using cli.
}
