% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.event_factor}
\alias{print.event_factor}
\title{Print event_factor objects.}
\usage{
\method{print}{event_factor}(x, ...)
}
\arguments{
\item{x}{An event_factor object.}

\item{...}{Additional arguments.}
}
\description{
Print event_factor objects.
}
