% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.event_matrix}
\alias{print.event_matrix}
\title{Print event_matrix objects.}
\usage{
\method{print}{event_matrix}(x, ...)
}
\arguments{
\item{x}{An event_matrix object.}

\item{...}{Additional arguments.}
}
\description{
Print event_matrix objects.
}
