% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{print.event_model}
\alias{print.event_model}
\title{Print an Event Model}
\usage{
\method{print}{event_model}(x, ...)
}
\arguments{
\item{x}{An \code{event_model} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Provides a concise summary of an \code{event_model} object using \code{cli}.
}
