% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.event_variable}
\alias{print.event_variable}
\title{Print event_variable objects.}
\usage{
\method{print}{event_variable}(x, ...)
}
\arguments{
\item{x}{An event_variable object.}

\item{...}{Additional arguments.}
}
\description{
Print event_variable objects.
}
