% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_betas.R
\name{print.fmri_betas}
\alias{print.fmri_betas}
\title{Pretty print method for fmri_betas objects}
\usage{
\method{print}{fmri_betas}(x, ...)
}
\arguments{
\item{x}{An object of class "fmri_betas"}

\item{...}{Additional arguments passed to the print function}
}
\description{
Pretty print method for fmri_betas objects
}
