% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{print.fmri_lm}
\alias{print.fmri_lm}
\title{Print an fmri_lm Object}
\usage{
\method{print}{fmri_lm}(x, ...)
}
\arguments{
\item{x}{An \code{fmri_lm} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function prints a summary of an \code{fmri_lm} object.
}
