% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{print.fmri_meta}
\alias{print.fmri_meta}
\title{Print Meta-Analysis Results}
\usage{
\method{print}{fmri_meta}(x, ...)
}
\arguments{
\item{x}{An fmri_meta object}

\item{...}{Additional print arguments}
}
\value{
Invisibly returns the input object x
}
\description{
Print Meta-Analysis Results
}
