% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{print.fmri_rlm}
\alias{print.fmri_rlm}
\title{Print method for fmri_rlm objects}
\usage{
\method{print}{fmri_rlm}(x, ...)
}
\arguments{
\item{x}{An fmri_rlm object}

\item{...}{Additional arguments passed to print.fmri_lm}
}
\description{
Print method for fmri_rlm objects
}
