% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.fmri_term}
\alias{print.fmri_term}
\title{Print fmri_term objects.}
\usage{
\method{print}{fmri_term}(x, ...)
}
\arguments{
\item{x}{An fmri_term object.}

\item{...}{Additional arguments.}
}
\description{
Print fmri_term objects.
}
