% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_ttest.R
\name{print.fmri_ttest_fit}
\alias{print.fmri_ttest_fit}
\title{Print method for fmri_ttest_fit}
\usage{
\method{print}{fmri_ttest_fit}(x, ...)
}
\arguments{
\item{x}{An fmri_ttest_fit object}

\item{...}{Additional print arguments}
}
\description{
Print method for fmri_ttest_fit
}
