% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{print.GaussianHRF}
\alias{print.GaussianHRF}
\title{Print method for Gaussian HRF objects}
\usage{
\method{print}{GaussianHRF}(x, ...)
}
\arguments{
\item{x}{A Gaussian HRF object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for Gaussian HRF objects
}
