% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{print.group_data}
\alias{print.group_data}
\title{Print Group Data Object}
\usage{
\method{print}{group_data}(x, ...)
}
\arguments{
\item{x}{A group_data object}

\item{...}{Additional print arguments}
}
\description{
Print Group Data Object
}
