% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{print.HRF}
\alias{print.HRF}
\title{Print method for HRF objects}
\usage{
\method{print}{HRF}(x, ...)
}
\arguments{
\item{x}{An HRF object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for HRF objects
}
