% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{print.poly_contrast_spec}
\alias{print.poly_contrast_spec}
\title{Print Polynomial Contrast Specification}
\usage{
\method{print}{poly_contrast_spec}(x, ...)
}
\arguments{
\item{x}{The polynomial contrast specification to print.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a polynomial contrast specification.
}
