% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R, R/baseline_model.R, R/contrast.R,
%   R/event-classes.R, R/event_model.R, R/event_vector.R, R/fmri_model.R,
%   R/fmrilm.R, R/fmrirlm.R
\name{print.afni_lm_spec}
\alias{print.afni_lm_spec}
\alias{print.baseline_model}
\alias{print.contrast_set}
\alias{print.contrast_spec}
\alias{print.contrast}
\alias{print.poly_contrast_spec}
\alias{print.contrast_diff_spec}
\alias{print.event}
\alias{print.event_model}
\alias{print.fmri_term}
\alias{print.convolved_term}
\alias{print.afni_hrf_convolved_term}
\alias{print.event_term}
\alias{print.fmri_model}
\alias{print.fmri_lm}
\alias{print.fmri_rlm}
\title{Print a Baseline Model}
\usage{
\method{print}{afni_lm_spec}(x, ...)

\method{print}{baseline_model}(x, ...)

\method{print}{contrast_set}(x, ...)

\method{print}{contrast_spec}(x, ...)

\method{print}{contrast}(x, ...)

\method{print}{poly_contrast_spec}(x, ...)

\method{print}{contrast_diff_spec}(x, ...)

\method{print}{event}(x, ...)

\method{print}{event_model}(x, ...)

\method{print}{fmri_term}(x, ...)

\method{print}{convolved_term}(x, ...)

\method{print}{afni_hrf_convolved_term}(x, ...)

\method{print}{event_term}(x, ...)

\method{print}{fmri_model}(x, ...)

\method{print}{fmri_lm}(x, ...)

\method{print}{fmri_rlm}(x, ...)
}
\arguments{
\item{x}{An fmri_rlm object}

\item{...}{Additional arguments passed to print.fmri_lm}
}
\value{
The input object, invisibly.
}
\description{
Displays key information about the baseline model components and a preview
of the design matrix.

Print a contrast set.

Print a contrast specification.

Print a contrast.

Print a polynomial contrast specification.

Print a contrast difference specification.

Provides a concise summary of an event object using cli.

Provides a concise summary of an \code{event_model} object using \code{cli}.

Provides a concise summary of an event_term object using cli.

Provides a colorful and informative printout.
}
\examples{
sframe <- fmrihrf::sampling_frame(blocklens = 5, TR = 1)
bmod <- baseline_model(sframe = sframe)
print(bmod)
}
