% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{print.SPMGHRF}
\alias{print.SPMGHRF}
\title{Print method for SPMG HRF objects}
\usage{
\method{print}{SPMGHRF}(x, ...)
}
\arguments{
\item{x}{An SPMG HRF object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for SPMG HRF objects
}
