% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{pull_stat}
\alias{pull_stat}
\title{Extract Statistical Measures from an fmri_lm Object}
\usage{
pull_stat(x, type, element)
}
\arguments{
\item{x}{An \code{fmri_lm} object.}

\item{type}{The type of statistic to extract: \code{"betas"}, \code{"contrasts"}, or \code{"F"}.}

\item{element}{The specific element to extract, such as \code{"estimate"}, \code{"se"}, \code{"stat"}, or \code{"prob"}.}
}
\value{
A tibble containing the requested statistical measures.
}
\description{
This function extracts statistical measures (e.g., estimates, standard errors) from an \code{fmri_lm} object.
}
\keyword{internal}
