% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank1_estimation.R
\name{r1_glm_betas}
\alias{r1_glm_betas}
\title{Rank‑1 GLM (stacked design)}
\usage{
r1_glm_betas(
  X,
  y,
  Z = NULL,
  hrf_basis,
  hrf_ref,
  maxit = 100,
  flip_sign = TRUE,
  use_box_constraints = FALSE,
  method = c("als", "svd")
)
}
\arguments{
\item{method}{\code{"als"} (default) or \code{"svd"}}
}
\description{
Rank‑1 GLM (stacked design)
}
