% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank1_estimation.R
\name{r1_glms_betas}
\alias{r1_glms_betas}
\title{Rank‑1 GLM (Mumford / split designs)}
\usage{
r1_glms_betas(
  X_list,
  X_all = NULL,
  y,
  Z = NULL,
  hrf_basis,
  hrf_ref,
  maxit = 100,
  method = c("als", "svd")
)
}
\arguments{
\item{method}{\code{"als"} (default) or \code{"svd"}}
}
\description{
Rank‑1 GLM (Mumford / split designs)
}
