% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_h5.R
\name{read_h5_full}
\alias{read_h5_full}
\title{Read All Data from HDF5 Files}
\usage{
read_h5_full(gd, stat = NULL)
}
\arguments{
\item{gd}{A group_data_h5 object}

\item{stat}{Character vector of statistics to extract}
}
\value{
Array with dimensions (voxels, subjects, statistics)
}
\description{
Reads complete data from all subjects' HDF5 files.
Warning: This can use a lot of memory for whole-brain data.
}
