% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_h5.R
\name{read_h5_metadata}
\alias{read_h5_metadata}
\title{Read Metadata from HDF5 File}
\usage{
read_h5_metadata(path)
}
\arguments{
\item{path}{Path to HDF5 file}
}
\value{
List with metadata including dimensions and labels
}
\description{
Read Metadata from HDF5 File
}
\keyword{internal}
