% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jobs.R
\name{read_jobs}
\alias{read_jobs}
\title{Read jobs exported by export_jobs()}
\usage{
read_jobs(dir)
}
\arguments{
\item{dir}{Directory containing \code{manifest.rds}.}
}
\value{
The list of \link{fmri_job}s.
}
\description{
Read jobs exported by export_jobs()
}
\seealso{
\code{\link[=export_jobs]{export_jobs()}}
}
