% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{read_nifti_chunk}
\alias{read_nifti_chunk}
\title{Read Data Chunk from NIfTI Files}
\usage{
read_nifti_chunk(gd, voxel_indices)
}
\arguments{
\item{gd}{A group_data_nifti object}

\item{voxel_indices}{Integer vector of voxel indices to read}
}
\value{
Matrix with dimensions (subjects, voxels)
}
\description{
Reads a subset of voxels from all subjects' NIfTI files
}
\keyword{internal}
