% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{read_nifti_data}
\alias{read_nifti_data}
\title{Read NIfTI Data}
\usage{
read_nifti_data(path, mask = NULL)
}
\arguments{
\item{path}{Path to NIfTI file}

\item{mask}{Optional mask to apply}
}
\value{
Array or vector of data
}
\description{
Read NIfTI Data
}
\keyword{internal}
