% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{read_nifti_header}
\alias{read_nifti_header}
\title{Read NIfTI Header Information}
\usage{
read_nifti_header(path)
}
\arguments{
\item{path}{Path to NIfTI file}
}
\value{
List with header information
}
\description{
Read NIfTI Header Information
}
\keyword{internal}
