% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{reconstruct_image}
\alias{reconstruct_image}
\title{Reconstruct 3D Image from Vector}
\usage{
reconstruct_image(values, object)
}
\arguments{
\item{values}{Numeric vector of values}

\item{object}{fmri_meta object with spatial information}
}
\value{
NeuroVol object
}
\description{
Reconstruct 3D Image from Vector
}
\keyword{internal}
