% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{regressors}
\alias{regressors}
\title{extract regressor set}
\usage{
regressors(x, ...)
}
\arguments{
\item{x}{a model object that contains regressors (or can generate them)}

\item{...}{extra args}
}
\description{
extract regressor set
}
\concept{regressors}
