% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{rescale_columns_01_to_neg1_pos1}
\alias{rescale_columns_01_to_neg1_pos1}
\title{Rescale columns of a design matrix to (-1,1).}
\usage{
rescale_columns_01_to_neg1_pos1(
  DM,
  handle_constant = c("to_zero", "error", "skip")
)
}
\arguments{
\item{DM}{A numeric matrix or data frame (e.g. from \code{design_matrix(event_model)}).}

\item{handle_constant}{Whether to force constant (zero-range) columns to zero, or
raise an error, or skip them. Default "to_zero" sets them all to 0.}
}
\value{
A matrix the same size as \code{DM} with each column scaled to (-1,1).
}
\description{
Rescale columns of a design matrix to (-1,1).
}
