% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{run}
\alias{run}
\title{run}
\usage{
run(x, ...)
}
\arguments{
\item{x}{the command to run}

\item{...}{extra args}
}
\description{
This function runs a command \code{x} with the provided extra arguments.
}
\details{
Run a command or analysis step.
}
