% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_backends.R
\name{run_backends}
\alias{run_backends}
\title{List registered execution backends}
\usage{
run_backends()
}
\value{
A character vector of backend names.
}
\description{
List registered execution backends
}
\seealso{
\code{\link[=register_run_backend]{register_run_backend()}}, \code{\link[=run_jobs]{run_jobs()}}
}
