% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{runwise_rlm}
\alias{runwise_rlm}
\title{Perform Runwise Robust Linear Modeling on fMRI Dataset}
\usage{
runwise_rlm(
  dataset,
  model,
  conlist,
  fcons,
  meta_weighting = c("inv_var", "equal"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{The \code{fmri_model} used for the analysis.}

\item{verbose}{Logical. Whether to display progress messages (default is \code{FALSE}).}
}
\description{
Perform Runwise Robust Linear Modeling on fMRI Dataset
}
\keyword{internal}
