% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{samples}
\alias{samples}
\title{extract samples}
\usage{
samples(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
get the sampling times for a regressor or sampling frame
}
