% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame.R
\name{samples.sampling_frame}
\alias{samples.sampling_frame}
\title{Extract samples from a sampling_frame}
\usage{
\method{samples}{sampling_frame}(x, blockids = NULL, global = FALSE, ...)
}
\arguments{
\item{x}{A sampling_frame object.}

\item{blockids}{A numeric vector of block IDs to extract the samples from. If NULL (default), all block IDs are used.}

\item{global}{A logical value. If TRUE, the global time (with respect to the first block) of each sample/acquisition is returned. If FALSE (default), the relative time (with respect to the last block) of each sample/acquisition is returned.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of sample times extracted from the specified \code{sampling_frame}.
}
\description{
This function extracts the relative or global time of each sample/acquisition from a \code{sampling_frame}.
}
\examples{
frame <- sampling_frame(blocklens = c(100, 100, 100), TR = 2, precision = 0.5)
# The relative time (with respect to the last block) in seconds of each sample/acquisition
sam <- samples(frame)
# The global time (with respect to the first block) of each sample/acquisition
gsam <- samples(frame, global = TRUE)

}
