% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame.R
\name{sampling_frame}
\alias{sampling_frame}
\title{construct a \code{sampling_frame}}
\usage{
sampling_frame(blocklens, TR, start_time = TR/2, precision = 0.1)
}
\arguments{
\item{blocklens}{the number of scans in each block, a \code{vector}}

\item{TR}{the repetition time in seconds; i.e. the spacing between consecutive image acquisitions.}

\item{start_time}{the offset of first scan of each block (default is \code{TR/2})}

\item{precision}{the discrete sampling interval used for convolution with hemodynamic response function.}
}
\description{
A \code{sampling_frame} describes the block structure and temporal sampling of an fMRI paradigm.
}
\examples{

frame <- sampling_frame(blocklens=c(100,100, 100), TR=2, precision=.5)

## the relative time (with respect to the last block) in seconds of each sample/acquisition
sam <- samples(frame)
## the global time (with respect to first block) of each sample/acquisition
gsam <- samples(frame, global=TRUE)
}
