% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-utils.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize Strings for Use in R Names}
\usage{
sanitize(x, allow_dot = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{allow_dot}{Logical, if \code{FALSE}, dots (\code{.}) are replaced with underscores (\verb{_}).}
}
\value{
A sanitized character vector.
}
\description{
Wraps \code{make.names} but allows control over dot replacement.
}
\examples{
sanitize("a.b c")
sanitize("a.b c", allow_dot = FALSE)
}
