% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{Scale}
\alias{Scale}
\title{Z-score (global) basis}
\usage{
Scale(x)
}
\arguments{
\item{x}{numeric vector (NAs allowed)}
}
\value{
object of class c("Scale","ParametricBasis")
}
\description{
Z-score (global) basis
}
