% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{se}
\alias{se}
\title{Extract Standard Errors from Meta-Analysis}
\usage{
se(object)
}
\arguments{
\item{object}{An fmri_meta object}
}
\value{
Matrix of standard errors
}
\description{
Extract Standard Errors from Meta-Analysis
}
