% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-methods.R
\name{shift.Reg}
\alias{shift.Reg}
\title{Shift method for the Reg class (and thus regressor)}
\usage{
\method{shift}{Reg}(x, shift_amount, ...)
}
\arguments{
\item{x}{the Reg or regressor object to shift.}

\item{shift_amount}{A numeric value indicating the amount of time (in the same units as onsets, typically seconds)
to shift the onsets by. Positive values shift events later in time, negative values shift them earlier.}

\item{...}{extra args (unused).}
}
\value{
A new object of the same class as \code{x} but with the \code{onsets} field adjusted by \code{shift_amount}.
}
\description{
Modifies a regressor object by shifting all its event onsets by a specified amount.
}
\concept{regressor_methods}
