% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{shift.regressor}
\alias{shift.regressor}
\title{Shift method for the regressor class}
\usage{
\method{shift}{regressor}(x, shift_amount, ...)
}
\arguments{
\item{x}{the regressor object}

\item{shift_amount}{A numeric value indicating the amount of time to shift the object by.
Positive values will shift the object to the right, while negative values will shift it to the left.}

\item{...}{extra args}
}
\description{
Shift method for the regressor class
}
\seealso{
Other shift: 
\code{\link{shift}()}
}
\concept{shift}
