% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{shortnames}
\alias{shortnames}
\alias{shortnames.event_term}
\alias{shortnames.feature_term}
\alias{shortnames.event_model}
\title{Short Names}
\usage{
shortnames(x, ...)

\method{shortnames}{event_term}(x, ...)

\method{shortnames}{feature_term}(x, ...)

\method{shortnames}{event_model}(x, ...)
}
\arguments{
\item{x}{The object to generate short names for.}

\item{...}{Additional arguments.}
}
\value{
A character vector of short names.
}
\description{
Generate short names for model terms and conditions.
}
