% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_ts}
\alias{sim_ts}
\title{Simulate fMRI Time Series}
\usage{
sim_ts(ncond, nreps = 12, amps = rep(1, ncond), isi = c(3, 6), TR = 1.5)
}
\arguments{
\item{ncond}{The number of conditions to simulate.}

\item{nreps}{The number of repetitions per condition (default is 12).}

\item{amps}{A vector of amplitudes for each condition (default is a vector of 1s with length ncond).}

\item{isi}{A vector specifying the range of inter-stimulus intervals to sample from (default is c(3, 6)).}

\item{TR}{The repetition time of the fMRI acquisition (default is 1.5 seconds).}
}
\value{
A list with two elements:
\itemize{
\item onset: A vector of the onset times for each trial.
\item mat: A matrix containing the simulated fMRI time series, with time in the first column and the simulated responses for each condition in the subsequent columns.
}
}
\description{
This function simulates an fMRI time series for multiple conditions with specified parameters.
}
\keyword{internal}
