% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_fmri_noise}
\alias{simulate_fmri_noise}
\title{Simulate fMRI Noise using ARMA Model}
\usage{
simulate_fmri_noise(n, ar = c(0.3), ma = c(0.5), sd = 1, seed = NULL)
}
\arguments{
\item{n}{The number of time points in the fMRI time series.}

\item{ar}{A numeric vector containing autoregressive (AR) coefficients (default is c(0.3)).}

\item{ma}{A numeric vector containing moving average (MA) coefficients (default is c(0.5)).}

\item{sd}{The standard deviation of the white noise component (default is 1).}

\item{seed}{An optional seed for reproducibility (default is NULL).}
}
\value{
A numeric vector containing the simulated fMRI noise.
}
\description{
This function simulates fMRI noise using an autoregressive moving average (ARMA) model.
}
\keyword{internal}
