% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hrf.R
\name{smooth_hrf_assignments}
\alias{smooth_hrf_assignments}
\title{Smooth HRF Assignments}
\usage{
smooth_hrf_assignments(
  reassigned_hrfs,
  L_best_hrfs,
  voxel_coords,
  lambda = 0.1,
  n_iterations = 5,
  k = 6
)
}
\arguments{
\item{reassigned_hrfs}{An integer vector of length n_voxels}

\item{L_best_hrfs}{A matrix of reference HRFs for each label}

\item{voxel_coords}{A matrix of voxel coordinates (n_voxels x 3), or similar}

\item{lambda}{Not actually used, but left here as a param}

\item{n_iterations}{Number of smoothing passes}

\item{k}{Number of neighbors}
}
\value{
A vector of the same length as \code{reassigned_hrfs}, updated (smoothed)
}
\description{
This function tries to smooth out a voxel-wise label map of HRF assignments
by iterating over each voxel, checking its neighbors, picking the "most
common" neighbor label if that reduces cost, etc.
}
\keyword{internal}
