% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Soft-threshold function}
\usage{
soft_threshold(x, threshold)
}
\arguments{
\item{x}{A numeric vector of input values}

\item{threshold}{A non-negative threshold value for the soft-thresholding operation}
}
\value{
A numeric vector with the soft-thresholded values
}
\description{
This function applies soft-thresholding to the input values, setting values below the threshold to zero
and shrinking the remaining values by the threshold amount.
}
\examples{
x <- c(-3, -2, -1, 0, 1, 2, 3)
threshold <- 1
soft_thresholded <- soft_threshold(x, threshold)
print(soft_thresholded)

}
