% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{split_onsets}
\alias{split_onsets}
\title{split an onset vector into a list}
\usage{
split_onsets(x, ...)
}
\arguments{
\item{x}{the object associated with the onset vector}

\item{...}{extra args}
}
\description{
split an onset vector into a list
}
