% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_latent_lm.R
\name{standard_error.fmri_latent_lm}
\alias{standard_error.fmri_latent_lm}
\title{Calculate the standard error for an fmri_latent_lm object}
\usage{
\method{standard_error}{fmri_latent_lm}(x, type = c("estimates", "contrasts"), recon = FALSE, ...)
}
\arguments{
\item{x}{An fmri_latent_lm object.}

\item{type}{A character string specifying whether to return the standard error for "estimates" or "contrasts".}

\item{recon}{Logical, if TRUE, the function calculates the standard error for reconstructed data.}
}
\value{
A tibble containing the standard error values.
}
\description{
Calculate the standard error for an fmri_latent_lm object
}
\seealso{
fmri_latent_lm
}
