% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_methods.R
\name{standard_error.fmri_lm}
\alias{standard_error.fmri_lm}
\title{Extract Standard Errors from an fmri_lm Object}
\usage{
\method{standard_error}{fmri_lm}(x, type = c("estimates", "contrasts"), ...)
}
\arguments{
\item{x}{An \code{fmri_lm} object.}

\item{type}{The type of standard errors to extract: "estimates" or "contrasts". Default is "estimates".}

\item{...}{Additional arguments (currently unused).}
}
\value{
A tibble of standard errors.
}
\description{
This function extracts standard errors from an \code{fmri_lm} object.
}
