% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{Standardized}
\alias{Standardized}
\title{Standardized basis}
\usage{
Standardized(x)
}
\arguments{
\item{x}{a numeric vector to standardize. Missing values are allowed and will be replaced with 0 after standardization.}
}
\value{
an instance of class \code{Standardized} extending \code{ParametricBasis}
}
\description{
Standardize a numeric vector by centering and scaling, handling NAs appropriately
}
