% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{sub_basis}
\alias{sub_basis}
\alias{sub_basis.Scale}
\alias{sub_basis.ScaleWithin}
\alias{sub_basis.RobustScale}
\title{sub_basis}
\usage{
sub_basis(x, subset)

\method{sub_basis}{Scale}(x, subset)

\method{sub_basis}{ScaleWithin}(x, subset)

\method{sub_basis}{RobustScale}(x, subset)
}
\arguments{
\item{x}{the object}

\item{subset}{the subset}
}
\description{
subset a parametric basis regressor
}
