% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_ttest.R
\name{summary.fmri_ttest_fit}
\alias{summary.fmri_ttest_fit}
\title{Summary method for fmri_ttest_fit}
\usage{
\method{summary}{fmri_ttest_fit}(object, ...)
}
\arguments{
\item{object}{An fmri_ttest_fit object}

\item{...}{Additional summary arguments}
}
\value{
Invisibly returns the input object
}
\description{
Summary method for fmri_ttest_fit
}
