% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{summary.group_data}
\alias{summary.group_data}
\title{Summary of Group Data Object}
\usage{
\method{summary}{group_data}(object, ...)
}
\arguments{
\item{object}{A group_data object}

\item{...}{Additional summary arguments}
}
\description{
Summary of Group Data Object
}
