% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fdr.R
\name{summary.spatial_fdr_result}
\alias{summary.spatial_fdr_result}
\title{Summary of Spatial FDR Results}
\usage{
\method{summary}{spatial_fdr_result}(object, ...)
}
\arguments{
\item{object}{A spatial_fdr_result object}

\item{...}{Additional summary arguments}
}
\value{
Invisibly returns the input object
}
\description{
Summary of Spatial FDR Results
}
