% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{t_to_beta_se}
\alias{t_to_beta_se}
\title{Convert t-statistics to Effect Sizes}
\usage{
t_to_beta_se(t, df, n = NULL)
}
\arguments{
\item{t}{T-statistic values}

\item{df}{Degrees of freedom}

\item{n}{Sample size (optional, improves SE estimation)}
}
\value{
List with beta and se estimates
}
\description{
Helper function to convert t-statistics and df to beta and SE estimates
}
