% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{term_indices}
\alias{term_indices}
\title{term_indices}
\usage{
term_indices(x, ...)
}
\arguments{
\item{x}{the model/term object}

\item{...}{additional arguments}
}
\description{
the term indices for the associated design matrix
}

