% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{term_matrices.fmri_model}
\alias{term_matrices.fmri_model}
\title{Construct term matrices for an fMRI model}
\usage{
\method{term_matrices}{fmri_model}(x, blocknum = NULL)
}
\arguments{
\item{x}{An object of class "fmri_model" containing the event and baseline models.}

\item{blocknum}{(Optional) A numeric vector specifying the block numbers to be included in the term matrices.
By default, all unique block numbers in the event model are included.}
}
\value{
A named list of term matrices, with event terms followed by baseline terms.
Attributes "event_term_indices" and "baseline_term_indices" store the indices of event and baseline terms,
"blocknum" stores the block numbers, and "varnames" stores the variable names.
}
\description{
This function constructs the term matrices for an fMRI model, which consists of event-related terms
and baseline-related terms. The term matrices are used for building the design matrix in fMRI data analysis.
}
\seealso{
fmri_model
}
