% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hrf.R
\name{tidy.fmri_hrf_estimate}
\alias{tidy.fmri_hrf_estimate}
\title{Tidy a smooth HRF estimate}
\usage{
\method{tidy}{fmri_hrf_estimate}(x, curve = NULL, voxel = NULL, ...)
}
\arguments{
\item{x}{An \code{fmri_hrf_estimate} object.}

\item{curve}{Optional curve names to retain.}

\item{voxel}{Optional voxel names or indices to retain.}

\item{...}{Unused.}
}
\value{
A tibble with one row per time, curve, and voxel.
}
\description{
Tidy a smooth HRF estimate
}
