% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{to_glt}
\alias{to_glt}
\title{convert a contrast to an AFNI 'GLT'}
\usage{
to_glt(x, ...)
}
\arguments{
\item{x}{the contrast to convert}

\item{...}{extra args}
}
\description{
convert a contrast to an AFNI 'GLT'
}
