% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{unpack_chunkwise}
\alias{unpack_chunkwise}
\title{Unpack Chunkwise Results}
\usage{
unpack_chunkwise(cres, event_indices, baseline_indices)
}
\arguments{
\item{cres}{The results of the chunkwise analysis.}

\item{event_indices}{The indices of the event-related effects.}

\item{baseline_indices}{The indices of the baseline-related effects.}
}
\value{
A list containing the betas, contrasts, and statistical information from the chunkwise results.
}
\description{
This function processes and unpacks the results of chunkwise analysis in fMRI data.
}
\keyword{internal}
