% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{validate_group_data}
\alias{validate_group_data}
\title{Validate Group Data Object}
\usage{
validate_group_data(x)
}
\arguments{
\item{x}{A group_data object}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Validate Group Data Object
}
\keyword{internal}
