% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_csv.R
\name{validate_group_data_csv}
\alias{validate_group_data_csv}
\title{Validate group_data_csv Object}
\usage{
validate_group_data_csv(x)
}
\arguments{
\item{x}{A group_data_csv object}
}
\description{
Validate group_data_csv Object
}
\keyword{internal}
