% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_h5.R
\name{validate_group_data_h5}
\alias{validate_group_data_h5}
\title{Validate group_data_h5 Object}
\usage{
validate_group_data_h5(x)
}
\arguments{
\item{x}{A group_data_h5 object}
}
\description{
Validate group_data_h5 Object
}
\keyword{internal}
