% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{validate_group_data_nifti}
\alias{validate_group_data_nifti}
\title{Validate group_data_nifti Object}
\usage{
validate_group_data_nifti(x)
}
\arguments{
\item{x}{A group_data_nifti object}
}
\description{
Validate group_data_nifti Object
}
\keyword{internal}
