% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{validate_hrf_params}
\alias{validate_hrf_params}
\title{Validate HRF parameters}
\usage{
validate_hrf_params(type, params)
}
\arguments{
\item{type}{HRF type}

\item{params}{List of parameters}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate HRF parameters
}
\keyword{internal}
