% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_job.R
\name{validate_job}
\alias{validate_job}
\title{Validate an fmri_job}
\usage{
validate_job(x)
}
\arguments{
\item{x}{An \link{fmri_job}.}
}
\value{
\code{TRUE} invisibly if valid; otherwise an error is raised.
}
\description{
Validate an fmri_job
}
