% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_nifti.R
\name{validate_nifti_dimensions}
\alias{validate_nifti_dimensions}
\title{Validate NIfTI Dimensions}
\usage{
validate_nifti_dimensions(paths, expected_dim)
}
\arguments{
\item{paths}{Character vector of NIfTI file paths}

\item{expected_dim}{Expected dimensions}
}
\description{
Validate NIfTI Dimensions
}
\keyword{internal}
