% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_lm_result_schema.R
\name{variance_model}
\alias{variance_model}
\title{Extract the structured variance model}
\usage{
variance_model(x)
}
\arguments{
\item{x}{A fitted \code{fmri_lm} object.}
}
\value{
An \code{fmri_lm_variance_model} object.
}
\description{
Extract the structured variance model
}
