% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracridge.R
\name{vec_len}
\alias{vec_len}
\title{Calculate Vector Length}
\usage{
vec_len(vec, axis = 1)
}
\arguments{
\item{vec}{A numeric vector or matrix.}

\item{axis}{An integer specifying the axis to compute the norm across.
\code{1} for columns and \code{2} for rows. Default is \code{1}.}
}
\value{
A numeric vector containing the Euclidean norms.
}
\description{
Computes the Euclidean (L2) norm of a vector or the norm across a specified axis of a matrix.
}
\examples{
vec <- c(3, 4)
vec_len(vec)  # Should return 5

mat <- matrix(1:6, nrow = 2, ncol = 3)
vec_len(mat, axis = 1)  # Norms of rows
vec_len(mat, axis = 2)  # Norms of columns

}
