% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{welch_t_cpp}
\alias{welch_t_cpp}
\title{Welch two-sample t-test across features}
\usage{
welch_t_cpp(Y, g_in)
}
\arguments{
\item{Y}{S x P matrix}

\item{g}{Length S vector of group indicators (1/2 or 0/1)}
}
\value{
List with muA, muB, t, df (Welch), nA, nB
}
\description{
Welch two-sample t-test across features
}
