% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_csv.R
\name{wide_to_long_format}
\alias{wide_to_long_format}
\title{Convert Wide Format to Long Format}
\usage{
wide_to_long_format(df, effect_cols, subject_col)
}
\arguments{
\item{df}{Data frame in wide format}

\item{effect_cols}{Effect column specification}

\item{subject_col}{Subject column name}
}
\value{
Data frame in long format
}
\description{
Convert Wide Format to Long Format
}
\keyword{internal}
