% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{wrap_chunked_lm_results}
\alias{wrap_chunked_lm_results}
\title{a beastly function that unravels multiple chunk-wise model fits results ...}
\usage{
wrap_chunked_lm_results(cres, event_indices = NULL)
}
\description{
a beastly function that unravels multiple chunk-wise model fits results ...
}
\keyword{internal}
