% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{write_glt.glt_contrast}
\alias{write_glt.glt_contrast}
\title{Write GLT to File}
\usage{
\method{write_glt}{glt_contrast}(x, fname = NULL)
}
\arguments{
\item{x}{The GLT contrast to write.}

\item{fname}{File path for the GLT specification}
}
\description{
Write a GLT contrast to a file.
}
