% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{write_results}
\alias{write_results}
\title{Write Results from fMRI Analysis}
\usage{
write_results(x, ...)
}
\arguments{
\item{x}{A fitted fMRI model object}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisible list of created file paths
}
\description{
Generic function to export statistical maps and analysis results from fitted fMRI models
to standardized file formats with appropriate metadata.
}
\concept{result_export}
