% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_fit.R
\name{z_to_r}
\alias{z_to_r}
\title{Back-transform Fisher's Z to Correlation}
\usage{
z_to_r(z)
}
\arguments{
\item{z}{Numeric vector or matrix; Fisher's Z values}
}
\value{
Numeric vector or matrix; correlations
}
\description{
Back-transform Fisher's Z to Correlation
}
\seealso{
\code{\link{fmri_meta}}
}
