% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-afni.R
\name{afni_hrf}
\alias{afni_hrf}
\title{construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.}
\usage{
afni_hrf(
  ...,
  basis = c("spmg1", "block", "dmblock", "tent", "csplin", "poly", "sin", "sine", "gam",
    "gamma", "spmg2", "spmg3", "wav"),
  onsets = NULL,
  durations = NULL,
  prefix = NULL,
  subset = NULL,
  nbasis = 1,
  contrasts = NULL,
  id = NULL,
  lag = 0,
  precision = 0.3,
  summate = TRUE,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{...}{One or more variable names (bare or character) or expressions involving variables
present in the \code{data} argument of \code{event_model}.}

\item{basis}{the impulse response function or the name of a pre-supplied function,
one of: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian", "tent", "bs".
Can also be an \code{HRF} object.}

\item{onsets}{optional onsets override. If missing, onsets will be taken from the LHS of the main model formula.}

\item{durations}{optional durations override. If missing, durations argument from \code{event_model} is used.}

\item{prefix}{a character string that is prepended to the variable names and used to identify the term.
Can be used to disambiguate two \code{hrf} terms with the same variable(s) but different onsets or basis functions.}

\item{subset}{an expression indicating the subset of 'onsets' to keep.}

\item{nbasis}{number of basis functions -- only used for hemodynamic response functions (e.g. bspline) that take a variable number of bases.}

\item{contrasts}{one or more \code{contrast_spec} objects created with the \code{contrast}, \code{pair_contrast} etc. functions.
Must be NULL, a single contrast spec, or a \emph{named} list of contrast specs.}

\item{id}{a unique \code{character} identifier used to refer to term, otherwise will be determined from variable names.}

\item{lag}{a temporal offset in seconds which is added to onset before convolution}

\item{precision}{sampling precision in seconds.}

\item{summate}{whether impulse amplitudes sum up when duration is greater than 0.}

\item{start}{the start of the window for sin/poly/csplin models}

\item{stop}{the stop time for sin/poly/csplin models}
}
\value{
an \code{afni_hrfspec} instance
}
\description{
construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.
}
