% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{afni_hrf}
\alias{afni_hrf}
\title{construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.}
\usage{
afni_hrf(
  ...,
  basis = c("spmg1", "block", "dmblock", "tent", "csplin", "poly", "sin", "gam",
    "spmg2", "spmg3", "wav"),
  onsets = NULL,
  durations = NULL,
  prefix = NULL,
  subset = NULL,
  nbasis = 1,
  contrasts = NULL,
  id = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{...}{the variable names, all of which must be present in the enclosing environment (e.g. an \code{event_model} object)}

\item{basis}{the impulse response function or the name of a pre-supplied function, one of: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian".}

\item{onsets}{optional onsets override. If missing, onsets will be taken from the \code{event_model}}

\item{durations}{optional durations override. If missing, onsets will be taken from the \code{event_model}}

\item{prefix}{a character string that is prepended to the variable names and used to identify the term.
Can be used to disambiguate two \code{hrf} terms with the same variable(s) but different onsets or basis functions.}

\item{subset}{an expression indicating the subset of 'onsets' to keep}

\item{nbasis}{number of basis functions -- only used for hemodynamic response functions (e.g. bspline) that take a variable number of bases.}

\item{contrasts}{one or more \code{contrast_spec} objects created with the \code{contrast} function.
If multiple contrasts are required, then these should be wrapped in a \code{list} or \code{contrast_set}.}

\item{id}{a  unique \code{character} identifier used to refer to term, otherwise will be determined from variable names.}

\item{start}{the start of the window for sin/poly/csplin models}

\item{stop}{the stop time for sin/poly/csplin models}
}
\description{
construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.
}
