% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R
\name{afni_lm}
\alias{afni_lm}
\title{Set up an fMRI linear model for AFNI's 3dDeconvolve}
\usage{
afni_lm(
  fmri_mod,
  dataset,
  working_dir = ".",
  polort = -1,
  jobs = 1,
  censor = NULL,
  options = list()
)
}
\arguments{
\item{fmri_mod}{An fmri_model object containing the event and baseline models.}

\item{dataset}{An fmri_dataset object containing the scan data and other necessary information.}

\item{working_dir}{The working directory (default is the current directory).}

\item{polort}{The number of polynomial baseline regressors (default is to suppress 'polort').}

\item{jobs}{The number of jobs to use with '3dDeconvolve' (default is 1).}

\item{censor}{A list of censoring vectors, one per run, or a single vector equal to the total number of scans (default is NULL).}

\item{options}{A list of options to be sent to 3dDeconvolve (default is an empty list).}
}
\value{
An afni_lm_spec object containing the fitted model, dataset, working directory, options, and command.
}
\description{
This function prepares an fMRI linear model for AFNI's 3dDeconvolve tool.
}
