% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/regressor.R
\name{amplitudes}
\alias{amplitudes}
\alias{amplitudes.regressor}
\title{amplitudes}
\usage{
amplitudes(x)

\method{amplitudes}{regressor}(x)
}
\arguments{
\item{x}{the object}
}
\description{
return the amplitude vector
}
