% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline}
\alias{baseline}
\title{Create a model specification for modeling low-frequency drift in fmri time series.}
\usage{
baseline(degree = 1, basis = c("poly", "bs", "ns"),
  name = paste0("baseline_", basis, "_", degree), constant = FALSE)
}
\arguments{
\item{degree}{number of polynomial terms for each image block}

\item{basis}{the type of polynomial basis.}

\item{name}{the name of the term}

\item{constant}{whether to include an intercept term}
}
\description{
Create a model specification for modeling low-frequency drift in fmri time series.
}
