% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline}
\alias{baseline}
\title{Create a Baseline Specification.}
\usage{
baseline(
  degree = 1,
  basis = c("constant", "poly", "bs", "ns"),
  name = NULL,
  intercept = c("runwise", "global", "none")
)
}
\arguments{
\item{degree}{Number of basis terms per image block (ignored for "constant").}

\item{basis}{Type of basis ("constant", "poly", "bs", or "ns").}

\item{name}{Optional name for the term.}

\item{intercept}{Type of intercept to include ("runwise", "global", or "none").}
}
\value{
A baselinespec list instance.
}
\description{
Generates a baselinespec for modeling low-frequency drift in fMRI time series.
}
