% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(degree = 5, basis = c("bs", "poly", "ns")[1],
  name = paste0("Baseline_", basis, "_", degree))
}
\arguments{
\item{basis}{the type of polynomial basis.}

\item{name}{the name of the term}

\item{number}{of polynomial terms for each image block}
}
\description{
A matrix of polynomial regressors for modleing low-frequency drift in fmri time series.
}

