% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(degree = 5, basis = c("bs", "poly", "ns"),
  name = paste0("Baseline_", basis, "_", degree), constant = FALSE)
}
\arguments{
\item{basis}{the type of polynomial basis.}

\item{name}{the name of the term}

\item{number}{of polynomial terms for each image block}
}
\description{
A matrix of polynomial regressors for modeling low-frequency drift in fmri time series.
}

