% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline_model}
\alias{baseline_model}
\title{baseline_model}
\usage{
baseline_model(basis = "bs", degree = 5, sampling_frame,
  nuisance_list = NULL)
}
\arguments{
\item{nuisance_list}{a list of nusiance matrices, one per block}

\item{baseline_model}{}
}
\description{
baseline_model
}

