% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{baseline_model}
\alias{baseline_model}
\title{Construct a Baseline Model.}
\usage{
baseline_model(
  basis = c("constant", "poly", "bs", "ns"),
  degree = 1,
  sframe,
  intercept = c("runwise", "global", "none"),
  nuisance_list = NULL
)
}
\arguments{
\item{basis}{Character; type of basis function ("constant", "poly", "bs", or "ns").}

\item{degree}{Integer; degree of the spline/polynomial function.}

\item{sframe}{A sampling_frame object.}

\item{intercept}{Character; whether to include an intercept ("runwise", "global", or "none").
(Automatically set to FALSE when basis == "constant".)}

\item{nuisance_list}{Optional list of nuisance matrices (one matrix per fMRI block).}
}
\value{
An object of class "baseline_model".
}
\description{
Builds a baseline model to account for noise and non–event-related variance.
This model may include a drift term, a block intercept term, and nuisance regressors.
}
\examples{
sframe <- sampling_frame(blocklens = c(100, 100), TR = 2)
bmod <- baseline_model(basis = "bs", degree = 3, sframe = sframe)
bmod_global <- baseline_model(basis = "bs", degree = 3, sframe = sframe, intercept = "global")
bmod_nointercept <- baseline_model(basis = "bs", degree = 3, sframe = sframe, intercept = "none")
stopifnot(ncol(design_matrix(bmod)) == 8)
}
