% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{block}
\alias{block}
\title{a block variable, which is constant over the span of a scanning run}
\usage{
block(x)
}
\arguments{
\item{x}{the block variable}
}
\value{
an instance of a class \code{blockspec}
}
\description{
a block variable, which is constant over the span of a scanning run
}
