% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{block_term}
\alias{block_term}
\title{block_term}
\usage{
block_term(varname, blockids, expanded_blockids, mat)
}
\arguments{
\item{varname}{the name of the term}
}
\description{
construct a constant term
}
