% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basis.R
\name{BSpline}
\alias{BSpline}
\title{BSpline}
\usage{
BSpline(x, degree)
}
\arguments{
\item{x}{a numeric vector at which to evaluate the spline. Missing values are not allowed in x}

\item{degree}{the degree of the piecewise polynomial}
}
\description{
BSpline
}

