% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R, R/all_generic.R
\name{build_afni_stims}
\alias{build_afni_stims}
\alias{build_afni_stims.event_term}
\title{Build AFNI stimulus definitions}
\usage{
build_afni_stims(x, ...)
}
\arguments{
\item{x}{Term object}
\item{...}{Additional arguments}
}
\value{
List of stimulus descriptors or \code{NULL}
}
\description{
Generate stimulus files for AFNI's 3dDeconvolve.
}
\family{AFNI}
\keyword{internal}
