% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{build_afni_stims}
\alias{build_afni_stims}
\title{generate a set of AFNI stimuli for '3dDeconvolve'}
\usage{
build_afni_stims(x, ...)
}
\arguments{
\item{x}{the term}

\item{...}{extra args}
}
\description{
generate a set of AFNI stimuli for '3dDeconvolve'
}
\keyword{internal}
