% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_vector.R
\name{cells}
\alias{cells}
\alias{cells.event_factor}
\alias{cells.event_term}
\alias{cells.covariate_convolved_term}
\title{The experimental cells of a design}
\usage{
cells(x, ...)

\method{cells}{event_factor}(x, drop.empty = TRUE, ...)

\method{cells}{event_term}(x, drop.empty = TRUE, ...)

\method{cells}{covariate_convolved_term}(x, ...)
}
\arguments{
\item{x}{An event_term object.}

\item{...}{Additional arguments to be passed to the function.}

\item{drop.empty}{Logical. If TRUE, empty cells will be removed.}
}
\value{
A list of data frames containing the cells of the event_factor object.

A tibble containing the cells of the event_term object, with an
additional "count" attribute.
}
\description{
return the experimental cells that are in a model term as a table

This function extracts cells from an event_factor object, optionally removing
empty cells. Note that the removal of empty cells is not implemented.

This function extracts cells from an event_term object, optionally removing
empty cells.
}
\examples{

evlist <- list(fac1=factor(c("A", "B", "A", "B")), 
               fac2=factor(c("1", "1", "2", "2")))
eterm <- event_term(evlist,onsets=1:4, blockids=rep(1,4))
cells(eterm)
}
