% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{chunkwise_lm.fmri_dataset}
\alias{chunkwise_lm.fmri_dataset}
\title{Chunkwise Linear Model for fMRI Dataset}
\usage{
\method{chunkwise_lm}{fmri_dataset}(
  dset,
  model,
  conlist,
  fcon,
  nchunks,
  robust = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dset}{An object of class \code{fmri_dataset}.}

\item{model}{The fMRI model used for the analysis.}

\item{conlist}{The list of contrasts used in the analysis.}

\item{fcon}{The F-contrasts used in the analysis.}

\item{nchunks}{The number of chunks to divide the dataset into.}

\item{robust}{Whether to use robust linear modeling (default is FALSE).}

\item{verbose}{Whether to display progress messages (default is FALSE).}
}
\value{
A list containing the unpacked chunkwise results.
}
\description{
This function performs a chunkwise linear model analysis on fMRI dataset,
splitting the dataset into chunks and running the linear model on each chunk.
}
\keyword{internal}
