% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{chunkwise_lm.fmri_dataset}
\alias{chunkwise_lm.fmri_dataset}
\title{Perform Chunkwise Linear Modeling on fMRI Dataset}
\usage{
\method{chunkwise_lm}{fmri_dataset}(
  dset,
  model,
  conlist,
  fcon,
  nchunks,
  robust = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dset}{An \code{fmri_dataset} object.}

\item{model}{The \code{fmri_model} used for the analysis.}

\item{conlist}{The list of contrasts used in the analysis.}

\item{fcon}{The F-contrasts used in the analysis.}

\item{nchunks}{The number of chunks to divide the dataset into.}

\item{robust}{Logical. Whether to use robust linear modeling (default is \code{FALSE}).}

\item{verbose}{Logical. Whether to display progress messages (default is \code{FALSE}).}
}
\value{
A list containing the unpacked chunkwise results.
}
\description{
This function performs a chunkwise linear model analysis on an fMRI dataset,
splitting the dataset into chunks and running the linear model on each chunk.
}
\keyword{internal}
