% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{chunkwise_lm}
\alias{chunkwise_lm}
\title{Run glm for each data chunk (responses split horizontally) and then concatenate chunks}
\usage{
chunkwise_lm(
  dset,
  model,
  conlist,
  fcon,
  nchunks,
  robust = FALSE,
  verbose = FALSE,
  full_dof = FALSE
)
}
\description{
Run glm for each data chunk (responses split horizontally) and then concatenate chunks
}
\keyword{internal}
