% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{chunkwise_rlm}
\alias{chunkwise_rlm}
\title{Perform Chunkwise Robust Linear Modeling on fMRI Dataset}
\usage{
chunkwise_rlm(dset, model, conlist, fcon, nchunks, verbose = FALSE)
}
\arguments{
\item{dset}{An \code{fmri_dataset} object.}

\item{model}{The \code{fmri_model} used for the analysis.}

\item{nchunks}{The number of chunks to divide the dataset into.}

\item{verbose}{Logical. Whether to display progress messages (default is \code{FALSE}).}
}
\description{
Perform Chunkwise Robust Linear Modeling on fMRI Dataset
}
\keyword{internal}
