% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{coef_image}
\alias{coef_image}
\title{Extract Image/Volume for Coefficient}
\usage{
coef_image(object, coef = 1, statistic = c("estimate", "se", "z", "p"))
}
\arguments{
\item{object}{An fmri_meta object}

\item{coef}{Coefficient name or index}

\item{statistic}{Type of statistic to extract ("estimate", "se", "z", "p")}
}
\value{
NeuroVol object or matrix
}
\description{
Extract Image/Volume for Coefficient
}
\examples{
toy_meta <- structure(
  list(
    coefficients = matrix(c(0.3, 0.1), nrow = 1,
      dimnames = list(NULL, c("A", "B"))),
    se = matrix(c(0.05, 0.06), nrow = 1)
  ),
  class = "fmri_meta"
)
coef_image(toy_meta, coef = "A")
}
