% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{columns}
\alias{columns}
\title{Extract Column Names or Identifiers}
\usage{
columns(x)
}
\arguments{
\item{x}{An object (typically a ParametricBasis)}
}
\value{
A character vector of column identifiers
}
\description{
Extract column names or identifiers from an object. For parametric basis objects,
this returns tokens representing the type of variables (categorical or continuous).
}
