% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_vector.R
\name{columns}
\alias{columns}
\alias{columns.event_term}
\alias{columns.event_seq}
\alias{columns.event_matrix}
\alias{columns.event_set}
\alias{columns.event_basis}
\title{columns}
\usage{
columns(x)

\method{columns}{event_term}(x)

\method{columns}{event_seq}(x)

\method{columns}{event_matrix}(x)

\method{columns}{event_set}(x)

\method{columns}{event_basis}(x)
}
\arguments{
\item{x}{the term}
}
\description{
return the column labels associated with the elements of a term.
}
\keyword{internal}
