% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{conditions.event_term}
\alias{conditions.event_term}
\title{extract conditions of an \code{event_term}}
\usage{
\method{conditions}{event_term}(x, drop.empty = TRUE, ...)
}
\arguments{
\item{x}{the model term}

\item{drop.empty}{whether to remove conditions with no cells (default is TRUE)}

\item{...}{extra args}
}
\description{
extract conditions of an \code{event_term}
}
\seealso{
Other conditions: 
\code{\link{conditions}()}
}
\concept{conditions}
