% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_generic.R
\name{conditions}
\alias{conditions}
\title{conditions}
\usage{
conditions(x, ...)
}
\arguments{
\item{x}{the model term}

\item{...}{}
}
\description{
return the set of condition labels associated with a model term
}

