% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{conditions}
\alias{conditions}
\title{Conditions}
\usage{
conditions(x, ...)
}
\arguments{
\item{x}{the model term}

\item{...}{extra args}
}
\description{
return the set of condition labels associated with a model term
}
\seealso{
Other conditions: 
\code{\link{conditions.event_term}()}
}
\concept{conditions}
