% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{conditions}
\alias{conditions}
\alias{conditions.event_model}
\title{conditions}
\usage{
conditions(x, ...)

\method{conditions}{event_model}(x, ...)
}
\arguments{
\item{x}{the model term}

\item{...}{extra args}
}
\description{
return the set of condition labels associated with a model term
}
