% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regressor.R
\name{conformLen}
\alias{conformLen}
\title{check vector is of length 1 or repeated for supplied length}
\usage{
conformLen(val, len)
}
\description{
check vector is of length 1 or repeated for supplied length
}

