<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc
% Please edit documentation in R/regressor.R
\name{conformLen}
\alias{conformLen}
\title{check vector is of length 1 or repeated for supplied length}
\usage{
conformLen(val, len)
}
\description{
check vector is of length 1 or repeated for supplied length
}

