% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{construct_event_term_from_spec}
\alias{construct_event_term_from_spec}
\title{Construct Event Term from Regressor Specification}
\usage{
construct_event_term_from_spec(regressor_spec, events_data, hrf_objects)
}
\arguments{
\item{regressor_spec}{A list containing the specification of a single regressor}

\item{events_data}{A data frame containing event information}

\item{hrf_objects}{A named list of HRF objects}
}
\value{
An hrfspec object or similar
}
\description{
This function constructs an event term based on the regressor specification,
events data, and HRF objects.
}
