% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Contrast Specification}
\usage{
contrast(form, name, where = NULL)
}
\arguments{
\item{form}{A formula describing the contrast.}

\item{name}{A character label for the contrast.}

\item{where}{An expression defining the subset over which the contrast is applied (default: NULL).}
}
\value{
A list containing the contrast specification.
}
\description{
Define a linear contrast using a formula expression.
}
\examples{
# A minus B contrast
contrast(~ A - B, name="A_B")

# With subsetting
contrast(~ A - B, name="A_B_block1", where = ~ block == 1)

}
