% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast_set}
\alias{contrast_set}
\title{contrast_set}
\usage{
contrast_set(...)
}
\arguments{
\item{...}{a variable length list of \code{contrast_spec} objects.}
}
\description{
A \code{list} o \code{contrast_spec} objects
}
\examples{
c1 <- contrast(~ A - B, name="A_B")
c2 <- contrast(~ B - C, name="B_C")
contrast_set(c1,c2)

}
