% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast_set}
\alias{contrast_set}
\title{Create a Set of Contrasts}
\usage{
contrast_set(...)
}
\arguments{
\item{...}{A variable-length list of contrast_spec objects.}
}
\value{
A list of contrast_spec objects with class "contrast_set".
}
\description{
Construct a list of contrast_spec objects.
}
\examples{
c1 <- contrast(~ A - B, name="A_B")
c2 <- contrast(~ B - C, name="B_C")
contrast_set(c1,c2)

}
