% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{contrast_weights}
\alias{contrast_weights}
\alias{contrast_weights.convolved_term}
\alias{contrast_weights.event_model}
\alias{contrast_weights.fmri_term}
\alias{design_matrix.convolved_term}
\alias{matrix_term}
\title{contrast_weights}
\usage{
contrast_weights(x, term, ...)

\method{contrast_weights}{convolved_term}(x)

\method{contrast_weights}{fmri_term}(x)

\method{contrast_weights}{event_model}(x)

\method{design_matrix}{convolved_term}(x, blockid = NULL)

matrix_term(varname, mat)
}
\arguments{
\item{x}{the object}
}
\description{
contrast_weights
}

