% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{contrast_weights}
\alias{contrast_weights}
\alias{contrast_weights.convolved_term}
\alias{contrast_weights.fmri_term}
\alias{contrast_weights.fmri_model}
\alias{contrast_weights.event_model}
\title{contrast_weights}
\usage{
contrast_weights(x, term, ...)

\method{contrast_weights}{convolved_term}(x)

\method{contrast_weights}{fmri_term}(x)

\method{contrast_weights}{fmri_model}(x)

\method{contrast_weights}{event_model}(x)
}
\arguments{
\item{x}{the object}

\item{term}{}

\item{...}{extra args}
}
\description{
contrast_weights
}
