% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{convolve.event_term}
\alias{convolve.event_term}
\title{Convolve an event-related design matrix with an HRF.}
\usage{
\method{convolve}{event_term}(
  x,
  hrf,
  sampling_frame,
  drop.empty = TRUE,
  summate = TRUE,
  precision = 0.3,
  ...
)
}
\arguments{
\item{x}{A data frame containing the input design matrix.}

\item{hrf}{A Hemodynamic Response Function to convolve the design matrix with.}

\item{sampling_frame}{A data frame specifying the sampling frame for the analysis.}

\item{drop.empty}{Logical. If TRUE, empty rows in the design matrix will be removed.}

\item{summate}{Logical. If TRUE, the convolved design matrix will be summed.}

\item{precision}{Numeric. The desired precision for the calculations.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A convolved design matrix, in tibble format.
}
\description{
This function takes an event-related design matrix and convolves it with
a specified Hemodynamic Response Function (HRF) to create a new design matrix
suitable for fMRI analysis. It also supports additional arguments for
flexibility and customization.
}
