% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{convolve}
\alias{convolve}
\title{convolve}
\usage{
convolve(x, hrf, sampling_frame, ...)
}
\arguments{
\item{x}{an event sequence}

\item{hrf}{a hemodynamic response function}

\item{sampling_frame}{the time series grid over which to sample the function.}

\item{...}{extra args}
}
\description{
This function convolves an event sequence \code{x} with a hemodynamic response function \code{hrf}
over a specified time series grid \code{sampling_frame}. It is a generic function that dispatches
to the appropriate method
}
\details{
Convolve a term \code{x} with a hemodynamic response function (HRF) over a \code{sampling_frame}.
}
