% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{convolve}
\alias{convolve}
\title{convolve}
\usage{
convolve(x, hrf, sampling_frame, ...)
}
\arguments{
\item{x}{an event sequence}

\item{hrf}{a hemodynamic response function}

\item{sampling_frame}{the time series grid over which to sample the function.}

\item{...}{extra args}
}
\description{
convolve a term \code{x} with a hemodynamic response over a \code{sampling_frame}
}
