% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{convolve_design}
\alias{convolve_design}
\title{Convolve HRF with Design Matrix}
\usage{
convolve_design(hrf, dmat, globons, durations, summate = TRUE)
}
\arguments{
\item{hrf}{A numeric vector representing the hemodynamic response function}

\item{dmat}{A design matrix with columns representing different experimental conditions}

\item{globons}{A numeric vector of global onsets for each event}

\item{durations}{A numeric vector of event durations}

\item{summate}{A logical value indicating whether to summate the convolved HRF (default: TRUE)}
}
\value{
A list of regressors, one for each column in the design matrix.
}
\description{
This function convolves a given hemodynamic response function (HRF) with the design matrix of an fMRI study.
It is useful for modeling the expected BOLD signal in response to experimental conditions.
}
