% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{convolve_event_term}
\alias{convolve_event_term}
\title{Convolve Event Term with HRF}
\usage{
convolve_event_term(event_term_obj, hrf_obj, sampling_frame)
}
\arguments{
\item{event_term_obj}{An event term object}

\item{hrf_obj}{An HRF object}

\item{sampling_frame}{A data frame defining the time points, must have 'time' col}
}
\value{
A numeric vector containing the convolved signal
}
\description{
This function convolves an event term with the specified HRF object
over the sampling frame.
}
