% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{correlation_map}
\alias{correlation_map}
\title{correlation_map}
\usage{
correlation_map(x, ...)
}
\arguments{
\item{x}{The model object (event_model, baseline_model, or fmri_model).}

\item{...}{Additional arguments for the correlation plot.}
}
\description{
A generic function to create a correlation heatmap for an fMRI design matrix.
}
