% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{covariate}
\alias{covariate}
\title{construct a covariate term that is not convolved with a hemodynamic response function}
\usage{
covariate(..., data, id = NULL, prefix = NULL)
}
\arguments{
\item{...}{a variable argument set of covariate names}

\item{data}{a data.frame containing the variables}

\item{id}{a  unique \code{character} identifier used to refer to term, otherwise will be determined from variable names.}

\item{prefix}{a character string that is prepended to the variable names and used to identify the term. 
Can be used to disambiguate two \code{hrf} terms with the same variable(s) but different onsets or basis functions.}
}
\description{
construct a covariate term that is not convolved with a hemodynamic response function
}
\examples{
df1 <- data.frame(x=rnorm(100), y=rnorm(100))
cv <- covariate(x,y,data=df1)
}
