% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{create_hrfspec}
\alias{create_hrfspec}
\title{Create an HRF specification programmatically}
\usage{
create_hrfspec(
  variable,
  hrf,
  onsets,
  durations = NULL,
  blockids = NULL,
  data = NULL
)
}
\arguments{
\item{variable}{Name of variable or parametric basis object}

\item{hrf}{HRF object to use}

\item{onsets}{Vector of onset times}

\item{durations}{Vector of durations (optional)}

\item{data}{Data frame containing variables (for parametric bases)}
}
\value{
An hrfspec object
}
\description{
Create an HRF specification programmatically
}
\keyword{internal}
