% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{create_parametric_basis}
\alias{create_parametric_basis}
\title{Create parametric basis from specification}
\usage{
create_parametric_basis(basis_spec, data)
}
\arguments{
\item{basis_spec}{Basis specification from YAML}

\item{data}{Data frame containing the variable}
}
\value{
A ParametricBasis object
}
\description{
Create parametric basis from specification
}
\keyword{internal}
