% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{data_chunks.matrix_dataset}
\alias{data_chunks.matrix_dataset}
\title{Create Data Chunks for matrix_dataset Objects}
\usage{
\method{data_chunks}{matrix_dataset}(x, nchunks = 1, runwise = FALSE, ...)
}
\arguments{
\item{x}{An object of class 'matrix_dataset'}

\item{nchunks}{The number of chunks to split the data into. Default is 1.}

\item{runwise}{If TRUE, creates run-wise chunks instead of arbitrary chunks}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of data chunks, each containing data, indices and chunk number
}
\description{
This function creates data chunks for matrix_dataset objects. It allows for the retrieval
of run-wise or sequence-wise data chunks, as well as arbitrary chunks.
}
