% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{data_chunks}
\alias{data_chunks}
\title{data_chunks}
\usage{
data_chunks(x, nchunks, ...)
}
\arguments{
\item{x}{the dataset}

\item{nchunks}{the numbe rof data chunks}

\item{...}{extra args}
}
\value{
an \code{iterator} returning on data chunk per iteration
}
\description{
return a set of data chunks
}
