% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R, R/fmri_model.R
\name{design_matrix}
\alias{design_matrix}
\alias{design_matrix.event_model}
\alias{design_matrix.convolved_term}
\alias{design_matrix.matrix_term}
\alias{design_matrix.fmri_model}
\title{design_matrix}
\usage{
design_matrix(x, ...)

\method{design_matrix}{event_model}(x, blockid = NULL, ...)

\method{design_matrix}{convolved_term}(x, blockid = NULL, ...)

\method{design_matrix}{matrix_term}(x, ...)

\method{design_matrix}{fmri_model}(x, blockid = NULL, ...)
}
\arguments{
\item{x}{the term}

\item{...}{additional arguments}

\item{blockid}{the block id to extract}
}
\description{
construct a design matrix from the term
}
