% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{design_matrix}
\alias{design_matrix}
\alias{design_matrix.event_model}
\alias{design_matrix.matrix_term}
\title{design_matrix}
\usage{
design_matrix(x, ...)

\method{design_matrix}{event_model}(x, blockid = NULL)

\method{design_matrix}{matrix_term}(x)
}
\arguments{
\item{x}{the term}

\item{...}{additional arguments}
}
\description{
construct a design matrix from the term
}
