% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despike.R
\name{despike}
\alias{despike}
\title{Despike Time Series Data}
\usage{
despike(x, k = 7, thresh = 6)
}
\arguments{
\item{x}{A numeric vector containing the time series data.}

\item{k}{The window size for the running median (default: 7).}

\item{thresh}{The threshold for identifying spikes (default: 6).}
}
\value{
A numeric vector with the spikes removed.
}
\description{
Remove spikes from time series data using running median and a threshold.
}
