% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/regressor.R
\name{durations}
\alias{durations}
\alias{durations.regressor}
\title{durations}
\usage{
durations(x)

\method{durations}{regressor}(x)
}
\arguments{
\item{x}{the object}
}
\description{
return a `durations` vector
}
