% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{estimate_betas}
\alias{estimate_betas}
\title{estimate trialwise beta coefficients for a dataset}
\usage{
estimate_betas(x, ...)
}
\arguments{
\item{x}{the dataset}

\item{...}{extra args}
}
\description{
estimate trialwise beta coefficients for a dataset
}
