% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/fmri_betas.R
\name{estimate_betas}
\alias{estimate_betas}
\alias{estimate_betas.fmri_dataset}
\alias{estimate_betas.matrix_dataset}
\title{estimate trialwise beta coefficients for a dataset}
\usage{
estimate_betas(x, ...)

\method{estimate_betas}{fmri_dataset}(
  x,
  fixed = NULL,
  ran,
  block,
  method = c("mixed", "pls", "pls_searchlight", "pls_global", "ols"),
  basemod = NULL,
  radius = 8,
  niter = 8,
  ncomp = 4,
  lambda = 0.01
)

\method{estimate_betas}{matrix_dataset}(
  x,
  fixed,
  ran,
  block,
  method = c("mixed", "pls", "pls_searchlight", "pls_global"),
  basemod = NULL,
  radius = 8,
  niter = 8,
  ncomp = 4,
  lambda = 0.01
)
}
\arguments{
\item{x}{the dataset}

\item{...}{extra args}

\item{fixed}{the fixed regressors that model constant effects (i.e. non-varying over trials)}

\item{ran}{the random (trialwise) regressors that model trialwise effects}

\item{block}{the block factor}

\item{method}{the regression method for estimating trialwise betas}

\item{basemod}{\code{baseline_model} instance to regress out of data before bet estimation}

\item{niter}{number of searchlight iterations for method "pls_searchlight"}

\item{ncomp}{number of pls components for method "pls" and "pls_searchlight" and "pls_global"}

\item{lambda}{lambda parameter (not currently used)}
}
\description{
estimate trialwise beta coefficients for a dataset
}
\details{
The \code{method} arguments allows for several beta estimation approaches
\itemize{
\item \code{mixed} uses a linear mixed effects modeling of trialwise random effects as implemented in the \code{rrBLUP} package.
\item \code{pls} uses separate partial least squares for each voxel to estimate trialwise betas
\item \code{pls_searchlight} estimates pls solutions over searchlight windows and averages the beta estimates
\item \code{pls_global} estimates a single multiresponse pls solution, where the \code{Y} matrix is the full data matrix.
\item \code{ols} ordinary least squares estimate of betas -- no regularization
}

facedes <- read.table(system.file("extdata", "face_design.txt", package = "fmrireg"), header=TRUE)
facesdes$frun <- factor(facedes$run)
scans <- paste0("rscan0", 1:6, ".nii")

dset <- fmri_dataset(scans=scans,mask="mask.nii",TR=1.5,run_length=rep(436,6),event_table=facedes)
fixed = onset ~ hrf(run)
ran = onset ~ trialwise()
block = ~ run
}
\examples{

}
