% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_betas.R
\name{estimate_betas}
\alias{estimate_betas}
\title{estimate trialwise betas for an fMRI dataset.}
\usage{
estimate_betas(fixed, ran, block, dataset, method = c("ridge", "pls",
  "pls_searchlight"), basedeg = 5, nuisance_list = NULL, radius = 8,
  niter = 8, ncomp = 4, lambda = 0.01)
}
\description{
estimate trialwise betas for an fMRI dataset.
}
\examples{
facedes <- read.table(system.file("extdata", "face_design.txt", package = "fmrireg"), header=TRUE)
facesdes$frun <- factor(facedes$run)
scans <- paste0("rscan0", 1:6, ".nii")
dset <- fmri_dataset(scans=scans,mask="mask.nii",TR=1.5,run_length=rep(436,6),event_table=facedes)
fixed = onset ~ hrf(run)
ran = onset ~ trialwise()
block = ~ run
}
