% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_betas.R
\name{estimate_hrf}
\alias{estimate_hrf}
\title{Estimate hemodynamic response function (HRF) using Generalized Additive Models (GAMs)}
\usage{
estimate_hrf(
  form,
  fixed = NULL,
  block,
  dataset,
  bs = c("tp", "ts", "cr", "ps"),
  rsam = seq(0, 20, by = 1),
  basemod = NULL,
  k = 8,
  fx = TRUE
)
}
\arguments{
\item{form}{A formula specifying the event model for the conditions of interest}

\item{fixed}{A formula specifying the fixed regressors that model constant effects (i.e., non-varying over trials); default is NULL}

\item{block}{A formula specifying the block factor}

\item{dataset}{An object representing the fMRI dataset}

\item{bs}{Basis function for the smooth term in the GAM; one of "tp" (default), "ts", "cr", or "ps"}

\item{rsam}{A sequence of time points at which the HRF is estimated (default: seq(0, 20, by = 1))}

\item{basemod}{A \code{baseline_model} instance to regress out of data before HRF estimation (default: NULL)}

\item{k}{the dimension of the basis, default is 8}

\item{fx}{indicates whether the term is a fixed d.f. regression spline (TRUE) or a penalized regression spline (FALSE); default is TRUE.}
}
\value{
A matrix with the estimated HRF values for each voxel
}
\description{
This function estimates the HRF using GAMs from the \code{mgcv} package.
The HRF can be estimated with or without fixed effects.
}
\examples{

# To be added

}
\seealso{
\code{\link{baseline_model}}, \code{\link{event_model}}, \code{\link{design_matrix}}
}
