% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{EV}
\alias{EV}
\title{EV}
\usage{
EV(
  vals,
  name,
  onsets,
  blockids,
  durations = 1,
  subset = rep(TRUE, length(onsets))
)
}
\arguments{
\item{vals}{Event values.}

\item{name}{Name of the event variable.}

\item{onsets}{Numeric vector of event onsets.}

\item{blockids}{Numeric vector of block IDs.}

\item{durations}{Numeric vector of event durations (or scalar).}

\item{subset}{Logical vector indicating which events to keep.}
}
\value{
An event object.
}
\description{
Factory function for creating event objects (e.g. event_factor, event_variable,
event_basis, event_matrix) based on the type of input values.
}
\examples{
ev_fac <- EV(factor(c("A", "B", "C")), "fac", onsets = c(1, 10, 20), blockids = rep(1, 3))

}
\keyword{internal}
