% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{evaluate.HRF}
\alias{evaluate.HRF}
\title{evaluate an HRF for a single event along a set of points}
\usage{
\method{evaluate}{HRF}(x, grid, amplitude = 1, duration = 0,
  precision = 0.1)
}
\arguments{
\item{grid}{the sampling grid, e.g. the points in time at which the HRF will be evaluated.}

\item{amplitude}{the scaling value for the event}

\item{duration}{the duration of the event}

\item{precision}{the temporal resolution used for computing summed responses when duration > 0}
}
\description{
evaluate an HRF for a single event along a set of points
}

