% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{evaluate.HRF}
\alias{evaluate.HRF}
\title{evaluate.HRF}
\usage{
\method{evaluate}{HRF}(
  x,
  grid,
  amplitude = 1,
  duration = 0,
  precision = 0.2,
  summate = TRUE,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{x}{The HRF function.}

\item{grid}{A vector of time points.}

\item{amplitude}{The scaling value for the event (default: 1).}

\item{duration}{The duration of the event (default: 0).}

\item{precision}{The temporal resolution used for computing summed responses when duration > 0 (default: 0.2).}

\item{summate}{Whether the HRF response increases its amplitude as a function of stimulus duration (default: TRUE).}

\item{normalize}{Scale output so that peak is 1 (default: FALSE).}
}
\value{
A vector of HRF values at the specified time points.
}
\description{
This function evaluates a hemodynamic response function (HRF) for a given set of time points (grid) and other parameters.
It is used to generate the predicted BOLD (blood-oxygen-level-dependent) signal changes in fMRI data analysis.
}
\examples{
hrf1 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=.1)
hrf2 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=.1, summate=FALSE)
}
