% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hrf.R, R/regressor.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.HRF}
\alias{evaluate.regressor}
\title{evaluate a function over a sampling grid}
\usage{
evaluate(x, grid, ...)

\method{evaluate}{HRF}(
  x,
  grid,
  amplitude = 1,
  duration = 0,
  precision = 0.2,
  summate = TRUE,
  normalize = FALSE
)

\method{evaluate}{regressor}(x, grid, precision = 0.2)
}
\arguments{
\item{x}{the object to evaluate}

\item{grid}{the sampling grid. A vector of real values in seconds.}

\item{...}{extra args}

\item{amplitude}{the scaling value for the event}

\item{duration}{the duration of the event}

\item{precision}{the sampling precision for the hrf. This parameter is passed to \code{evaluate.HRF}}

\item{summate}{whether the HRF response increases its amplitude as a function of stimulus duration.}

\item{normalize}{scale output so that peak is 1.}
}
\description{
given an object to be evaluated and an input sample "grid", evaluate the object.

evalute a regressor function over a sampling grid
}
\examples{

hrf1 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=1)

# the same, except now turn off temporal summation.
hrf2 <- evaluate(HRF_SPMG1, grid=seq(0,20,by=1.5), duration=2, precision=1,summate=FALSE)

frame <- sampling_frame(blocklens=100, TR=2)
reg <- regressor(onsets=c(10,20), hrf=HRF_SPMG1)
evaluate(reg, samples(frame))
}
