% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hrf.R, R/regressor.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.HRF}
\alias{evaluate.regressor}
\title{evaluate}
\usage{
evaluate(x, grid, ...)

\method{evaluate}{HRF}(x, grid, amplitude = 1, duration = 0,
  precision = 0.1)

\method{evaluate}{regressor}(x, grid, precision = 0.1)
}
\arguments{
\item{x}{the object to evaluate}

\item{grid}{the sampling grid}

\item{...}{extra args}

\item{amplitude}{the scaling value for the event}

\item{duration}{the duration of the event}

\item{precision}{the temporal resolution used for computing summed responses when duration > 0}

\item{grid}{the sampling grid. A vector of real values in seconds.}

\item{precision}{the sampling precision for the hrf. This parameter is passed to \code{evaluate.HRF}}
}
\description{
evaluate a regression or hemodynamic response function
}
\examples{
frame <- sampling_frame(blocklens=100, TR=2)
reg <- regressor(onsets=c(10,20), hrf=HRF_SPMG1)
evaluate(reg, samples(frame))
}
