<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_basis}
\alias{event_basis}
\title{event_basis}
\usage{
event_basis(basis, onsets, blockids = 1, durations = NULL)
}
\arguments{
\item{basis}{}

\item{onsets}{}

\item{blockids}{}

\item{durations}{}
}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_basis}
\alias{event_basis}
\title{Create a event set from a basis object of type \code{\linkS4class{ParametricBasis}}.}
\usage{
event_basis(basis, onsets, blockids = 1, durations = NULL)
}
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc
\description{
Create a event set from a basis object of type \code{\linkS4class{ParametricBasis}}.
}

