% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_basis}
\alias{event_basis}
\title{event_basis}
\usage{
event_basis(
  basis,
  onsets,
  blockids = 1,
  durations = NULL,
  subset = rep(TRUE, length(onsets))
)
}
\arguments{
\item{basis}{the basis object}

\item{onsets}{the event onsets.}

\item{blockids}{the block ids associated with each event (must be non-decreasing)}

\item{durations}{the duration of each event.}

\item{subset}{a \code{logical} vector indicating the subset of events to keep}
}
\description{
Create a event set from a basis object of type \code{\linkS4class{ParametricBasis}}.
}
