% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_basis}
\alias{event_basis}
\title{event_basis}
\usage{
event_basis(basis, onsets, blockids = 1, durations = NULL)
}
\description{
Create a event set from a basis object of type \code{\linkS4class{ParametricBasis}}.
}

