% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_basis}
\alias{event_basis}
\title{Create an event set from a ParametricBasis object.}
\usage{
event_basis(
  basis,
  onsets,
  blockids = 1,
  durations = 0,
  subset = rep(TRUE, length(onsets))
)
}
\arguments{
\item{basis}{A ParametricBasis object.}

\item{onsets}{Numeric vector of onsets.}

\item{blockids}{Numeric vector of block IDs (default: 1).}

\item{durations}{Numeric vector of durations (default: 0).}

\item{subset}{Logical vector for subsetting (default: rep(TRUE, length(onsets))).}
}
\value{
An object of class "event_basis" and "event_seq".
}
\description{
Create an event set from a ParametricBasis object.
}
\examples{
basis <- BSpline(1:21, 3)
onsets <- seq(0, 20, length.out = 21)
blockids <- rep(1, length(onsets))
ebasis <- event_basis(basis, onsets, blockids)

}
