% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_factor}
\alias{event_factor}
\title{Create a categorical event sequence from a factor.}
\usage{
event_factor(
  fac,
  name,
  onsets,
  blockids = rep(1, length(fac)),
  durations = rep(0, length(fac))
)
}
\arguments{
\item{fac}{A factor representing the categorical event sequence.}

\item{name}{Name of the event sequence.}

\item{onsets}{Numeric vector of onsets.}

\item{blockids}{Numeric vector of block IDs (default: rep(1, length(fac))).}

\item{durations}{Numeric vector of durations (default: rep(0, length(fac))).}
}
\value{
An object of class "event_factor" and "event_seq".
}
\description{
Create a categorical event sequence from a factor.
}
\examples{
efac <- event_factor(factor(c("a", "b", "c", "a", "b", "c")), "abc", 
        onsets = seq(1, 100, length.out = 6))

}
\seealso{
\code{\link{event_model}}
}
