<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_factor}
\alias{event_factor}
\title{event_factor}
\usage{
event_factor(fac, name, onsets, blockids = 1, durations = NULL)
}
\arguments{
\item{fac}{}

\item{name}{}

\item{onsets}{}

\item{blockids}{}

\item{durations}{}
}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_factor}
\alias{event_factor}
\usage{
event_factor(fac, name, onsets, blockids = 1, durations = NULL)
}
>>>>>>> f1758cc34cfbefabfc88a2ae7f6f2059d8809ccc
\description{
Create an categorical event sequence from a \code{factor}
}

