% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_factor}
\alias{event_factor}
\title{event_factor}
\usage{
event_factor(
  fac,
  name,
  onsets,
  blockids = rep(1, length(fac)),
  durations = rep(0, length(fac))
)
}
\arguments{
\item{fac}{a factor}

\item{name}{the name of the event variable}

\item{onsets}{the event onsets.}

\item{blockids}{the block ids associated with each event (must be non-decreasing)}

\item{durations}{the duration of each event.}
}
\description{
Create an categorical event sequence from a \code{factor}
}
\examples{

efac <- event_factor(factor(c("a", "b", "c", "a", "b", "c")), "abc", onsets=seq(1,100,length.out=6))
}
