% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_matrix}
\alias{event_matrix}
\title{event_matrix}
\usage{
event_matrix(mat, name, onsets, blockids = rep(1, ncol(mat)), durations = NULL)
}
\arguments{
\item{mat}{a matrix of values, one row per event, indicating the amplitude/intensity of each event.}

\item{name}{the name of the event variable}

\item{onsets}{the event onsets.}

\item{blockids}{the block ids associated with each event (must be non-decreasing)}

\item{durations}{the duration of each event.}
}
\description{
Create a continuous valued event set from a \code{matrix}
}
\examples{

mat <- matrix(rnorm(200), 100, 2)
onsets <- seq(1, 1000, length.out=100)
durations <- rep(1, 100)
blockids <- rep(1, 100)

eset <- event_matrix(mat, "eset", onsets,durations,blockids)

}
