% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{event_model.list}
\alias{event_model.list}
\title{event_model.list}
\usage{
\method{event_model}{list}(
  x,
  data,
  block,
  sampling_frame,
  drop_empty = TRUE,
  durations = 0,
  precision = 0.3,
  ...
)
}
\arguments{
\item{x}{A list of HRF specifications. All elements in \code{x} must be of type \code{hrfspec}.}

\item{data}{The dataset containing the event information.}

\item{block}{The block variable, either a formula or a vector of block values.}

\item{sampling_frame}{The time series grid over which to sample the function.}

\item{drop_empty}{Logical indicating whether to drop empty events. Default is \code{TRUE}.}

\item{durations}{A numeric vector of event durations. Default is 0 for all events.}

\item{precision}{Numeric value indicating the precision of the model. Default is 0.3.}

\item{...}{Additional arguments.}
}
\value{
An event model object.
}
\description{
This function constructs an event model from a list of hemodynamic response function
(HRF) specifications \code{x}, along with other parameters like data, block, sampling frame,
and durations. It is specifically designed to handle lists of HRF specifications.
}
\details{
Construct an event model from a list of HRF specifications.
}
