% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{event_model}
\alias{event_model}
\title{construct an event model}
\usage{
event_model(
  x,
  data,
  block,
  sampling_frame,
  drop_empty = TRUE,
  durations = 0,
  ...
)
}
\arguments{
\item{x}{the model specification, typically a \code{formula} (see examples)}

\item{data}{the data containing experimental design}

\item{block}{formula for the block structure}

\item{sampling_frame}{the sampling frame defining the temporal and block structure}

\item{drop_empty}{whether to drop empty factor levels}

\item{durations}{the event durations}

\item{...}{extra args}
}
\description{
A data sructure representing an event-based fMRI regression model
}
\examples{
event_data <- data.frame(fac=c("a", "B", "A", "B"), onsets=c(1,10,20,80), run=c(1,1,1,1))
sframe <- sampling_frame(blocklens=50, TR=2)
evmodel <- event_model(onsets ~ hrf(fac), data=event_data, block= ~ run, sampling_frame=sframe)
}
