% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{event_table}
\alias{event_table}
\alias{event_table.convolved_term}
\title{event_table}
\usage{
event_table(x)

\method{event_table}{convolved_term}(x)
}
\arguments{
\item{x}{the term}
}
\value{
A \code{data.frame} containing the event table.
}
\description{
This function extracts the event table from the provided term \code{x} and returns it
as a \code{data.frame}.
}
\details{
Extract event_table as a \code{data.frame}.
}
