% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_term}
\alias{event_term}
\title{event_term}
\usage{
event_term(evlist, onsets, blockids, durations = 1, subset = NULL)
}
\description{
Create a event model term from a list of variables.
}

