% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_term}
\alias{event_term}
\title{event_term}
\usage{
event_term(evlist, onsets, blockids, durations = 1, subset = NULL)
}
\arguments{
\item{evlist}{a list of named variables}

\item{onsets}{the onset times from the experimental events in seconds}

\item{blockids}{the block number associated with each onset}
}
\description{
Create a event model term from a named list of variables.
}

