% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_term}
\alias{event_term}
\title{Create an event model term from a named list of variables.}
\usage{
event_term(evlist, onsets, blockids, durations = 1, subset = NULL)
}
\arguments{
\item{evlist}{A list of named variables.}

\item{onsets}{A vector of onset times for the experimental events in seconds.}

\item{blockids}{A vector of block numbers associated with each onset.}

\item{durations}{A vector of event durations (default is 1).}

\item{subset}{A logical vector indicating the subset of onsets to retain (default is NULL).}
}
\value{
A list containing the following components:
\itemize{
\item varname: A character string representing the variable names, concatenated with colons.
\item events: A list of event variables.
\item subset: A logical vector indicating the retained onsets.
\item event_table: A tibble containing event information.
\item onsets: A vector of onset times.
\item blockids: A vector of block numbers.
\item durations: A vector of event durations.
}
}
\description{
This function generates an event model term from a list of named variables,
along with their onsets, block IDs, and durations. Optionally, a subset of
onsets can be retained.
}
\examples{
x1 <- factor(rep(letters[1:3], 10))
x2 <- factor(rep(1:3, each=10))
eterm <- event_term(list(x1=x1,x2=x2), onsets=seq(1,100,length.out=30), 
                    blockids=rep(1,30))

x1 <- rnorm(30)
x2 <- factor(rep(1:3, each=10))
eterm <- event_term(list(x1=x1,x2=x2), onsets=seq(1,100,length.out=30), 
                    blockids=rep(1,30), subset=x1>0)

}
