% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_variable}
\alias{event_variable}
\title{Create a continuous event sequence from a numeric vector.}
\usage{
event_variable(vec, name, onsets, blockids = 1, durations = 0)
}
\arguments{
\item{vec}{Numeric vector representing continuous event values.}

\item{name}{Name of the event sequence.}

\item{onsets}{Numeric vector of onsets.}

\item{blockids}{Numeric vector of block IDs (default: 1).}

\item{durations}{Numeric vector of durations (default: 0).}
}
\value{
An object of class "event_variable" and "event_seq".
}
\description{
Create a continuous event sequence from a numeric vector.
}
\examples{
evar <- event_variable(c(1, 2, 3, 4, 5, 6), "example_var", onsets = seq(1, 100, length.out = 6))

}
\seealso{
\code{\link{event_factor}}
}
