% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_variable}
\alias{event_variable}
\title{event_variable}
\usage{
event_variable(vec, name, onsets, blockids = 1, durations = NULL)
}
\arguments{
\item{vec}{the vector event variable values}

\item{name}{the name of the event variable}

\item{onsets}{the event onsets.}

\item{blockids}{the block ids associated with each event (must be non-decreasing)}

\item{durations}{the duration of each event.}
}
\description{
Create a continuous valued event sequence from a \code{numeric} vector.
}
