% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{event_variable}
\alias{event_variable}
\title{event_variable}
\usage{
event_variable(vec, name, onsets, blockids = 1, durations = NULL)
}
\arguments{
\item{name}{the name of the variable}

\item{onsets}{the event onsets in seconds}

\item{blockids}{the index of the block/scan in which the event occurs}

\item{durations}{the durations of each event in seconds}
}
\description{
Create a continuous valued event sequence from a \code{numeric} vector.
}
