% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{Fcontrasts}
\alias{Fcontrasts}
\alias{FContrasts.convolved_term}
\title{Fcontrasts}
\usage{
Fcontrasts(x, ...)

FContrasts.convolved_term(x)
}
\arguments{
\item{x}{the term}

\item{...}{extra args}
}
\description{
generate a set of Fcontrasts for a model term
}
