% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_stats.R
\name{fit_contrasts}
\alias{fit_contrasts}
\title{Fit Contrasts}
\usage{
fit_contrasts(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments passed to methods}
}
\value{
Contrast results (format depends on method)
}
\description{
Generic function for fitting contrasts to model objects.
}
