% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{fit_contrasts}
\alias{fit_contrasts}
\title{fit_contrasts}
\usage{
fit_contrasts(lmfit, conmat, colind)
}
\arguments{
\item{lmfit}{the \code{lm} object}

\item{conmat}{the contrast \code{matrix} or contrast \code{vector}}

\item{colind}{the subset column indices in the design associated with the contrast.}
}
\description{
fit_contrasts
}

