% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{fitted_hrf}
\alias{fitted_hrf}
\title{fitted_hrf}
\usage{
fitted_hrf(x, sample_at, ...)
}
\arguments{
\item{x}{An object for which the fitted HRF should be computed.}

\item{sample_at}{A vector of time points at which the HRF should be sampled.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
A fitted HRF for the object.
}
\description{
This generic function computes the fitted hemodynamic response function (HRF) for an object.
The method needs to be implemented for specific object types.
}
