% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{fmri_dataset}
\alias{fmri_dataset}
\title{fmri_dataset}
\usage{
fmri_dataset(scans, mask, TR, run_length, event_table = data.frame(),
  aux_table = tibble::as_tibble(list()), base_path = ".")
}
\arguments{
\item{scans}{a vector of file names of the images comprising the dataset}

\item{mask}{name of the binary mask file indicating the voxels to include in analysis.}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{run_length}{the number of scans in each run.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}

\item{run_length}{the number of images in each session}

\item{aux_data}{a \code{list} of auxilliary data such as nuisance variables that may enter a 
regression model and are not convolved with hemodynamic response function.}
}
\description{
fmri_dataset
}

