% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fmri_glm}
\alias{fmri_glm}
\title{fmri_glm}
\usage{
fmri_glm(formula, baseline_formula, dataset, basis = HRF_SPMG1, durations,
  drop_empty = TRUE, analyze_by = c("run", "slice"))
}
\description{
fmri_glm
}

