% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_latent_lm.R
\name{fmri_latent_lm}
\alias{fmri_latent_lm}
\title{Fast fmri regression model estimation from a latent component dataset}
\usage{
fmri_latent_lm(
  formula,
  block,
  baseline_model = NULL,
  dataset,
  durations,
  drop_empty = TRUE,
  robust = FALSE,
  autocor = c("none", "auto", "ar1", "ar2", "arma"),
  bootstrap = FALSE,
  nboot = 1000,
  ...
)
}
\arguments{
\item{formula}{the model formula for experimental events}

\item{block}{the model formula for block structure}

\item{baseline_model}{the \code{baseline_model} object}

\item{dataset}{an object derived from \code{fmri_dataset} containing the time-series data}

\item{durations}{a vector of event durations}

\item{drop_empty}{whether to remove factor levels with size of zero}

\item{robust}{whether to use robust fitting (TRUE or FALSE)}

\item{autocor}{experimental auto-correlation correction on components}

\item{bootstrap}{whether to compute bootstrapped parameter estimates}

\item{nboot}{number of bootstrap iterations}

\item{...}{extra args}
}
\description{
Fast fmri regression model estimation from a latent component dataset
}
\details{
dataset must be of type \code{latent_dataset}, which itself require a \code{LatentNeuroVec} input
}
\note{
This method is currently experimental.
}
