% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fmri_lm}
\alias{fmri_lm}
\title{Fit a linear regression model for fMRI data analysis}
\usage{
fmri_lm(
  formula,
  block,
  baseline_model = NULL,
  dataset,
  durations,
  drop_empty = TRUE,
  robust = FALSE,
  strategy = c("runwise", "chunkwise"),
  nchunks = 10,
  ...
)
}
\arguments{
\item{formula}{The model formula for experimental events.}

\item{block}{The model formula for block structure.}

\item{baseline_model}{(Optional) The \code{baseline_model} object. Default is NULL.}

\item{dataset}{An object derived from \code{fmri_dataset} containing the time-series data.}

\item{durations}{A vector of event durations.}

\item{drop_empty}{Whether to remove factor levels with a size of zero. Default is TRUE.}

\item{robust}{Whether to use robust fitting. Default is FALSE.}

\item{strategy}{The data splitting strategy, either "runwise" or "chunkwise". Default is "runwise".}

\item{nchunks}{Number of data chunks when strategy is \code{chunkwise}. Default is 10.}

\item{...}{Extra arguments.}
}
\value{
A fitted linear regression model for fMRI data analysis.
}
\description{
This function fits a linear regression model for fMRI data analysis using the specified model formula,
block structure, and dataset. The model can be fit using either a runwise or chunkwise data splitting strategy,
and robust fitting can be enabled if desired.
}
\examples{
# Example usage of fmri_lm function
# ...
}
\seealso{
fmri_dataset, fmri_lm_fit
}
