% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fmri_lm_fit}
\alias{fmri_lm_fit}
\title{fmri_lm_fit}
\usage{
fmri_lm_fit(
  fmrimod,
  dataset,
  strategy = c("chunkwise", "runwise"),
  robust = FALSE,
  nchunks = 10,
  meta_weighting = c("inv_var", "equal"),
  ...
)
}
\arguments{
\item{fmrimod}{an object of type \code{fmri_model}}

\item{dataset}{an object derived from \code{fmri_dataset} containing the time-series data}

\item{strategy}{the data splitting strategy}

\item{robust}{whether to use robust fitting (TRUE or FALSE)}

\item{nchunks}{number of daa chunks when strategy is \code{chunkwise}}

\item{meta_weighting}{for runwise analyses, wheter to use
inverse variance ("inv_var") or equal ("equal") weighting}

\item{...}{extra args}
}
\description{
fmri_lm_fit
}
