% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{fmri_lm_fit}
\alias{fmri_lm_fit}
\title{Fit an fMRI linear regression model with a specified fitting strategy}
\usage{
fmri_lm_fit(
  fmrimod,
  dataset,
  strategy = c("chunkwise", "runwise"),
  robust = FALSE,
  nchunks = 10,
  ...
)
}
\arguments{
\item{fmrimod}{An object of type \code{fmri_model}.}

\item{dataset}{An object derived from \code{fmri_dataset} containing the time-series data.}

\item{strategy}{The data splitting strategy, either "chunkwise" or "runwise". Default is "chunkwise".}

\item{robust}{Whether to use robust fitting. Default is FALSE.}

\item{nchunks}{Number of data chunks when strategy is \code{chunkwise}. Default is 10.}

\item{...}{Extra arguments.}
}
\value{
A fitted fMRI linear regression model with the specified fitting strategy.
}
\description{
This function fits an fMRI linear regression model using the specified fmri_model object, dataset,
and data splitting strategy (either "runwise" or "chunkwise"). It is primarily an internal function
used by the fmri_lm function.
}
\seealso{
fmri_lm, fmri_model, fmri_dataset
}
\keyword{internal}
