% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{fmri_mem_dataset}
\alias{fmri_mem_dataset}
\title{fmri_mem_dataset}
\usage{
fmri_mem_dataset(scans, mask, TR, event_table = data.frame(),
  base_path = ".")
}
\arguments{
\item{scans}{a vector of objects of class \code{\linkS4class{BrainVector}}}

\item{mask}{a object of type \code{\linkS4class{LogicalBrainVolume}} indicating the voxels to include in analyses.}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}
}
\description{
fmri_mem_dataset
}
