% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{fmri_mem_dataset}
\alias{fmri_mem_dataset}
\title{fmri_mem_dataset}
\usage{
fmri_mem_dataset(
  scans,
  mask,
  TR,
  run_length = sapply(scans, function(x) dim(x)[4]),
  event_table = data.frame(),
  base_path = ".",
  censor = NULL
)
}
\arguments{
\item{scans}{a \code{list} of objects of class \code{\linkS4class{NeuroVec}}}

\item{mask}{a binary mask of class  \code{\linkS4class{NeuroVol}} indicating the set of voxels to include in analyses}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{run_length}{a \code{vector} of one or more integers indicating the number of scans in each run.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}

\item{base_path}{the file path to be prepended to relative file names.}

\item{censor}{a binary vector indicating which scans to remove.}
}
\description{
fmri_mem_dataset
}
\examples{

d <- c(10,10,10,10)
nvec <- neuroim2::NeuroVec(array(rnorm(prod(d)), d), space=neuroim2::NeuroSpace(d))
mask <- neuroim2::NeuroVol(array(rnorm(10*10*10), d[1:3]), space=neuroim2::NeuroSpace(d[1:3]))
mask[mask < .5] <- 0
dset <- fmri_mem_dataset(list(nvec),mask, TR=2)

iter <- data_chunks(dset, nchunks=100)
}
