% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_model.R
\name{fmri_model}
\alias{fmri_model}
\title{fmri_model}
\usage{
fmri_model(event_model, baseline_model)
}
\arguments{
\item{event_model}{the \code{event_model} object}

\item{baseline_model}{the \code{baseline_model} object}
}
\description{
fmri_model
}
