% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{fmri_rlm}
\alias{fmri_rlm}
\title{fmri_rlm}
\usage{
fmri_rlm(
  formula,
  block,
  baseline_model = NULL,
  dataset,
  durations,
  drop_empty = TRUE,
  nchunks = 10,
  strategy = c("runwise", "slicewise", "all")
)
}
\arguments{
\item{formula}{The model formula for experimental events.}

\item{block}{The model formula for block structure.}

\item{baseline_model}{(Optional) The \code{baseline_model} object. Default is NULL.}

\item{dataset}{An object derived from \code{fmri_dataset} containing the time-series data.}

\item{durations}{A vector of event durations.}

\item{drop_empty}{Whether to remove factor levels with a size of zero. Default is TRUE.}

\item{nchunks}{Number of data chunks when strategy is \code{chunkwise}. Default is 10.}

\item{strategy}{The data splitting strategy, either "runwise" or "chunkwise". Default is "runwise".}
}
\description{
robust linear modeling of fmri data
}
\examples{
etab <- data.frame(onset=c(1,30,15,25), fac=factor(c("A", "B", "A", "B")), run=c(1,1,2,2))
etab2 <- data.frame(onset=c(1,30,65,75), fac=factor(c("A", "B", "A", "B")), run=c(1,1,1,1))
mat <- matrix(rnorm(100*100), 100,100)
dset <- matrix_dataset(mat, TR=1, run_length=c(50,50),event_table=etab)
dset2 <- matrix_dataset(mat, TR=1, run_length=c(100),event_table=etab2)
#lm.1 <- fmri_rlm(onset ~ hrf(fac), block= ~ run,dataset=dset)
#lm.2 <- fmri_rlm(onset ~ hrf(fac), block= ~ run,dataset=dset2)
}
\keyword{experimental}
