% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{fmri_rlm}
\alias{fmri_rlm}
\title{fmri_rlm}
\usage{
fmri_rlm(
  formula,
  block,
  baseline_model = NULL,
  dataset,
  durations,
  drop_empty = TRUE,
  contrasts = NULL,
  strategy = c("runwise", "slicewise", "all")
)
}
\arguments{
\item{formula}{the model formula for experimental events}

\item{block}{the model formula for block structure}

\item{baseline_model}{the \code{baseline_model} object}

\item{dataset}{an object derived from \code{fmri_dataset} containing the time-series data}

\item{durations}{a vector of event durations}

\item{drop_empty}{whether to remove factor levels with size of zero}

\item{contrasts}{a set of contrasts}

\item{strategy}{the data splitting strategy}
}
\description{
robust linear modeling of fmri data
}
\examples{
etab <- data.frame(onset=c(1,30,15,25), fac=factor(c("A", "B", "A", "B")), run=c(1,1,2,2))
etab2 <- data.frame(onset=c(1,30,65,75), fac=factor(c("A", "B", "A", "B")), run=c(1,1,1,1))
mat <- matrix(rnorm(100*100), 100,100)
dset <- matrix_dataset(mat, TR=1, run_length=c(50,50),event_table=etab)
dset2 <- matrix_dataset(mat, TR=1, run_length=c(100),event_table=etab2)
lm.1 <- fmri_rlm(onset ~ hrf(fac), block= ~ run,dataset=dset)
lm.2 <- fmri_rlm(onset ~ hrf(fac), block= ~ run,dataset=dset2)
}
