% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrireg.R
\docType{package}
\name{fmrireg}
\alias{fmrireg}
\title{fmrireg: regresssion tools for fMRI data}
\description{
fmrireg provides functions for generating experimental design matrices appropriate for analyzing fMRI data with regression.
}
