% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_empirical_hrf}
\alias{gen_empirical_hrf}
\title{generate an empirical hemodynamic response function}
\usage{
gen_empirical_hrf(t, y, name = "empirical_hrf")
}
\arguments{
\item{t}{time}

\item{y}{values of hrf at time \code{t[i]}}

\item{name}{name of the genrated HRF}
}
\description{
generate an empirical hemodynamic response function
}
\examples{

y <- -poly(0:24, 2)[,2]
emphrf <- gen_empirical_hrf(0:24, y) 
## plot(emphrf(seq(0,24,by=.5)), type='l')
}
