% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_hrf}
\alias{gen_hrf}
\title{gen_hrf}
\usage{
gen_hrf(hrf, lag = 0, width = NULL, precision = 0.1, ...)
}
\arguments{
\item{hrf}{a function mapping from time --> signal}

\item{lag}{optional lag in seconds}

\item{width}{optional block width in seconds}

\item{precision}{sampling precision in seconds}

\item{...}{extra parameters for the \code{hrf} function}
}
\description{
gen_hrf
}
