% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{gen_hrf}
\alias{gen_hrf}
\title{gen_hrf}
\usage{
gen_hrf(hrf, lag = 0, width = NULL, precision = 0.1, ...)
}
\arguments{
\item{hrf}{a function mapping from time --> signal}

\item{lag}{optional lag in seconds}

\item{width}{optional block width in seconds}

\item{precision}{sampling precision in seconds}

\item{...}{extra parameters for the \code{hrf} function}
}
\description{
gen_hrf
}
\examples{

## generate and hrf using spmg canonical hrf, a lag of 3, and a width of 2.
grf <- gen_hrf(hrf_spmg1, lag=3, width=2)
grf(0:20)

grf <- gen_hrf(hrf_gaussian, lag=3, width=2, mean=3, sd=1)

}
