% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_model_utils.R, R/fmrilm.R
\name{get_formula.fmri_model}
\alias{get_formula.fmri_model}
\title{Model Utilities for fmri_model Objects}
\usage{
\method{get_formula}{fmri_model}(x, ...)

\method{get_formula}{fmri_model}(x, ...)
}
\arguments{
\item{x}{An \code{fmri_model} object.}

\item{...}{Additional arguments.}
}
\value{
A formula representing the model.

A formula representing the model.
}
\description{
Functions for creating and manipulating fmri_model objects

This function extracts the formula from an \code{fmri_model} object.
}
\keyword{internal}
