% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/fmri_model_utils.R, R/fmrilm.R
\name{get_formula}
\alias{get_formula}
\alias{get_formula.fmri_model}
\title{get_data}
\usage{
get_formula(x, ...)

\method{get_formula}{fmri_model}(x, ...)

\method{get_formula}{fmri_model}(x, ...)
}
\arguments{
\item{x}{The object to extract a formula from.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A formula.
}
\description{
Functions for creating and manipulating fmri_model objects
}
\examples{
fm <- fmrireg:::.demo_fmri_model()
get_formula(fm)
}
\keyword{Generic}
\keyword{Retrieve}
\keyword{a}
\keyword{expose}
\keyword{fitted}
\keyword{for}
\keyword{formula}
\keyword{internal}
\keyword{model}
\keyword{modelling}
\keyword{object}
\keyword{objects.}
\keyword{original}
\keyword{the}
\keyword{to}
\keyword{underlying}
\keyword{used}
