% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{get_formula}
\alias{get_formula}
\alias{get_formula.fmri_model}
\title{Get the formula representation of an fMRI model}
\usage{
get_formula(x, ...)
\method{get_formula}{fmri_model}(x)
}
\arguments{
\item{x}{the object}
\item{...}{extra args}
}
\value{
A formula representing the model.
}
\description{
This function extracts the formula from an \code{fmri_model} object.
}
