% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{getHRF}
\alias{getHRF}
\title{getHRF}
\usage{
getHRF(
  name = c("gam", "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian", "tent",
    "bs"),
  nbasis = 5,
  span = 24,
  lag = 0,
  width = 0,
  summate = TRUE,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{name}{The name of the HRF function. Available options: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian", "tent", "bs".}

\item{nbasis}{The number of basis functions (if relevant, default: 5).}

\item{span}{The temporal window over which the basis sets span (default: 24).}

\item{lag}{The time lag parameter (default: 0).}

\item{width}{The width parameter (default: 0).}

\item{summate}{Whether the HRF response increases its amplitude as a function of stimulus duration (default: TRUE).}

\item{normalize}{Whether to scale output so that the peak is 1 (default: FALSE).}

\item{...}{Additional arguments passed to the gen_hrf function.}
}
\value{
An HRF object with the specified properties.
}
\description{
This function retrieves a specified hemodynamic response function (HRF) by name and creates an HRF object with specified properties.
}
\examples{
hrf_obj <- getHRF(name = "gamma", nbasis = 5, span = 24)
}
\keyword{internal}
