% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{getHRF}
\alias{getHRF}
\title{getHRF}
\usage{
getHRF(name = c("gam", "gamma", "spmg1", "spmg2", "spmg3", "bspline",
  "gaussian", "tent", "bs"), nbasis = 5, lag = 0, ...)
}
\arguments{
\item{name}{the name of the hrf function}

\item{nbasis}{the number of basis functions (if relevant)}
}
\description{
getHRF
}
