% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{getHRF}
\alias{getHRF}
\title{getHRF}
\usage{
getHRF(name = c("gamma", "spmg1", "spmg2", "spmg3", "bspline"), nbasis = 5,
  ...)
}
\arguments{
\item{name}{the name of the hrf function}

\item{nbasis}{the numbe rof basis functions (if relevant)}
}
\description{
getHRF
}
