% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{global_onsets}
\alias{global_onsets}
\title{return the "global" onsets of an object.}
\usage{
global_onsets(x, onsets, ...)
}
\arguments{
\item{x}{the object}

\item{onsets}{the relative onset times of the events}

\item{...}{extra args}
}
\description{
global onsets are defined as cumlative time over runs, i.e. it does not reset to zero for each run.
}
