% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{HRF}
\alias{HRF}
\alias{HRF_GAMMA}
\alias{HRF_GAUSSIAN}
\alias{HRF_BSPLINE}
\alias{HRF_SPMG1}
\title{HRF Constructor Function}
\usage{
HRF(fun, name, nbasis = 1, span = 24, param_names = NULL)

HRF_GAMMA(t, shape = 6, rate = 1)

HRF_GAUSSIAN(t, mean = 6, sd = 2)

HRF_BSPLINE(t, span = 24, N = 5, degree = 3)

HRF_SPMG1(t, P1 = 5, P2 = 15, A1 = 0.0833)
}
\arguments{
\item{fun}{A function representing the hemodynamic response, mapping from time to BOLD response.}

\item{name}{A string specifying the name of the function.}

\item{nbasis}{An integer representing the number of basis functions, e.g., the columnar dimension of the HRF. Default is 1.}

\item{span}{A numeric value representing the span in seconds of the HRF. Default is 24.}

\item{param_names}{A character vector containing the names of the parameters for the HRF function.}
}
\value{
An HRF object with the specified properties.
}
\description{
The \code{HRF} function creates an object representing a hemodynamic response function (HRF). It is a class constructor for HRFs.
}
\section{Functions}{
\itemize{
\item \code{HRF_GAMMA()}: Gamma HRF function object

\item \code{HRF_GAUSSIAN()}: Gaussian HRF function object

\item \code{HRF_BSPLINE()}: BSpline HRF function object

\item \code{HRF_SPMG1()}: SPMG1 HRF function object

}}
\examples{
hrf <- HRF(hrf_gamma, "gamma", nbasis=1, param_names=c("shape", "rate"))
resp <- evaluate(hrf, seq(0, 24, by=1))

}
\concept{HRF}
\keyword{internal}
