% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{HRF}
\alias{HRF}
\alias{AFNI_HRF}
\alias{HRF_GAUSSIAN}
\alias{HRF_BSPLINE}
\alias{HRF_SPMG1}
\alias{HRF_SPMG2}
\alias{HRF_SPMG3}
\title{HRF constructor function}
\usage{
HRF(fun, name, nbasis = 1, span = 24, param_names = NULL)

AFNI_HRF(name, nbasis, params)

HRF_GAUSSIAN(t, mean = 6, sd = 2)

HRF_BSPLINE(t, span = 20, N = 5, degree = 3)

HRF_SPMG1(t, P1 = 5, P2 = 15)

HRF_SPMG2(t)

HRF_SPMG3(t)
}
\arguments{
\item{fun}{hemodynamic response function mapping from time --> BOLD response}

\item{name}{the name of the function}

\item{nbasis}{the number of basis functions, e.g. the columnar dimension of the hrf.}

\item{span}{the span in seconds of the HRF}

\item{param_names}{the names of the parameters}
}
\description{
a class used to represent a hemodynamic response function.
}
\section{Functions}{
\itemize{
\item \code{AFNI_HRF()}: AFNI hrf

\item \code{HRF_GAUSSIAN()}: Gaussian hrf

\item \code{HRF_BSPLINE()}: B-spline hrf

\item \code{HRF_SPMG1()}: SPMG1 hrf

\item \code{HRF_SPMG2()}: SPMG2 hrf

\item \code{HRF_SPMG3()}: SPMG3 hrf

}}
\examples{

hrf <- HRF(hrf_gamma, "gamma", nbasis=1, param_names=c("shape", "rate"))
resp <- evaluate(hrf, seq(0,24,by=1))

}
\keyword{internal}
