% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{HRF}
\alias{HRF}
\title{HRF constructor function}
\usage{
HRF(fun, name, nbasis = 1)
}
\arguments{
\item{fun}{hemodynamic response function mapping from time --> BOLD response}

\item{name}{the name of the function}

\item{nbasis}{the number of basis, e.g. the columnar dimension of the response.}
}
\description{
HRF constructor function
}

