% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{HRF}
\alias{HRF}
\title{HRF constructor function}
\usage{
HRF(fun, name, nbasis = 1, param_names = NULL)
}
\arguments{
\item{fun}{hemodynamic response function mapping from time --> BOLD response}

\item{name}{the name of the function}

\item{nbasis}{the number of basis functions, e.g. the columnar dimension of the hrf.}
}
\description{
a class used to represent a hemodynamic response function.
}
\examples{

hrf <- HRF(hrf_gamma, "gamma", nbasis=1, param_names=c("shape", "rate"))
resp <- evaluate(hrf, seq(0,24,by=1))

}

