% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\name{hrf}
\alias{hrf}
\title{hrf}
\usage{
hrf(..., basis = HRF.GAMMA, onsets = NULL, durations = NULL,
  prefix = NULL, subset = NULL, precision = 0.2)
}
\arguments{
\item{...}{the variable names}

\item{basis}{the impulse response function.}

\item{onsets}{optional onsets override. If missing, onsets will be taken from overall model specification.}

\item{durations}{optional durations override. If missing, onsets will be taken from overall model specification.}
}
\description{
hemodynamic response specification
}

