% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{hrf_gaussian}
\alias{hrf_gaussian}
\title{hrf_gaussian}
\usage{
hrf_gaussian(t, mean = 6, sd = 2)
}
\arguments{
\item{t}{time}

\item{mean}{the mean of Gaussian pdf}

\item{sd}{the standard deviation of Gaussian pdf}
}
\description{
A hemodynamic response function using the gamma density function
}
