% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_smoothing_kernel.R
\name{hrf_smoothing_kernel}
\alias{hrf_smoothing_kernel}
\title{Compute an HRF smoothing kernel}
\usage{
hrf_smoothing_kernel(len, TR = 2, form)
}
\arguments{
\item{len}{The number of scans.}

\item{TR}{The repetition time (default is 2 seconds).}

\item{form}{the \code{trialwise} formula expression, see examples.}
}
\value{
a smoothing matrix
}
\description{
This function computes a temporal similarity matrix from a series of hemodynamic response functions.
}
\examples{
form <- onsets ~ trialwise(basis="gaussian")
sk <- hrf_smoothing_kernel(100, TR=1.5, form)
}
