% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_smoothing_kernel.R
\name{hrf_smoothing_kernel}
\alias{hrf_smoothing_kernel}
\title{compute an hrf smoothing kernel}
\usage{
hrf_smoothing_kernel(len, TR = 2, basis = "spmg1")
}
\arguments{
\item{len}{the number of scans}

\item{TR}{the repetition time}

\item{basis}{the hemodynamic response function}
}
\description{
A function to compute a temporal similarity matrix from a series of hemodynamic response functions
}
