% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_time.R
\name{ihs_latent_solve}
\alias{ihs_latent_solve}
\title{Iterative Hessian Sketch (multi-RHS), returning M and Ginv}
\usage{
ihs_latent_solve(X, Z, m, iters = 3L)
}
\description{
Iterative Hessian Sketch (multi-RHS), returning M and Ginv
}
\keyword{internal}
