% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{latent_dataset}
\alias{latent_dataset}
\title{A dataset that encapsulates a dimension-reduced subspace of "latent variables".}
\usage{
latent_dataset(lvec, TR, run_length, event_table = data.frame())
}
\arguments{
\item{lvec}{an instance of class \code{LatentNeuroVec}}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{run_length}{a \code{vector} of itnegers indicating the number of scans in each run.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}
}
\description{
A dataset that encapsulates a dimension-reduced subspace of "latent variables".
}
\examples{

a matrix with 100 rows and 1000 columns (voxels)
X <- matrix(rnorm(100*1000), 100, 1000)
pres <- prcomp(X)
basis <- pres$x[,1:25]
loadings <- pres$rotation[,1:25]
offset <- colMeans(X)
lvec <- LatentNeuroVec(basis, loadings, NeuroSpace(c(10,10,10,100)), mask=rep(TRUE,1000), offset=offset)
dset <- latent_dataset(lvec, TR=2, run_length=100)

}
