% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{list_available_hrfs}
\alias{list_available_hrfs}
\title{List all available hemodynamic response functions (HRFs)}
\usage{
list_available_hrfs(details = FALSE)
}
\arguments{
\item{details}{Logical; if TRUE, attempt to add descriptions (basic for now).}
}
\value{
A data frame with columns: name, type (object/generator), nbasis_default.
}
\description{
Reads the internal HRF registry to list available HRF types.
}
