% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{longnames}
\alias{longnames}
\alias{longnames.convolved_term}
\alias{longnames.afni_hrf_convolved_term}
\alias{longnames.event_model}
\alias{longnames.event_term}
\alias{longnames.matrix_term}
\title{extract long names of variable}
\usage{
longnames(x, ...)

longnames(x, ...)

\method{longnames}{convolved_term}(x)

\method{longnames}{afni_hrf_convolved_term}(x)

\method{longnames}{event_model}(x)

\method{longnames}{event_term}(x)

\method{longnames}{matrix_term}(x)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
get the extended names of a set of variable levels
}
