% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_time.R
\name{make_time_sketch}
\alias{make_time_sketch}
\title{Build a time sketch matrix S (m x T)}
\usage{
make_time_sketch(Tlen, ctrl)
}
\arguments{
\item{Tlen}{Integer time length}

\item{ctrl}{List(method = "gaussian"|"countsketch", m, iters)}
}
\value{
Dense or sparse sketch matrix S
}
\description{
Build a time sketch matrix S (m x T)
}
\keyword{internal}
