% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{matrix_dataset}
\alias{matrix_dataset}
\title{matrix_dataset}
\usage{
matrix_dataset(datamat, TR, run_length, event_table = data.frame())
}
\arguments{
\item{datamat}{a \code{matrix} each column is a voxel time-series}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{run_length}{the number of scans in each run.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}
}
\description{
matrix_dataset
}
