% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_dataset.R
\name{matrix_dataset}
\alias{matrix_dataset}
\title{matrix_dataset}
\usage{
matrix_dataset(datamat, TR, run_length, event_table = data.frame())
}
\arguments{
\item{datamat}{a \code{matrix} each column is a voxel time-series}

\item{TR}{the repetition time in seconds of the scan-to-scan interval.}

\item{run_length}{a \code{vector} of one or more integers indicating the number of scans in each run.}

\item{event_table}{a \code{data.frame} containing the event onsets and experimental variables.}
}
\description{
matrix_dataset
}
\examples{

## a matrix with 100 rows and 100 columns (voxels)
X <- matrix(rnorm(100*100), 100, 100)
dset <- matrix_dataset(X, TR=2, run_length=100)

## an iterator with 5 chunks
iter <- data_chunks(dset, nchunks=5)
y <- foreach(chunk = iter) \%do\% { colMeans(chunk$data) }
length(y) == 5

## an iterator with 100 chunks
iter <- data_chunks(dset, nchunks=100)
y <- foreach(chunk = iter) \%do\% { colMeans(chunk$data) }
length(y) == 100

## a matrix_dataset with 200 rows, 100 columns and 2 runs
X <- matrix(rnorm(200*100), 200, 100)
dset <- matrix_dataset(X, TR=2, run_length=c(100,100))

## get a "runwise" iterator. For every iteration an entire run's worth of data is returned.
iter <- data_chunks(dset, runwise=TRUE)
y <- foreach(chunk = iter) \%do\% { colMeans(chunk$data) }
length(y) == 2
}
