% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{nuisance}
\alias{nuisance}
\title{Create a Nuisance Specification}
\usage{
nuisance(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
An object of class "nuisancespec".
}
\description{
Returns a nuisance term specification from a numeric matrix.
}
\examples{
mat <- matrix(rnorm(10), nrow = 5)
nuisance(mat)
}
