% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{one_against_all_contrast}
\alias{one_against_all_contrast}
\title{A contrast that compares each level vs the average of the other levels.}
\usage{
one_against_all_contrast(levels, facname, where = NULL)
}
\arguments{
\item{levels}{the set of factor levels to compare.}

\item{facname}{the name of the factor containing the supplied levels.}

\item{where}{the subset over which the contrast is computed}
}
\description{
A contrast that compares each level vs the average of the other levels.
}
\examples{

fac <- factor(rep(c("A", "B", "C"), 2))
con <- one_against_all_contrast(levels(fac), "fac")
}
