% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/regressor.R
\name{onsets}
\alias{onsets}
\alias{onsets.regressor}
\title{get event onsets of a variable}
\usage{
onsets(x)

\method{onsets}{regressor}(x)
}
\arguments{
\item{x}{the object}
}
\description{
return an `onset` vector
}
