% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{pair_contrast}
\alias{pair_contrast}
\title{pair_contrast}
\usage{
pair_contrast(A, B, name, where = NULL)
}
\arguments{
\item{A}{the first \code{formula} expression in the contrast}

\item{B}{the second \code{formula} expression in the contrast}

\item{name}{the name of the contrast (mandatory)}

\item{where}{the subset over which the contrast is computed}
}
\description{
sum-to-zero contrast between two logical expressions
}
\examples{

# a hypothetical experiment with a factor 'category' that takes on values of 'face' and 'scene'

pair_contrast(~ category == "face", ~ category == "scene", name="face_scene")
}
