% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{Pairwise Contrasts}
\usage{
pairwise_contrasts(levels, where = NULL)
}
\arguments{
\item{levels}{A vector of factor levels to be compared.}

\item{where}{An optional formula specifying the subset over which the contrast is computed.}
}
\value{
A contrast_set object containing pairwise contrasts for all combinations of factor levels.
}
\description{
Construct pairwise contrasts for all combinations of factor levels.
}
\examples{
pairwise_contrasts(c("A", "B", "C"))

}
