% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{poly_contrast}
\alias{poly_contrast}
\title{poly_contrast}
\usage{
poly_contrast(A, name, where = NULL, degree = 1, value_map = NULL)
}
\arguments{
\item{A}{the contrast expression as a formula}

\item{name}{the name of the contrast}

\item{where}{the subset of conditions to apply contrast to}

\item{degree}{the degree of the polynomial}

\item{value_map}{optional list that maps between levels of a factor and a numeric value.}
}
\description{
A polynomial contrast evaluated over a set of (orderable) factor levels
}
\examples{

pcon <- poly_contrast(~ time, name="poly_time_3", degree=3)
}
