% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{print.event_term}
\alias{print.event_term}
\title{Print event_term objects.}
\usage{
\method{print}{event_term}(x, ...)
}
\arguments{
\item{x}{An event_term object.}

\item{...}{Additional arguments.}
}
\description{
Print event_term objects.
}
