% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{print.fmri_lm}
\alias{print.fmri_lm}
\title{Print an fmri_lm_result object}
\usage{
\method{print}{fmri_lm}(x, ...)
}
\arguments{
\item{x}{An fmri_lm_result object.}

\item{...}{Additional arguments (unused).}
}
\description{
Provides a colorful and informative printout.
}
