% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_model.R, R/fmrilm.R, R/fmrirlm.R
\name{print.fmri_model}
\alias{print.fmri_model}
\alias{print.fmri_lm}
\alias{print.fmri_rlm}
\title{Print an fmri_lm_result object}
\usage{
\method{print}{fmri_model}(x, ...)

\method{print}{fmri_lm}(x, ...)

\method{print}{fmri_rlm}(x, ...)
}
\arguments{
\item{x}{An fmri_rlm object}

\item{...}{Additional arguments passed to print.fmri_lm}
}
\description{
Provides a colorful and informative printout.
}
