% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-methods.R
\name{print.Reg}
\alias{print.Reg}
\title{Print method for Reg objects}
\usage{
\method{print}{Reg}(x, ...)
}
\arguments{
\item{x}{A \code{Reg} object.}

\item{...}{Not used.}
}
\description{
Provides a concise summary of the regressor object using the cli package.
}
