% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{pvalues}
\alias{pvalues}
\title{Compute P-values from Meta-Analysis}
\usage{
pvalues(object, two_tailed = TRUE)
}
\arguments{
\item{object}{An fmri_meta object}

\item{two_tailed}{Logical. Use two-tailed test (default: TRUE)}
}
\value{
Matrix of p-values
}
\description{
Compute P-values from Meta-Analysis
}
