% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridataset-imports.R, R/fmrihrf-imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{matrix_dataset}
\alias{fmri_mem_dataset}
\alias{latent_dataset}
\alias{fmri_dataset}
\alias{read_fmri_config}
\alias{get_data}
\alias{get_data_matrix}
\alias{get_mask}
\alias{data_chunks}
\alias{as.matrix_dataset}
\alias{HRF}
\alias{as_hrf}
\alias{gen_hrf}
\alias{sampling_frame}
\alias{hrf_spmg1}
\alias{regressor}
\alias{blocklens}
\alias{blockids}
\alias{samples}
\alias{global_onsets}
\alias{evaluate}
\alias{nbasis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fmridataset}{\code{\link[fmridataset]{as.matrix_dataset}}, \code{\link[fmridataset]{data_chunks}}, \code{\link[fmridataset]{fmri_dataset}}, \code{\link[fmridataset]{fmri_mem_dataset}}, \code{\link[fmridataset]{get_data}}, \code{\link[fmridataset]{get_data_matrix}}, \code{\link[fmridataset]{get_mask}}, \code{\link[fmridataset]{latent_dataset}}, \code{\link[fmridataset]{matrix_dataset}}, \code{\link[fmridataset]{read_fmri_config}}}

  \item{fmrihrf}{\code{\link[fmrihrf]{as_hrf}}, \code{\link[fmrihrf]{blockids}}, \code{\link[fmrihrf]{blocklens}}, \code{\link[fmrihrf]{evaluate}}, \code{\link[fmrihrf]{gen_hrf}}, \code{\link[fmrihrf]{global_onsets}}, \code{\link[fmrihrf:HRF-class]{HRF}}, \code{\link[fmrihrf]{hrf_spmg1}}, \code{\link[fmrihrf]{nbasis}}, \code{\link[fmrihrf]{regressor}}, \code{\link[fmrihrf]{samples}}, \code{\link[fmrihrf]{sampling_frame}}}
}}

