% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridataset-imports.R, R/fmridesign-imports.R,
%   R/fmrihrf-imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{matrix_dataset}
\alias{fmri_mem_dataset}
\alias{latent_dataset}
\alias{fmri_dataset}
\alias{read_fmri_config}
\alias{get_data}
\alias{get_data_matrix}
\alias{get_mask}
\alias{data_chunks}
\alias{as.matrix_dataset}
\alias{event_model}
\alias{baseline_model}
\alias{event_factor}
\alias{event_variable}
\alias{event_matrix}
\alias{event_basis}
\alias{condition_basis_list}
\alias{covariate}
\alias{nuisance}
\alias{hrf}
\alias{design_matrix}
\alias{conditions}
\alias{cells}
\alias{convolve}
\alias{sanitize}
\alias{onsets}
\alias{durations}
\alias{elements}
\alias{blockids}
\alias{event_terms}
\alias{baseline_terms}
\alias{term_matrices}
\alias{is_categorical}
\alias{is_continuous}
\alias{term_indices}
\alias{split_by_block}
\alias{event_table}
\alias{event_term}
\alias{construct}
\alias{convolve_design}
\alias{sub_basis}
\alias{basis_suffix}
\alias{feature_suffix}
\alias{trialwise}
\alias{Ident}
\alias{Poly}
\alias{BSpline}
\alias{Scale}
\alias{ScaleWithin}
\alias{Standardized}
\alias{RobustScale}
\alias{baseline}
\alias{block}
\alias{pair_contrast}
\alias{column_contrast}
\alias{unit_contrast}
\alias{oneway_contrast}
\alias{poly_contrast}
\alias{interaction_contrast}
\alias{one_against_all_contrast}
\alias{Fcontrasts}
\alias{contrast_weights}
\alias{plot_contrasts}
\alias{design_map}
\alias{HRF}
\alias{as_hrf}
\alias{gen_hrf}
\alias{sampling_frame}
\alias{hrf_spmg1}
\alias{regressor}
\alias{blocklens}
\alias{samples}
\alias{global_onsets}
\alias{evaluate}
\alias{nbasis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fmridataset}{\code{\link[fmridataset]{as.matrix_dataset}}, \code{\link[fmridataset]{data_chunks}}, \code{\link[fmridataset]{fmri_dataset}}, \code{\link[fmridataset]{fmri_mem_dataset}}, \code{\link[fmridataset]{get_data}}, \code{\link[fmridataset]{get_data_matrix}}, \code{\link[fmridataset]{get_mask}}, \code{\link[fmridataset]{latent_dataset}}, \code{\link[fmridataset]{matrix_dataset}}, \code{\link[fmridataset]{read_fmri_config}}}

  \item{fmridesign}{\code{\link[fmridesign]{baseline}}, \code{\link[fmridesign]{baseline_model}}, \code{\link[fmridesign]{baseline_terms}}, \code{\link[fmridesign]{basis_suffix}}, \code{\link[fmridesign]{block}}, \code{\link[fmridesign:fmrihrf-reexports]{blockids}}, \code{\link[fmridesign]{BSpline}}, \code{\link[fmridesign]{cells}}, \code{\link[fmridesign]{column_contrast}}, \code{\link[fmridesign]{condition_basis_list}}, \code{\link[fmridesign]{conditions}}, \code{\link[fmridesign]{construct}}, \code{\link[fmridesign]{contrast_weights}}, \code{\link[fmridesign]{convolve}}, \code{\link[fmridesign]{convolve_design}}, \code{\link[fmridesign]{covariate}}, \code{\link[fmridesign]{design_map}}, \code{\link[fmridesign]{design_matrix}}, \code{\link[fmridesign:fmrihrf-reexports]{durations}}, \code{\link[fmridesign]{elements}}, \code{\link[fmridesign]{event_basis}}, \code{\link[fmridesign]{event_factor}}, \code{\link[fmridesign]{event_matrix}}, \code{\link[fmridesign]{event_model}}, \code{\link[fmridesign]{event_table}}, \code{\link[fmridesign]{event_term}}, \code{\link[fmridesign]{event_terms}}, \code{\link[fmridesign]{event_variable}}, \code{\link[fmridesign]{Fcontrasts}}, \code{\link[fmridesign]{feature_suffix}}, \code{\link[fmridesign]{hrf}}, \code{\link[fmridesign]{Ident}}, \code{\link[fmridesign]{interaction_contrast}}, \code{\link[fmridesign]{is_categorical}}, \code{\link[fmridesign]{is_continuous}}, \code{\link[fmridesign]{nuisance}}, \code{\link[fmridesign]{one_against_all_contrast}}, \code{\link[fmridesign]{oneway_contrast}}, \code{\link[fmridesign:fmrihrf-reexports]{onsets}}, \code{\link[fmridesign]{pair_contrast}}, \code{\link[fmridesign]{plot_contrasts}}, \code{\link[fmridesign]{Poly}}, \code{\link[fmridesign]{poly_contrast}}, \code{\link[fmridesign]{RobustScale}}, \code{\link[fmridesign]{sanitize}}, \code{\link[fmridesign]{Scale}}, \code{\link[fmridesign]{ScaleWithin}}, \code{\link[fmridesign]{split_by_block}}, \code{\link[fmridesign]{Standardized}}, \code{\link[fmridesign]{sub_basis}}, \code{\link[fmridesign]{term_indices}}, \code{\link[fmridesign]{term_matrices}}, \code{\link[fmridesign]{trialwise}}, \code{\link[fmridesign]{unit_contrast}}}

  \item{fmrihrf}{\code{\link[fmrihrf]{as_hrf}}, \code{\link[fmrihrf]{blockids}}, \code{\link[fmrihrf]{blocklens}}, \code{\link[fmrihrf]{evaluate}}, \code{\link[fmrihrf]{gen_hrf}}, \code{\link[fmrihrf]{global_onsets}}, \code{\link[fmrihrf:HRF-class]{HRF}}, \code{\link[fmrihrf]{hrf_spmg1}}, \code{\link[fmrihrf]{nbasis}}, \code{\link[fmrihrf]{regressor}}, \code{\link[fmrihrf]{samples}}, \code{\link[fmrihrf]{sampling_frame}}}
}}

