% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{regressor}
\alias{regressor}
\title{regressor}
\usage{
regressor(onsets, hrf, duration = 0, amplitude = 1, span = 20)
}
\arguments{
\item{hrf}{a hemodynamic response function}

\item{duration}{duration of events (default is 0)}

\item{amplitude}{scaling vector (default is 1)}

\item{span}{the temporal winmixed.dow of the response function (default is 20)}

\item{onset}{the event onsets in seconds}
}
\description{
construct a regressor function that can be used to generate a regression fucntion 
from a set of onset times and a hemodynamic response function
}

