% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{regressor}
\alias{regressor}
\title{regressor}
\usage{
regressor(onsets, hrf = HRF_SPMG1, duration = 0, amplitude = 1,
  span = 24, summate = TRUE)
}
\arguments{
\item{hrf}{a hemodynamic response function, e.g. \code{HRF_SPMG1}}

\item{duration}{duration of events (default is 0)}

\item{amplitude}{scaling vector (default is 1)}

\item{span}{the temporal window of the impulse response function (default is 24)}

\item{summate}{whether to summate hrf response as a function of the duration of an event.}

\item{onset}{the event onsets in seconds}
}
\value{
an S3 list of type \code{regressor}
}
\description{
construct a regressor object that can be used to generate regressor variables
from a set of onset times and a hemodynamic response function
}
\examples{

reg <- regressor(c(10,12,14,16,18, 40), HRF_SPMG1)
pred <- evaluate(reg, seq(0,100,by=2))
nbasis(reg) == 1
}
