% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R, R/all_generic.R
\name{run.afni_lm_spec}
\alias{run.afni_lm_spec}
\alias{run}
\title{run a command}
\usage{
\method{run}{afni_lm_spec}(x, outdir, reml = FALSE, execute = TRUE)

run(x, ...)
}
\arguments{
\item{x}{the command to run}

\item{outdir}{the output folder}

\item{reml}{whether to run 3dREMLFit}

\item{execute}{whether to execute the command or only output shell '3dDeconvolve.sh' script}

\item{...}{extra args}
}
\description{
run a command
}
