% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{runwise_lm}
\alias{runwise_lm}
\title{Perform Runwise Linear Modeling on fMRI Dataset}
\usage{
runwise_lm(
  dset,
  model,
  contrast_objects,
  robust = FALSE,
  verbose = FALSE,
  use_fast_path = FALSE
)
}
\arguments{
\item{dset}{An \code{fmri_dataset} object.}

\item{model}{The \code{fmri_model} used for the analysis.}

\item{contrast_objects}{The list of full contrast objects.}

\item{robust}{Logical. Whether to use robust linear modeling (default is \code{FALSE}).}

\item{verbose}{Logical. Whether to display progress messages (default is \code{FALSE}).}
}
\value{
A list containing the combined results from runwise linear model analysis.
}
\description{
This function performs a runwise linear model analysis on an fMRI dataset by
running the linear model for each data run and combining the results.
}
\keyword{internal}
