% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilm.R
\name{runwise_lm}
\alias{runwise_lm}
\title{Run glm for each data run (responses split vertically) and then combine over runs via meta-analysis}
\usage{
runwise_lm(dset, model, conlist, fcon, robust = FALSE, verbose = FALSE)
}
\description{
Run glm for each data run (responses split vertically) and then combine over runs via meta-analysis
}
