% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrirlm.R
\name{runwise_rlm}
\alias{runwise_rlm}
\title{Perform Runwise Robust Linear Modeling on fMRI Dataset}
\usage{
runwise_rlm(
  dset,
  model,
  contrast_objects,
  verbose = FALSE,
  use_fast_path = FALSE,
  progress = FALSE
)
}
\arguments{
\item{dset}{An \code{fmri_dataset} object.}

\item{model}{The \code{fmri_model} used for the analysis.}

\item{contrast_objects}{The list of full contrast objects.}

\item{verbose}{Logical. Whether to display progress messages (default is \code{FALSE}).}

\item{progress}{Logical. Display a progress bar for run processing. Default is \code{FALSE}.}
}
\description{
Perform Runwise Robust Linear Modeling on fMRI Dataset
}
\keyword{internal}
