% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{shortnames}
\alias{shortnames}
\alias{shortnames.matrix_term}
\title{shortnames}
\usage{
shortnames(x, ...)

\method{shortnames}{matrix_term}(x)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
shortnames
}
