% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{shortnames}
\alias{shortnames}
\alias{shortnames}
\title{shortnames}
\usage{
shortnames(x, ...)

shortnames(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}

\item{x}{the object}

\item{...}{extra args}
}
\description{
shortnames

return regression design as a set of matrices strored in an environment
}
