% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/event_model.R
\name{shortnames}
\alias{shortnames}
\alias{shortnames.event_model}
\alias{shortnames.convolved_term}
\alias{shortnames.event_term}
\alias{shortnames.matrix_term}
\title{extract short names of variable}
\usage{
shortnames(x, ...)

\method{shortnames}{event_model}(x, ...)

\method{shortnames}{convolved_term}(x, ...)

\method{shortnames}{event_term}(x, ...)

\method{shortnames}{matrix_term}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
get the short names of a set of variable levels
}
