% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{single_trial_regressor}
\alias{single_trial_regressor}
\title{single_trial_regressor}
\usage{
single_trial_regressor(
  onsets,
  hrf = HRF_SPMG1,
  duration = 0,
  amplitude = 1,
  span = 24
)
}
\arguments{
\item{onsets}{the event onset in seconds, must be of length 1.}

\item{hrf}{a hemodynamic response function, e.g. \code{HRF_SPMG1}}

\item{duration}{duration of the event (default is 0)}

\item{amplitude}{scaling vector (default is 1)}

\item{span}{the temporal window of the impulse response function (default is 24)}
}
\value{
an S3 list of type \code{single_trial_regressor}
}
\description{
construct a regressor object that has a single onset
}
\examples{

reg <- single_trial_regressor(c(10), HRF_SPMG1)
pred <- evaluate(reg, seq(0,100,by=2))
nbasis(reg) == 1
}
