% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_vector.R
\name{split_onsets.event_term}
\alias{split_onsets.event_term}
\title{Split onsets of an event_term object}
\usage{
\method{split_onsets}{event_term}(x, sframe, global = FALSE, blocksplit = FALSE, ...)
}
\arguments{
\item{x}{An event_term object}

\item{sframe}{A data frame representing the sampling frame}

\item{global}{A logical value indicating whether to use global onsets (default: FALSE)}

\item{blocksplit}{A logical value indicating whether to split onsets by block identifiers (default: FALSE)}

\item{...}{Additional arguments passed to other methods}
}
\value{
A list of numeric vectors representing the split onsets for each factor level or block identifier
}
\description{
This function splits the onsets of an event_term object based on its factor levels or block identifiers.
It is useful for processing fMRI data when analyzing event-related designs with multiple conditions or blocks.
}
