% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{term_indices}
\alias{term_indices}
\title{term_indices}
\usage{
term_indices(x, ...)
}
\arguments{
\item{x}{the model/term object}

\item{...}{additional arguments}
}
\value{
A vector of term indices.
}
\description{
This function retrieves the term indices for the associated design matrix
from the provided model or term object \code{x}.
}
\details{
Get the term indices for the associated design matrix.
}
