% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_model_utils.R
\name{term_matrices.fmri_model}
\alias{term_matrices.fmri_model}
\title{Extract term matrices from fmri_model}
\usage{
\method{term_matrices}{fmri_model}(x, blocknum = NULL, ...)
}
\arguments{
\item{x}{An fmri_model object}

\item{blocknum}{Optional vector of block numbers to extract matrices for}

\item{...}{Additional arguments (currently unused)}
}
\value{
A list of matrices, one for each term in the model
}
\description{
Extract design matrices for individual terms from an fmri_model object.
}
