% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{terms.baseline_model}
\alias{terms.baseline_model}
\title{Return the Terms of a Baseline Model.}
\usage{
\method{terms}{baseline_model}(x, ...)
}
\arguments{
\item{x}{A baseline_model object.}
\item{...}{Passed to underlying term constructors (unused here).}
}
\value{
A list of terms (drift, block, nuisance).
}
\description{
Return the Terms of a Baseline Model.
}
