% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{tidy}
\alias{tidy}
\title{Tidy generic}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed to methods}
}
\description{
Minimal tidy generic to support fmri_meta tidy() without requiring broom.
}
