% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{trialwise}
\alias{trialwise}
\title{trialwise}
\usage{
trialwise(..., basis = HRF_SPMG1, onsets = NULL, durations = NULL,
  prefix = NULL, subset = NULL, precision = 0.2, nbasis = 1,
  contrasts = list(), id = NULL)
}
\arguments{
\item{...}{the variable names, all of which must be present in the enclosing environment (e.g. an \code{event_model} object)}

\item{basis}{the impulse response function or the name of a pre-supplied function, one of: "gamma", "spmg1", "spmg2", "spmg3", "bspline", "gaussian".}

\item{onsets}{optional onsets override. If missing, onsets will be taken from the \code{event_model}}

\item{durations}{optional durations override. If missing, onsets will be taken from the \code{event_model}}

\item{prefix}{a character string that is prepended to the variable names and used to identify the term. 
Can be used to disambiguate two \code{hrf} terms with the same variable(s) but different onsets or basis functions.}

\item{subset}{an expression indicating the subset of 'onsets' to keep}

\item{precision}{sampling precision in seconds}

\item{nbasis}{number of basis functions -- only used for hemodynamic response functions (e.g. bspline) that take a variable number of bases.}

\item{contrasts}{one or more \code{contrast_spec} objects created with the \code{contrast} function. 
If multiple contrasts are required, then these should be wrapped in a \code{list} or \code{contrast_set}.}

\item{id}{a  unique \code{character} identifier used to refer to term, otherwise will be determined from variable names.}
}
\description{
This function is to be used in formulas for fitting functions, e.g. onsets ~ trialwise(fac1) ...
}
\examples{


## trialwise can be used with a factor with a single level because it splits each element in to separate regressor
fac <- factor(rep(1, 10))
twise <- trialwise(fac)
}
