% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{unit_contrast}
\alias{unit_contrast}
\title{unit_contrast}
\usage{
unit_contrast(A, name = NULL, where = TRUE, split_by = NULL)
}
\arguments{
\item{A}{the contrast expression as a formula}

\item{name}{the name of the contrast}

\item{where}{the subset of conditions to apply contrast to}
}
\description{
unit_contrast
}
