% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmridsl_builder.R
\name{validate_hrf_params}
\alias{validate_hrf_params}
\title{Validate HRF parameters}
\usage{
validate_hrf_params(type, params)
}
\description{
(Optional function, if you want to extend param checking)
}
\keyword{internal}
