% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{write_glt}
\alias{write_glt}
\title{Write GLT to File}
\usage{
write_glt(x, ...)
}
\arguments{
\item{x}{The GLT contrast object}

\item{...}{Additional arguments passed to methods}
}
\description{
Write a generic function for writing GLT contrast to a file.
}
