% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{write_glt}
\alias{write_glt}
\title{write a GLT file to disk}
\usage{
write_glt(x, fname)
}
\arguments{
\item{x}{the object}

\item{fname}{the file name to write to}
}
\description{
write a GLT file to disk
}
