% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{write_results.fmri_meta}
\alias{write_results.fmri_meta}
\title{Write Meta-Analysis Results}
\usage{
\method{write_results}{fmri_meta}(x, path = ".", prefix = "meta", format = c("nifti", "h5"), ...)
}
\arguments{
\item{x}{An fmri_meta object}

\item{path}{Output directory}

\item{prefix}{File name prefix}

\item{format}{Output format ("nifti" or "h5")}

\item{...}{Additional arguments passed to write_results.fmri_lm}
}
\value{
Invisible list of created files
}
\description{
Write Meta-Analysis Results
}
