% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_meta_methods.R
\name{zscores}
\alias{zscores}
\title{Compute Z-scores from Meta-Analysis}
\usage{
zscores(object)
}
\arguments{
\item{object}{An fmri_meta object}
}
\value{
Matrix of z-scores
}
\description{
Compute Z-scores from Meta-Analysis
}
